### 2024-05-13 #6749 @editorialbot
[PRE REVIEW]: multipers: Multiparameter Persistence for Machine Learning
**Submitting author:** @DavidLapous (<a href="http://orcid.org/0009-0003-5559-3712">David Loiseaux</a>)
**Repository:** https://github.com/DavidLapous/multipers
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
