### 2024-05-14 d703e8f
Activated more templates for simplextree / slicers.
