+++ _tempita_grid_gen.py
@@ -1,50 +1,125 @@
+import pickle
+from itertools import product
+
 ##  Slicers : CPP NAME, CTYPE, PYTHON_TYPE, IS_SIMPLICIAL, IS_VINE, IS_KCRITICAL, CVALUE_TYPE, PYVALUE_TYPE, COLUMN_TYPE, SHORT_DTYPE
-slicers = [ # this is temporarily necessary
-  ("GeneralVineClementTruc<>", "GeneralVineClementTruc","_SlicerClement", False, True,False, "float", "np.float32", None, "f32", "Finitely_critical_multi_filtration[float]"),
-  ("SimplicialVineGraphTruc", "SimplicialVineGraphTruc","_SlicerVineGraph", True,True,False, "float", "np.float32",None, "f32", "Finitely_critical_multi_filtration[float]"),
-  ("SimplicialVineMatrixTruc<>", "SimplicialVineMatrixTruc","_SlicerVineSimplicial", True, True,False, "float", "np.float32",None, "f32", "Finitely_critical_multi_filtration[float]"),
-  ("SimplicialNoVineMatrixTruc<>", "SimplicialNoVineMatrixTruc", "_SlicerNoVineSimplicial",True,False,False, "float", "np.float32",None, "f32", "Finitely_critical_multi_filtration[float]"),
+slicers = [  # this is temporarily necessary
+    (
+        "GeneralVineClementTruc<>",
+        "GeneralVineClementTruc",
+        "_SlicerClement",
+        False,
+        True,
+        False,
+        "float",
+        "np.float32",
+        None,
+        "f32",
+        "Finitely_critical_multi_filtration[float]",
+    ),
+    (
+        "SimplicialVineGraphTruc",
+        "SimplicialVineGraphTruc",
+        "_SlicerVineGraph",
+        True,
+        True,
+        False,
+        "float",
+        "np.float32",
+        None,
+        "f32",
+        "Finitely_critical_multi_filtration[float]",
+    ),
+    (
+        "SimplicialVineMatrixTruc<>",
+        "SimplicialVineMatrixTruc",
+        "_SlicerVineSimplicial",
+        True,
+        True,
+        False,
+        "float",
+        "np.float32",
+        None,
+        "f32",
+        "Finitely_critical_multi_filtration[float]",
+    ),
+    (
+        "SimplicialNoVineMatrixTruc<>",
+        "SimplicialNoVineMatrixTruc",
+        "_SlicerNoVineSimplicial",
+        True,
+        False,
+        False,
+        "float",
+        "np.float32",
+        None,
+        "f32",
+        "Finitely_critical_multi_filtration[float]",
+    ),
 ]
-columns_name=[ # only one column is necessary
-  "Available_columns::"+stuff
-  for stuff in (
-  "INTRUSIVE_SET",
-  # "LIST",
-  # "SET",
-  # "HEAP",
-  # "VECTOR",
-  # "NAIVE_VECTOR",
-  # "UNORDERED_SET",
-  # "INTRUSIVE_LIST",
-  )
+columns_name = [  # only one column is necessary
+    "Available_columns::" + stuff
+    for stuff in (
+        "INTRUSIVE_SET",
+        # "LIST",
+        # "SET",
+        # "HEAP",
+        # "VECTOR",
+        # "NAIVE_VECTOR",
+        # "UNORDERED_SET",
+        # "INTRUSIVE_LIST",
+    )
 ]
 
 ## Value types : CTYPE, PYTHON_TYPE, short
 value_types = [
-  ("int32_t", "np.int32",   "i32"), # necessary
-  # ("int64_t", "np.int64",   "i64"),
[…]
