+++ README.md
@@ -1,4 +1,5 @@
 # Multipers
+[![PyPI](https://img.shields.io/pypi/v/multipers?color=green)](https://pypi.org/project/multipers)
 [![Downloads](https://static.pepy.tech/badge/multipers)](https://pepy.tech/project/multipers)
 <br>
 Scikit-style multiparameter persistent homology python library. 
