+++ paper/paper.bib
@@ -59,6 +59,7 @@
   year = {2023},
   month = mar,
   journal = {arXiv.org},
+  doi = {10.48550/arxiv.2303.14040},
   urldate = {2024-04-16},
   abstract = {In this article, we study Euler characteristic techniques in topological data analysis. Pointwise computing the Euler characteristic of a family of simplicial complexes built from data gives rise to the so-called Euler characteristic profile. We show that this simple descriptor achieve state-of-the-art performance in supervised tasks at a very low computational cost. Inspired by signal analysis, we compute hybrid transforms of Euler characteristic profiles. These integral transforms mix Euler characteristic techniques with Lebesgue integration to provide highly efficient compressors of topological signals. As a consequence, they show remarkable performances in unsupervised settings. On the qualitative side, we provide numerous heuristics on the topological and geometric information captured by Euler profiles and their hybrid transforms. Finally, we prove stability results for these descriptors as well as asymptotic guarantees in random settings.},
   langid = {english},
@@ -77,16 +78,19 @@
   file = {/user/dloiseau/home/Zotero/storage/WWRX5EQI/Alonso et al. - Filtration-domination in bifiltered graphs.pdf}
 }
 
-@inproceedings{function_delaunay,
+@incollection{function_delaunay,
   title = {Delaunay {{Bifiltrations}} of {{Functions}} on {{Point Clouds}}},
-  booktitle = {2024 {{Joint Mathematics Meetings}} ({{JMM}} 2024)},
-  author = {Alonso, Angel Javier and Kerber, Michael and Lam, Tung and Lesnick, Michael},
+  booktitle = {Proceedings of the 2024 {{Annual ACM-SIAM Symposium}} on {{Discrete Algorithms}} ({{SODA}})},
+  author = {Alonso, {\'A}ngel Javier and Kerber, Michael and Lam, Tung and Lesnick, Michael},
   year = {2024},
   month = jan,
-  publisher = {AMS},
-  urldate = {2024-04-16},
-  abstract = {The Delaunay filtration  {\textbackslash}mathcal \{D\}(X)          of a point cloud  X{\textbackslash}s...},
-  file = {/user/dloiseau/home/Zotero/storage/NFBJ3KVV/32354.html}
+  series = {Proceedings},
+  pages = {4872--4891},
+  publisher = {{Society for Industrial and Applied Mathematics}},
+  doi = {10.1137/1.9781611977912.173},
+  urldate = {2024-05-14},
+  abstract = {The Delaunay filtration D.(X) of a point cloud X {$\subset$} {$\mathbb{R}$}d is a central tool of computational topology. Its use is justified by the topological equivalence of D. (X) and the offset (i.e., union-of-balls) filtration of X. Given a function {$\gamma$} : X {$\rightarrow$} {$\mathbb{R}$}, we introduce a Delaunay bifiltration DC.({$\gamma$}) that satisfies an analogous topological equivalence, ensuring that DC. ({$\gamma$}) topologically encodes the offset filtrations of all sublevel sets of {$\gamma$}, as well as the topological relations between them. DC.({$\gamma$}) is of size , which for d odd matches the worst-case size of D. (X). Adapting the Bowyer-Watson algorithm for computing Delaunay triangulations, we give a simple, practical algorithm to compute DC.({$\gamma$}) in time  Our implementation, based on CGAL, computes DC. ({$\gamma$}) with modest overhead compared to computing D. (X), and handles tens of thousands of points in {$\mathbb{R}$}3 within seconds.},
[…]
