+++ docs/paper.bib
@@ -15,6 +15,24 @@
   file = {/user/dloiseau/home/Zotero/storage/AVCY36LB/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/QH98BQFF/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/LPFXRBHW/2203.html}
 }
 
+@article{code_persistable,
+  title = {Persistable: Persistent and Stable Clustering},
+  shorttitle = {Persistable},
+  author = {Scoccola, Luis and Rolle, Alexander},
+  year = {2023},
+  month = mar,
+  journal = {Journal of Open Source Software},
+  volume = {8},
+  number = {83},
+  pages = {5022},
+  issn = {2475-9066},
+  doi = {10.21105/joss.05022},
+  urldate = {2024-05-09},
+  abstract = {Scoccola et al., (2023). Persistable: persistent and stable clustering. Journal of Open Source Software, 8(83), 5022, https://doi.org/10.21105/joss.05022},
+  langid = {english},
+  file = {/user/dloiseau/home/Zotero/storage/3DS9DWXU/Scoccola et Rolle - 2023 - Persistable persistent and stable clustering.pdf}
+}
+
 @article{erstaircode,
   title = {Elder-{{Rule-Staircodes}} for {{Augmented Metric Spaces}}},
   author = {Cai, Chen and Kim, Woojin and Memoli, Facundo and Wang, Yusu},
@@ -35,15 +53,15 @@
   file = {/user/dloiseau/home/Zotero/storage/QN5PC85K/Cai et al. - 2021 - Elder-Rule-Staircodes for Augmented Metric Spaces.pdf;/user/dloiseau/home/Zotero/storage/TKL8FVMS/2003.html}
 }
 
-@misc{eulearning,
+@article{eulearning,
   title = {Euler {{Characteristic Tools For Topological Data Analysis}}},
   author = {Hacquard, Olympio and Lebovici, Vadim},
   year = {2023},
   month = mar,
   journal = {arXiv.org},
+  doi = {10.48550/arxiv.2303.14040},
   urldate = {2024-04-16},
   abstract = {In this article, we study Euler characteristic techniques in topological data analysis. Pointwise computing the Euler characteristic of a family of simplicial complexes built from data gives rise to the so-called Euler characteristic profile. We show that this simple descriptor achieve state-of-the-art performance in supervised tasks at a very low computational cost. Inspired by signal analysis, we compute hybrid transforms of Euler characteristic profiles. These integral transforms mix Euler characteristic techniques with Lebesgue integration to provide highly efficient compressors of topological signals. As a consequence, they show remarkable performances in unsupervised settings. On the qualitative side, we provide numerous heuristics on the topological and geometric information captured by Euler profiles and their hybrid transforms. Finally, we prove stability results for these descriptors as well as asymptotic guarantees in random settings.},
-  howpublished = {https://arxiv.org/abs/2303.14040v2},
   langid = {english},
   file = {/user/dloiseau/home/Zotero/storage/GUG2VCX3/Hacquard and Lebovici - 2023 - Euler Characteristic Tools For Topological Data An.pdf}
 }
@@ -52,6 +70,7 @@
   title = {Filtration-Domination in Bifiltered Graphs},
   booktitle = {2023 Proceedings of the Symposium on Algorithm Engineering and Experiments ({{ALENEX}})},
[…]
