### 2024-05-14 #6749 @DavidLapous
> if you have any suggestions for potential reviewers then please mention them here in this thread

There are a few JOSS papers in Topological Data Analysis (TDA)-Machine Learning (ML) in Python, whose authors / previous reviewer could be possible reviewers:
 - Giotto : https://joss.theoj.org/papers/10.21105/joss.04846
 - Persistable : https://joss.theoj.org/papers/10.21105/joss.05022 (Although Luis Scoccola could have a conflict of interest)
 - Ripser: https://joss.theoj.org/papers/10.21105/joss.00925

Also, from the list of reviewer having TDA / ML
 - [Yossi Bokor](https://reviewers.joss.theoj.org/reviewers/2686)
 - [Matt Piekenbrock](https://reviewers.joss.theoj.org/reviewers/3708)
 - [EduPH](https://reviewers.joss.theoj.org/reviewers/520)
 - [sauln](https://reviewers.joss.theoj.org/reviewers/400)
 - [@leotrs](https://reviewers.joss.theoj.org/reviewers/1435)

> MISSING DOIs

I've updated the DOIs I could, but it seems that, e.g., NeurIPS and JMLR do not provide DOIs. Should I do something about it ?

> ⚠️ An error happened when generating the pdf

From the logs, the `paper.jats` file create some conflict and some pictures were missing; I've fixed that.
