+++ README.md
@@ -49,33 +49,3 @@ If I missed something, or you want to add something, feel free to open an issue.
 Feel free to contribute, report a bug on a pipeline, or ask for documentation by opening an issue.<br>
 
 
-## For mac users 
-Due to apple's clang compiler, one may have to disable a compilator optimization to compile `multipers`: in the `setup.py` file, add the 
-```bash
--fno-aligned-new
-```
-line in the `extra_compile_args` list. You should have should end up with something like the following.
-```python
-extensions = [
-    Extension(
-        f"multipers.{module}",
-        sources=[
-            f"multipers/{module}.pyx",
-        ],
-        language="c++",
-        extra_compile_args=[
-            "-Ofast",
-            "-std=c++20",
-            "-fno-aligned-new", # Uncomment this if you have trouble compiling on macos.
-            "-Wall",
-        ],
-        include_dirs=cpp_dirs,
-        define_macros=[("NPY_NO_DEPRECATED_API", "NPY_1_7_API_VERSION")],
-        libraries=["tbb", "tbbmalloc"],
-        library_dirs=library_dirs,
-    )
-    for module in cython_modules
-]
-```
-#### Alternatives
-One may try to use the `clang` compiler provided by conda or brew. If you have a simpler alternative, please let me know ;)
