### 2024-05-14 #6749 @crvernon
@DavidLapous for the references, it is OK to not have a DOI for references that do not provide them.  All good.
