### 2024-05-17 4c7bb76
fix: warning compilations, and for non-default matrix columns
