### 2024-05-23 #6749 @RMeli
Thank you @yossibokorbleile and @peekxc for agreeing to review for JOSS. I'll soon start the review process, closing this PRE-REVIEW issue and opening a REVIEW issue for the actual review.

If it's your first time reviewing for JOSS, please have a look at the following pages:
* [Reviewing for JOSS](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#)
* [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)
* [Review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html)

A good way to review is to open issues in the software repository, and [link them](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/autolinked-references-and-urls#issues-and-pull-requests) to the (soon-to-be-open) review issue.

Do not hesitate to ping me with any questions you might have.
