### 2024-06-13 1f7784c
add Differentiability paper
