### 2024-06-16 6cb02fe
feat: turn a complex to a flag complex
