+++ multipers/tests/test_simplextreemulti.py
@@ -59,40 +59,93 @@ def test_3():
 
 
 def test_4():
-    st = mp.SimplexTreeMulti(num_parameters=2, kcritical=True, dtype = np.float64)
-    st.insert([0,1,2], [0,1])
-    st.insert([0,1,2], [1,0])
-    st.remove_maximal_simplex([0,1,2])
-    st.insert([0,1,2], [1,2])
-    st.insert([0,1,2], [2,1])
-    st.insert([0,1,2],[1.5,1.5])
-    st.insert([0,1,2], [2.5,.5])
-    st.insert([0,1,2], [.5,2.5])
+    st = mp.SimplexTreeMulti(num_parameters=2, kcritical=True, dtype=np.float64)
+    st.insert([0, 1, 2], [0, 1])
+    st.insert([0, 1, 2], [1, 0])
+    st.remove_maximal_simplex([0, 1, 2])
+    st.insert([0, 1, 2], [1, 2])
+    st.insert([0, 1, 2], [2, 1])
+    st.insert([0, 1, 2], [1.5, 1.5])
+    st.insert([0, 1, 2], [2.5, 0.5])
+    st.insert([0, 1, 2], [0.5, 2.5])
 
     s = mp.Slicer(st, is_kcritical=True)
 
-    assert np.array_equal(s.get_filtrations_values(),    array([[0. , 1. ],
-       [1. , 0. ],
-       [0. , 1. ],
-       [1. , 0. ],
-       [0. , 1. ],
-       [1. , 0. ],
-       [0. , 1. ],
-       [1. , 0. ],
-       [0. , 1. ],
-       [1. , 0. ],
-       [0. , 1. ],
-       [1. , 0. ],
-       [1. , 2. ],
-       [2. , 1. ],
-       [1.5, 1.5],
-       [2.5, 0.5],
-       [0.5, 2.5]])), "Invalid conversion from kcritical st to kcritical slicer."
-    death_curve = np.asarray(mp.module_approximation(s, box = [[0,0],[3,3]]).get_module_of_degree(1)[0].get_death_list())
-    assert np.array_equal(death_curve, array([[2. , 1.5],
-       [2.5, 1. ],
-       [np.inf, 0.5],
-       [1.5, 2. ],
-       [1. , 2.5],
-       [0.5, np.inf]]))
+    assert np.array_equal(
+        s.get_filtrations_values(),
+        array(
+            [
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [0.0, 1.0],
+                [1.0, 0.0],
+                [1.0, 2.0],
+                [2.0, 1.0],
+                [1.5, 1.5],
+                [2.5, 0.5],
+                [0.5, 2.5],
+            ]
+        ),
+    ), "Invalid conversion from kcritical st to kcritical slicer."
+    death_curve = np.asarray(
+        mp.module_approximation(s, box=[[0, 0], [3, 3]])
+        .get_module_of_degree(1)[0]
+        .get_death_list()
+    )
+    assert np.array_equal(
+        death_curve,
+        array(
+            [
+                [2.0, 1.5],
+                [2.5, 1.0],
+                [np.inf, 0.5],
+                [1.5, 2.0],
+                [1.0, 2.5],
+                [0.5, np.inf],
+            ]
+        ),
+    )
 
+
+def test_make_filtration_non_decreasing():
[…]
