### 2024-06-17 9059e8a
feat: weak delaunay bifiltration autodiff
