+++ multipers/gudhi/truc.h
@@ -600,9 +600,35 @@ public:
     }
   }
 
-  std::vector<typename PersBackend::cycle_type>
-  get_representative_cycles(bool update = true) {
-    return persistence.get_representative_cycles(update);
+  // dim, num_cycle_of_dim, num_faces_in_cycle, vertices_in_face
+  std::vector<std::vector<std::vector<std::vector< unsigned int >>>>
+  get_representative_cycles(bool update = true, const std::set<int>& dims={}) {
+    // iterable iterable simplex key
+    auto cycles_key = persistence.get_representative_cycles(update);
+    auto num_cycles = cycles_key.size();
+    std::vector<std::vector<std::vector<std::vector< unsigned int >>>> out(structure.max_dimension()+1);
+    for (auto& cycles_of_dim : out)
+      cycles_of_dim.reserve(num_cycles);
+    for (auto& cycle : cycles_key) {
+      if (structure.dimension(cycle[0]) == 0 && (dims.size() == 0 || dims.contains(0))){
+        out[0].push_back({{static_cast<unsigned int >(cycle[0])}});
+        continue;
+      }
+      int cycle_dim = structure.dimension(cycle[0]); // all faces have the same dim
+      // if (!(dims.size() == 0 || dims.contains(cycle_dim)))
+      //   continue;
+      std::vector<std::vector<unsigned int>> cycle_faces(cycle.size());
+      for (std::size_t i=0; i<cycle_faces.size(); ++i ){
+        const auto& B = structure[cycle[i]];
+        cycle_faces[i].resize(B.size());
+        for (std::size_t j=0; j < B.size(); ++j){
+          cycle_faces[i][j] = static_cast<unsigned int>(B[j]);
+        }
+      }
+      out[cycle_dim].push_back(cycle_faces);
+    }
+    return out;
+
   }
   const std::vector<std::size_t> get_current_order() const {
     return generator_order;
