### 2024-06-28 bca4ac3
compatibility check with numpy2 and gudhi3.10
