### 2024-06-28 7623c25
force numpy2 in workflow
