+++ docs/build_doc.sh
@@ -1,4 +1,4 @@
 rm -rf source 
-sphinx-apidoc -f -o source .. ../setup.py
+sphinx-apidoc -f -P -o source .. ../setup.py
 rm -rf _build 
 make html
