### 2024-07-02 e9aca02
doc: update & simplification
