+++ docs/conf.py
@@ -27,7 +27,9 @@ extensions = [
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = "sphinx_rtd_theme"
-html_favicon = 'icon.svg'
+html_favicon = "icon.svg"
+autodoc_typehints = "description"
+
 
 # -- Notebooks
 nb_execution_timeout = 10
