+++ multipers/_signed_measure_meta.py
@@ -6,16 +6,14 @@ from multipers.grids import compute_grid, sms_in_grid
 from multipers.plots import plot_signed_measures
 from multipers.point_measure_integration import clean_sms
 from multipers.rank_invariant import rank_from_slicer
-from multipers.simplex_tree_multi import (
-    SimplexTreeMulti_type,
-    _available_strategies,
-    is_simplextree_multi,
-)
+from multipers.simplex_tree_multi import (SimplexTreeMulti_type,
+                                          _available_strategies,
+                                          is_simplextree_multi)
 from multipers.slicer import Slicer_type, is_slicer
 
 
 def signed_measure(
-    simplextree: Union[SimplexTreeMulti_type, Slicer_type],
+    filtered_complex: Union[SimplexTreeMulti_type, Slicer_type],
     degree: Optional[int] = None,
     degrees=[None],
     mass_default=None,
@@ -31,28 +29,37 @@ def signed_measure(
     grid_conversion: Optional[list] = None,
     coordinate_measure: bool = False,
     num_collapses: int = 0,
-    clean: bool = False,
+    clean: Optional[bool] = None,
     **infer_grid_kwargs,
 ) -> list[tuple[np.ndarray, np.ndarray]]:
     """
     Computes the signed measures given by the decomposition of the hilbert
-    function or the euler characteristic.
+    function or the euler characteristic, or the rank invariant.
 
     Input
     -----
-     - simplextree:SimplexTreeMulti, the multifiltered simplicial complex.
-       Its recommended to squeeze the simplextree first.
-     - mass_default: Either None, or 'auto' or 'inf', or array-like of floats.
-       Where to put the default mass to get a zero-mass measure.
+     - filtered_complex: given by a simplextree or a slicer.
      - degree:int|None / degrees:list[int] the degrees to compute.
        None represents the euler characteristic.
+     - mass_default: Either None, or 'auto' or 'inf', or array-like of floats.
+       Where to put the default mass to get a zero-mass measure.
+     - grid_strategy: If not squeezed yet, the strategy to coarsen the grid; see ``strategy`` in :func:`multipers.grids.compute_grid`.
+     - invariant: The invariant to use, either "hilbert", "rank", or "euler".
      - plot:bool, plots the computed measures if true.
      - n_jobs:int, number of jobs.
        Defaults to #cpu, but when doing parallel computations of signed measures, we recommend setting this to 1.
      - verbose:bool, prints c++ logs.
+     - expand_collapse: when the input is a simplextree, only expands the complex when computing 1-dimensional slices. Meant to reduce memory footprint at some computational expense.
+     - backend:str when the input is a simplextree, reduce first the filtered complex using an external library
+     see ``backend`` in :func:`multipers.io.reduce_complex`.
+     - grid_conversion: If given, re-evaluates the final signed measure in this grid.
+     - coordinate_measure: bool, if True, compute the signed measure as a coordinates given in grid_conversion.
+     - num_collapses: int, if `filtered_complex` is a simplextree, does some collapses if possible.
+     - clean: reduces the output signed measure. Only useful for euler computations.
 
     Output
     ------
+
     `[signed_measure_of_degree for degree in degrees]`
     with `signed_measure_of_degree` of the form `(dirac location, dirac weights)`.
     """
@@ -62,20 +69,22 @@ def signed_measure(
         assert len(degrees) == 1
     if len(degrees) == 0:
         return []
-    if is_slicer(simplextree):
-        if grid_conversion is None and not simplextree.is_squeezed:
+    if clean is None:
+        clean = None in degrees
+    if is_slicer(filtered_complex):
+        if grid_conversion is None and not filtered_complex.is_squeezed:
[…]
