### 2024-07-03 d56eabe
fix: bunch of refactorization (for autodiff fixes) with signed measures This fixes various issues that killed auto-differentiation if, e.g., input was a simplex tree and rank invariant, or failed to recover autodiff after some mpfree calls. This also solved some issues where, if the input was a slicer, imposing a `mass_default` location was not possible / autodiff-able.
