+++ multipers/point_measure_integration.pyx
@@ -18,6 +18,9 @@ ctypedef fused some_int:
     int
 
 ctypedef fused some_float:
+    int32_t
+    int64_t
+    int
     float
     double
 
@@ -119,7 +122,7 @@ def clean_signed_measure(some_float[:,:] pts, some_int[:] weights, dtype = np.fl
     num_diracs, num_parameters = pts.shape[:2]
     for i in range(num_diracs):
         key = tuple(pts[i]) # size cannot be known at compiletime
-        out[tuple(pts[i])] = out.get(key,0)+ weights[i]
+        out[key] = out.get(key,0)+ weights[i]
     num_keys = len(out)
     new_pts = np.fromiter(out.keys(), dtype=np.dtype((dtype,num_parameters)), count=num_keys) 
     new_weights = np.fromiter(out.values(), dtype=np.int32, count=num_keys)
@@ -136,4 +139,38 @@ def clean_sms(sms):
     """
     return tuple(clean_signed_measure(pts,weights) for pts,weights in sms)
 
+def zero_out_sm(pts,weights, mass_default):
+    """
+    Zeros out the modules outside of \f$ \{ x\in \mathbb R^n \mid x \le \mathrm{mass_default}\}\f$.
+    """
+    from itertools import product
+    # merge pts, weights
+    PTS = np.concatenate([pts,weights[:,None]], axis=1)[None]
+    
+    num_diracs, num_parameters=pts.shape
+    
+    # corners of the square of dim num_parameters. shape : (num_corners,num_parameters)
+    C = np.fromiter(product(*([[1,0]]*num_parameters)), dtype=np.dtype((np.bool_, num_parameters)),)
+    Cw = 1-2*((~C).sum(axis=1)%2)
+    # add 1 in the end to copy the shape 
+    C = np.concatenate([C, np.ones((C.shape[0],1))], axis=1).astype(np.bool_)
+    
+    # signs of each corner
+    Cw = 1-2*((~C).sum(axis=1)%2)
+
+    # something useless in the end to ensure that the shape is correct (as we added 1 in pts)
+    mass_default = np.concatenate([mass_default,[np.nan]]) 
+    mass_default = mass_default[None,None,:] ## num c, num_pts, num_params
+    # each point `pt` becomes a corner of the square [`pt`, `mass_default`]
+    new= np.where(C[:,None,:], PTS,mass_default).reshape(-1,(num_parameters+1))
+    new_pts = new[:,:-1]
+    new_masses = new[:,-1]*np.repeat(Cw,num_diracs)
+    ## a bunch of stuff are in the same place, better clean it
+    return clean_signed_measure(new_pts.astype(np.float64), new_masses.astype(np.int64), dtype=np.float64)
+ 
+def zero_out_sms(sms, mass_default):
+    """
+    Zeros out the modules outside of \f$ \{ x\in \mathbb R^n \mid x \le \mathrm{mass_default}\}\f$.
+    """
+    return tuple(zero_out_sm(pts,weights, mass_default) for pts,weights in sms)
 
