### 2024-07-03 7ac211d
fix: added an assert
