### 2024-07-04 #6773 @DavidLapous
👋 @RMeli @peekxc @yossibokorbleile Thanks for taking the time to review this!

I'm curious of what's failing on your end for the installation. If you're not using windows (which is not supported, cf DavidLapous/multipers#6), a `pip install multipers tqdm matplotlib`  with `python>=3.10` should pass all the tests, and notebooks parts that do not require external libraries.
There was also very recently a major numpy upgrade (which was unlocked with [Gudhi 3.10](https://github.com/GUDHI/gudhi-devel/releases/tag/tags%2Fgudhi-release-3.10.0) in the dependencies), for which I need to make a release, with recompiled binaries (cf. DavidLapous/multipers@bca4ac3). I'll do that in a minute.

Could you open an issue / email me if you still have an issue there ?
