### 2024-07-04 19bb19b
fix: added missing dependencies
