+++ setup.py
@@ -117,6 +117,8 @@ python_dependencies = [
     "scikit-learn",
     "joblib",
     "pot",
+    "tqdm",
+    "matplotlib",
 ]
 
 
@@ -154,7 +156,7 @@ if __name__ == "__main__":
         url="https://github.com/DavidLapous/multipers",
         # long_description=long_description,
         # long_description_content_type='text/markdown'
-        version="2.0.2",
+        version="2.0.3",
         license="MIT",
         keywords="TDA Persistence Multiparameter sklearn",
         ext_modules=cythonize(
