### 2024-07-04 a7b04eb
fix: more coherent choices in signed_measures
