### 2024-07-04 9900246
fix: removed superfluous plot
