### 2024-07-04 6c71358
fix: rank was not pushed back in the grid.
