### 2024-07-04 aeb0c52
fix: Invariant was overwritten if no degree was given.
