### 2024-07-04 DavidLapous/multipers#12 @peekxc
(I should add that installing `multipers` from pypi via pip works fine, this is just about installing from source)
