### 2024-07-04 fcb8b93
fix: cpp issues with int's infinity which is 0
