### 2024-07-04 DavidLapous/multipers#12 @DavidLapous
Thanks for this investigation.

For the dependencies, I need to investigate a bit, as, in the case of "from source compilation", I haven't updated that in a while;
My issue is that I want to use the less source as possible (i.e., only from conda) and have similar installation between OSes which may add unnecessary problems.
Also, I don't have access to a macOS machine, so if you have a working environment, I'd be happy to have it.

According to the current [GitHub workflow](https://github.com/DavidLapous/multipers/blob/19bb19b624e589856851fa761fd99d4d4e9f5859/.github/workflows/python.yml#L42-L43) (which builds + tests + sends to pypi)  the following environment should work without any issue on any up-to-date Linux or macOS machine
```sh
mamba install boost-cpp  gudhi numpy=2 cython pytest scikit-learn matplotlib boost joblib tqdm scipy tbb tbb-devel -c conda-forge
pip install pykeops filtration-domination pot --upgrade
```

Regarding your compilation error, this is due to an (old) Apple clang issue, which I used to point out, but I forgot to re-add to the documentation when I rewrote it. This should be solved by disabling a compilation optimization, by uncommenting a line [in the setup file](https://github.com/DavidLapous/multipers/blob/19bb19b624e589856851fa761fd99d4d4e9f5859/setup.py#L137).
There is also [this workaround](https://github.com/DavidLapous/multipers/blob/19bb19b624e589856851fa761fd99d4d4e9f5859/.github/workflows/python.yml#L72-L74) in the workflow which is possible.
