### 2024-07-05 #6773 @yossibokorbleile
@RMeli @DavidLapous I have managed to get multipers installed now using pip (still unsure what the initial issue was, but a fresh environment sorted it all out). I will leave checking compilation from source to @peekxc
