### 2024-07-05 DavidLapous/multipers#13 @yossibokorbleile
Mac M1/M2/M? support
Hiya! Mac M1/M2/... are not compatible with CUDA, and so KeOps does not work. I keep getting
```
[KeOps] Warning : Cuda libraries were not detected on the system or could not be loaded ; using cpu only mode
[pyKeOps] Compiling pykeops cpp 25c52bb6f5 module ...
[KeOps] Warning : There were warnings or errors :
g++: error: unrecognized command-line option '-Xclang'

OK
```
when I run
```
points = mp.data.three_annulus(5000,5000)
grades = - KDE(bandwidth=.1, return_log=True).fit(points).score_samples(points)
```
in the `function_delaunay_test.ipynb` notebook.
