### 2024-07-05 #6773 @yossibokorbleile
@RMeli @DavidLapous I have created the issue https://github.com/DavidLapous/multipers/issues/13 which is about Mac M1/M2/... support. I can test the functionality on a Linux environment, but this will delay the day I can finish this review by.
