### 2024-07-05 DavidLapous/multipers#13 @DavidLapous
Hey @yossibokorbleile, thanks for pointing that out;

First (unrelated) thing, I'm not really maintaining the notebooks in the tutorial folder anymore, as IIRC, they are all in the documentation in the docs/notebooks folder. I'm planning to remove this folder after checking that everything is in the doc.

For PyKeops on ARM + macOS, this is not linked with cuda and should work fine, but I opened an issue for a very similar problem (cf getkeops/keops#332) which was solved then, I hope what's in there helps you fix this.
Otherwise, feel free to open an issue there, I don't own a mac anymore, so it's hard for me to debug this.

You can also try to launch this with the `conda install cxx-compiler` installed.

In the meantime, the `KDE` can be completely replaced by
```python
from sklearn.neighbors import KernelDensity as KDE
```
This will be slower and not pytorch-auto-differentiable, but should work fine otherwise.
