### 2024-07-05 DavidLapous/multipers#13 @yossibokorbleile
Thanks!

Replacing KDE worked initially, until I tried
```
(simplextree,), = mmp.PointCloud2SimplexTree(expand_dim=2,bandwidths=[0.2], num_collapses="full", threshold=2, complex='rips').fit_transform([X])
```
in `multipers_intro.ipynb`.

I will just continue with testing functionality, and ignore anything `KeOps` related.
