### 2024-07-05 DavidLapous/multipers#13 @DavidLapous
Yes, this specific pipeline doesn't really have a non-keops alternative, but you can replace `bandwidths` by `masses`, which doesn't require keops, and should produce somehow similar pictures. This basically replaces a Gaussian Kernel Density Estimation (KDE) by the Distance To Measure (DTM), c.f., e.g., [this tda paper](https://arxiv.org/pdf/1811.04757).
