### 2024-07-06 eb4874f
workaround: prevents calling twice a minimal presentation
