+++ multipers/_slicer_meta.py
@@ -139,6 +139,7 @@ def Slicer(
         )
         if st.is_squeezed:
             slicer.filtration_grid = st.filtration_grid
+        slicer.minpres_degree = st.minpres_degree
     elif is_simplextree_multi(st) and backend == "graph":
         slicer = _slicer_from_simplextree(st, backend, vineyard)
         if st.is_squeezed:
