### 2024-07-06 5a2d4ef
fix: fixed various signed_measure issues
