+++ multipers/_signed_measure_meta.py
@@ -2,13 +2,16 @@ from typing import Iterable, Optional, Union
 
 import numpy as np
 
+import multipers as mp
 from multipers.grids import compute_grid, sms_in_grid
 from multipers.plots import plot_signed_measures
 from multipers.point_measure_integration import clean_sms, zero_out_sms
 from multipers.rank_invariant import rank_from_slicer
-from multipers.simplex_tree_multi import (SimplexTreeMulti_type,
-                                          _available_strategies,
-                                          is_simplextree_multi)
+from multipers.simplex_tree_multi import (
+    SimplexTreeMulti_type,
+    _available_strategies,
+    is_simplextree_multi,
+)
 from multipers.slicer import Slicer_type, is_slicer
 
 
@@ -29,7 +32,7 @@ def signed_measure(
     coordinate_measure: bool = False,
     num_collapses: int = 0,
     clean: Optional[bool] = None,
-    vineyard:bool=False,
+    vineyard: bool = False,
     **infer_grid_kwargs,
 ) -> list[tuple[np.ndarray, np.ndarray]]:
     """
@@ -50,7 +53,7 @@ def signed_measure(
        Defaults to #cpu, but when doing parallel computations of signed measures, we recommend setting this to 1.
      - verbose:bool, prints c++ logs.
      - expand_collapse: when the input is a simplextree, only expands the complex when computing 1-dimensional slices. Meant to reduce memory footprint at some computational expense.
-     - backend:str  reduces first the filtered complex using an external library, 
+     - backend:str  reduces first the filtered complex using an external library,
      see ``backend`` in :func:`multipers.io.reduce_complex`.
      - grid_conversion: If given, re-evaluates the final signed measure in this grid.
      - coordinate_measure: bool, if True, compute the signed measure as a coordinates given in grid_conversion.
@@ -67,7 +70,9 @@ def signed_measure(
     if len(degrees) == 1 and degrees[0] is None and degree is not None:
         degrees = [degree]
     if None in degrees:
-        assert len(degrees) == 1 and (invariant is None or invariant == "euler"), "Can only compute one invariant at the time."
+        assert len(degrees) == 1 and (
+            invariant is None or invariant == "euler"
+        ), "Can only compute one invariant at the time."
         invariant = "euler"
     if len(degrees) == 0:
         return []
@@ -96,20 +101,25 @@ def signed_measure(
 
     if mass_default is None:
         mass_default = mass_default
-    elif mass_default == "inf":
-        mass_default = np.array([np.inf] * filtered_complex.num_parameters)
-    elif mass_default == "auto": # this will not work with torch, but we don't want to autodiff this anyway
-        mass_default = np.array(
-            [1.1 * np.max(f) - 0.1 * np.min(f) for f in grid_conversion]
-        )
+    elif isinstance(mass_default, str):
+        if mass_default == "auto":
+            mass_default = np.array(
+                [1.1 * np.max(f) - 0.1 * np.min(f) for f in grid_conversion]
+            )
+        elif mass_default == "inf":
+            mass_default = np.array([np.inf] * filtered_complex.num_parameters)
+        else:
+            raise NotImplementedError
     else:
         mass_default = np.asarray(mass_default)
         assert (
             mass_default.ndim == 1
             and mass_default.shape[0] == filtered_complex.num_parameters
         )
-    
+
     INPUT_ARGS = locals()
+    INPUT_ARGS.pop("filtered_complex")
+
     if not filtered_complex.is_squeezed:
         filtered_complex_ = filtered_complex.grid_squeeze(
             grid_conversion, coordinates=True
@@ -122,15 +132,28 @@ def signed_measure(
         grid_conversion
[…]
