### 2024-07-06 9420f46
tests: added some for signed_measure
