+++ multipers/tests/__init__.py
@@ -21,15 +21,12 @@ def sort_sm(sms):
     return tuple((sm[0][idx], sm[1][idx]) for sm in sms)
 
 
-def assert_sm(sm1, sm2, exact=True, max_error=1e-5):
+def assert_sm_pair(sm1, sm2, exact=True, max_error=1e-3, reg=0.1):
     if not exact:
         from multipers.distances import sm_distance
 
-        for sm1_, sm2_ in zip(sm1, sm2, strict=True):
-            d = sm_distance(sm1_, sm2_)
-            assert (
-                d < max_error
-            ), f"Failed comparison:\n{sm1_}\n{sm2_},\n with distance {d}."
+        d = sm_distance(sm1, sm2, reg=0.1)
+        assert d < max_error, f"Failed comparison:\n{sm1}\n{sm2},\n with distance {d}."
         return
     assert np.all(
         [
@@ -38,3 +35,9 @@ def assert_sm(sm1, sm2, exact=True, max_error=1e-5):
             for a, b in zip(x, y, strict=True)
         ]
     ), f"Failed comparison:\n-----------------\n{sm1}\n-----------------\n{sm2}"
+
+
+def assert_sm(*args, exact=True, max_error=1e-5, reg=0.1):
+    sms = tuple(args)
+    for i in range(len(sms) - 1):
+        assert_sm_pair(sms[i], sms[i + 1], exact=exact, max_error=max_error, reg=reg)
