### 2024-07-07 5786ace
fix: various fixes of signed_measure, some tests
