+++ multipers/rank_invariant.pyx
@@ -22,142 +22,8 @@ import multipers.grids as mpg
 import multipers.slicer as mps
 from multipers.slicer cimport *
 
-# ctypedef pair[vector[vector[indices_type]], vector[tensor_dtype]] signed_measure_type
-
-# cdef extern from "multi_parameter_rank_invariant/rank_invariant.h" namespace "Gudhi::multiparameter::rank_invariant":
-    # void compute_rank_invariant_python(const intptr_t, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type, bool) except + nogil
-
-    # void compute_rank_invariant_python(GeneralNoVineTruc, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type) except + nogil
-    # void compute_rank_invariant_python(GeneralVineTruc, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type) except + nogil
-    # void compute_rank_invariant_python(SimplicialVineMatrixTruc, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type) except + nogil
-    # void compute_rank_invariant_python(SimplicialVineGraphTruc, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type) except + nogil
-    # void compute_rank_invariant_python(GeneralVineClementTruc, tensor_dtype* , const vector[indices_type], const vector[indices_type], indices_type) except + nogil
-
 from multipers.ml.signed_betti import rank_decomposition_by_rectangles
 from multipers.point_measure_integration import sparsify
-
-# def rank_invariant(simplextree, vector[indices_type] degrees, mass_default=None, plot=False, indices_type n_jobs=0, bool verbose=False, bool expand_collapse=False):
-#     """
-#     Computes the signed measures given by the decomposition of the hilbert function.
-#
-#     Input
-#     -----
-#      - simplextree:SimplexTreeMulti, the multifiltered simplicial complex
-#      - degrees:array-like of ints, the degrees to compute
-#      - mass_default: Either None, or 'auto' or 'inf', or array-like of floats. Where to put the default mass to get a zero-mass measure.
-#      - plot:bool, plots the computed measures if true.
-#      - n_jobs:int, number of jobs. Defaults to #cpu, but when doing parallel computations of signed measures, we recommend setting this to 1.
-#      - verbose:bool, prints c++ logs.
-#
-#     Output
-#     ------
-#     `[signed_measure_of_degree for degree in degrees]`
-#     with `signed_measure_of_degree` of the form `(dirac location, dirac weights)`.
-#     """
-#     assert simplextree.is_squeezed, "Squeeze grid first."
-#     assert simplextree.dtype == np.int32
-#     cdef bool zero_pad = mass_default is not None
-#     grid_conversion = [np.asarray(f) for f in simplextree.filtration_grid]
-#     # assert simplextree.num_parameters == 2
-#     grid_shape = np.array([len(f) for f in grid_conversion])
-#
-#     if mass_default is None:
-#         mass_default = mass_default
-#     else:
-#         mass_default = np.asarray(mass_default)
-#         assert mass_default.ndim == 1 and mass_default.shape[0] == simplextree.num_parameters
-#     if zero_pad:
-#         for i, _ in enumerate(grid_shape):
-#             grid_shape[i] += 1 # adds a 0
-#         for i,f in enumerate(grid_conversion):
-#             grid_conversion[i] = np.concatenate([f, [mass_default[i]]])
-#
-#     assert len(grid_shape) == simplextree.num_parameters, "Grid shape size has to be the number of parameters."
[…]
