### 2024-07-07 5ad9c71
ci: fix gudhi, readded pot to conda
