+++ multipers/gudhi/gudhi/Fields/Multi_field_small.h
@@ -1,11 +1,17 @@
 /*    This file is part of the Gudhi Library - https://gudhi.inria.fr/ - which is released under MIT.
  *    See file LICENSE or go to https://gudhi.inria.fr/licensing/ for full license details.
- *    Author(s):       Hannah Schreiber
+ *    Author(s):       Hannah Schreiber, Clément Maria
  *
- *    Copyright (C) 2022 Inria
+ *    Copyright (C) 2022-24 Inria
  *
  *    Modification(s):
  *      - YYYY/MM Author: Description of the modification
+ */
+
+/**
+ * @file Multi_field_small.h
+ * @author Hannah Schreiber, Clément Maria
+ * @brief Contains the @ref Multi_field_element_with_small_characteristics class.
  */
 
 #ifndef MATRIX_FIELD_MULTI_SMALL_H_
@@ -14,483 +20,512 @@
 #include <utility>
 #include <vector>
 #include <limits.h>
-#include <gmpxx.h>
 #include <numeric>
 
 namespace Gudhi {
-namespace persistence_matrix {
+namespace persistence_fields {
 
-//productOfAllCharacteristics_ has to fit in an unsigned int, ie, productOfAllCharacteristics_ < 4294967295
-template<unsigned int minimum, unsigned int maximum>
+/**
+ * @class Multi_field_element_with_small_characteristics Multi_field_small.h gudhi/Fields/Multi_field_small.h
+ * @ingroup persistence_fields
+ *
+ * @brief Class representing an element of a multi-field, such that the product of all characteristics fits into
+ * the given @p Unsigned_integer_type template argument. The characteristics will corresponds to all prime numbers
+ * in the interval given as other template arguments.
+ *
+ * @tparam minimum Interval closed lower bound.
+ * @tparam maximum Interval closed upper bound.
+ * @tparam Unsigned_integer_type A native unsigned integer type: unsigned int, long unsigned int, etc.
+ * Will be used as the field element type.
+ */
+template <unsigned int minimum, unsigned int maximum, typename Unsigned_integer_type = unsigned int,
+          class = std::enable_if_t<std::is_unsigned_v<Unsigned_integer_type> > >
 class Multi_field_element_with_small_characteristics {
-public:
-	using element_type = unsigned int;
-	template <class T>
-	using isInteger = std::enable_if_t<std::is_integral_v<T> >;
+ public:
+  using element_type = Unsigned_integer_type; /**< Type for the elements in the field. */
+  using characteristic_type = element_type;   /**< Type for the field characteristic. */
+  template <class T>
+  using isInteger = std::enable_if_t<std::is_integral_v<T> >;
 
-	Multi_field_element_with_small_characteristics();
-	Multi_field_element_with_small_characteristics(unsigned int element);
-	Multi_field_element_with_small_characteristics(int element);
-	Multi_field_element_with_small_characteristics(const Multi_field_element_with_small_characteristics& toCopy);
-	Multi_field_element_with_small_characteristics(Multi_field_element_with_small_characteristics&& toMove) noexcept;
+  /**
+   * @brief Default constructor. Sets the element to 0.
+   */
+  Multi_field_element_with_small_characteristics() : element_(0) {
+    static_assert(maximum >= 2, "Characteristics have to be positive.");
+    static_assert(minimum <= maximum, "The given interval is not valid.");
+    static_assert(minimum != maximum || _is_prime(minimum), "The given interval does not contain a prime number.");
+    static_assert(productOfAllCharacteristics_ != 1, "The given interval does not contain a prime number.");
+  }
+  /**
+   * @brief Constructor setting the element to the given value.
+   *
+   * @param element Value of the element.
+   */
+  template <typename Integer_type, class = isInteger<Integer_type> >
[…]
