+++ multipers/gudhi/gudhi/Persistence_matrix/matrix_dimension_holders.h
@@ -2,103 +2,183 @@
  *    See file LICENSE or go to https://gudhi.inria.fr/licensing/ for full license details.
  *    Author(s):       Hannah Schreiber
  *
- *    Copyright (C) 2022-23 Inria
+ *    Copyright (C) 2022-24 Inria
  *
  *    Modification(s):
  *      - YYYY/MM Author: Description of the modification
  */
 
+/**
+ * @file matrix_dimension_holders.h
+ * @author Hannah Schreiber
+ * @brief Contains the @ref Matrix_max_dimension_holder @ref Matrix_all_dimension_holder classes
+ * and the @ref Dummy_matrix_dimension_holder structure.
+ */
+
 #ifndef PM_MATRIX_DIM_HOLDER_H
 #define PM_MATRIX_DIM_HOLDER_H
 
-#include <utility>	//std::swap, std::move & std::exchange
+#include <utility>  //std::swap, std::move & std::exchange
 #include <vector>
 
 namespace Gudhi {
 namespace persistence_matrix {
 
-struct Dummy_matrix_dimension_holder{
-	template<typename dimension_type>
-	Dummy_matrix_dimension_holder([[maybe_unused]] dimension_type maximalDimension){}
+/**
+ * @ingroup persistence_matrix
+ *
+ * @brief Empty structure.
+ * Inheritated instead of @ref Matrix_max_dimension_holder or @ref Matrix_all_dimension_holder, when the maximal
+ * dimension of a matrix is not stored.
+ */
+struct Dummy_matrix_dimension_holder {
+  template <typename dimension_type>
+  Dummy_matrix_dimension_holder([[maybe_unused]] dimension_type maximalDimension) {}
 
-	friend void swap([[maybe_unused]] Dummy_matrix_dimension_holder& d1, [[maybe_unused]] Dummy_matrix_dimension_holder& d2){}
+  friend void swap([[maybe_unused]] Dummy_matrix_dimension_holder& d1,
+                   [[maybe_unused]] Dummy_matrix_dimension_holder& d2) {}
 };
 
-template<typename dimension_type>
-class Matrix_max_dimension_holder
+/**
+ * @ingroup persistence_matrix
+ *
+ * @brief Class managing the maximal dimension of a face represented in the inheritating matrix, when the option of
+ * face removal is not enabled.
+ * 
+ * @tparam dimension_type Dimension value type. Has to be an integer type.
+ * If unsigned, the maximal value of the type should not be attained during a run.
+ */
+template <typename dimension_type>
+class Matrix_max_dimension_holder 
 {
-public:
-	Matrix_max_dimension_holder(dimension_type maximalDimension = -1) : maxDim_(maximalDimension){};
-	Matrix_max_dimension_holder(const Matrix_max_dimension_holder& toCopy) : maxDim_(toCopy.maxDim_){};
-	Matrix_max_dimension_holder(Matrix_max_dimension_holder&& other) noexcept 
-		: maxDim_(std::exchange(other.maxDim_, -1)){};
+ public:
+  /**
+   * @brief Default constructor. If a dimension is specified, stores it as the maximal value.
+   * 
+   * @param maximalDimension Value of the maximal dimension. Has to be either positive or -1. Default value: -1.
+   */
+  Matrix_max_dimension_holder(dimension_type maximalDimension = -1) : maxDim_(maximalDimension){};
+  /**
+   * @brief Copy constructor.
+   * 
+   * @param toCopy Matrix to copy.
+   */
+  Matrix_max_dimension_holder(const Matrix_max_dimension_holder& toCopy) : maxDim_(toCopy.maxDim_){};
[…]

[…]
