### 2024-07-08 906ad5b
fix: separate templates
