### 2024-07-08 c3bde76
fix: minpres of degree 0 wasn't properly handled
