### 2024-07-08 f22870d
tests: fixed some tests with mpfree
