### 2024-07-08 6687b5e
feat: check if external software is working
