### 2024-07-10 b3ff8cd
fix: unspecified invariant was leading to an error
