### 2024-07-10 1f482d8
optim: ignore infinite simplices in non-vine computations
