### 2024-07-11 02224da
feat: slicer dimension pruning
