### 2024-07-11 3aaa5f5
fix: workaround of a PersistenceMatrix bug by pruning unneccessary gens
