+++ multipers/gudhi/truc.h
@@ -34,18 +34,17 @@ public:
       if (stuff == 0)
         num_vertices_++;
     }
-    max_dimension_ = generator_dimensions.size() > 0 ? *std::max_element(generator_dimensions.begin(),
-                                         generator_dimensions.end()) : 0;
-
+    max_dimension_ = generator_dimensions.size() > 0
+                         ? *std::max_element(generator_dimensions.begin(),
+                                             generator_dimensions.end())
+                         : 0;
   };
 
   PresentationStructure(const PresentationStructure &other)
       : generators(other.generators),
         generator_dimensions(other.generator_dimensions),
         num_vertices_(other.num_vertices_),
-        max_dimension_(other.max_dimension_) {
-
-  }
+        max_dimension_(other.max_dimension_) {}
   /* PresentationStructure &operator=(const PresentationStructure &other) { */
   /*   generators = other.generators; */
   /*   generator_dimensions = other.generator_dimensions; */
@@ -69,8 +68,7 @@ public:
         stream << truc << ", ";
 
       if (!stuff.empty())
-        stream << "\b"
-               << "\b ";
+        stream << "\b" << "\b ";
 
       stream << "},\n";
     }
@@ -80,8 +78,7 @@ public:
     for (const auto &stuff : structure.generator_dimensions)
       stream << stuff << ", ";
     if (structure.size() > 0) {
-      stream << "\b"
-             << "\b";
+      stream << "\b" << "\b";
     }
     stream << "}\n";
     return stream;
@@ -99,8 +96,16 @@ public:
   }
   inline std::size_t size() const { return generators.size(); };
   unsigned int num_vertices() const { return num_vertices_; };
-  unsigned int max_dimension() const {
-    return max_dimension_;
+  unsigned int max_dimension() const { return max_dimension_; }
+  int prune_above_dimension(int dim) {
+    int idx = std::lower_bound(generator_dimensions.begin(),
+                               generator_dimensions.end(), dim + 1) -
+              generator_dimensions.begin();
+    generators.resize(idx);
+    generator_dimensions.resize(idx);
+    max_dimension_ =
+        generator_dimensions.size() ? generator_dimensions.back() : -1;
+    return idx;
   }
 
 private:
@@ -144,8 +149,7 @@ public:
         stream << truc << ", ";
 
       if (!stuff.empty())
-        stream << "\b"
-               << "\b ";
+        stream << "\b" << "\b ";
 
       stream << "},\n";
     }
@@ -167,14 +171,15 @@ public:
   inline unsigned int num_vertices() { return num_vertices_; }
   inline unsigned int max_dimension() const { return max_dimension_; }
 
+  int prune_above_dimension(int dim) { throw; }
+
 private:
   std::vector<std::vector<unsigned int>> boundaries;
   unsigned int num_vertices_;
   unsigned int max_dimension_;
 };
 
-template <class PersBackend, class Structure,
-          class MultiFiltration>
+template <class PersBackend, class Structure, class MultiFiltration>
 class Truc {
 public:
   using Filtration_value = MultiFiltration;
@@ -234,7 +239,11 @@ public:
   };
 
   inline std::size_t num_generators() const { return structure.size(); }
-  inline std::size_t num_parameters() const { return num_generators() == 0 ? 0 : this->generator_filtration_values[0].num_parameters(); }
+  inline std::size_t num_parameters() const {
+    return num_generators() == 0
+               ? 0
+               : this->generator_filtration_values[0].num_parameters();
+  }
 
   template <class SubFiltration, bool original_order = true>
   inline void push_to_out(
@@ -273,7 +282,7 @@ public:
   inline PersBackend compute_persistence_out(
[…]
