### 2024-07-11 ee04c31
feat: automatic generator pruning for faster computations, more verbose
