+++ docs/paper.bib
@@ -1,72 +1,4 @@
-@misc{botnanlesnick,
-  title = {An {{Introduction}} to {{Multiparameter Persistence}}},
-  author = {Botnan, Magnus Bakke and Lesnick, Michael},
-  year = {2022},
-  month = mar,
-  number = {arXiv:2203.14289},
-  eprint = {2203.14289},
-  primaryclass = {cs, math},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2203.14289},
-  urldate = {2022-06-11},
-  abstract = {In topological data analysis (TDA), one often studies the shape of data by constructing a filtered topological space, whose structure is then examined using persistent homology. However, a single filtered space often does not adequately capture the structure of interest in the data, and one is led to consider multiparameter persistence, which associates to the data a space equipped with a multiparameter filtration. Multiparameter persistence has become one of the most active areas of research within TDA, with exciting progress on several fronts. In this article, we introduce multiparameter persistence and survey some of this recent progress, with a focus on ideas likely to lead to practical applications in the near future.},
-  archiveprefix = {arxiv},
-  keywords = {Computer Science - Computational Geometry,Mathematics - Algebraic Topology,Mathematics - Representation Theory},
-  file = {/user/dloiseau/home/Zotero/storage/AVCY36LB/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/QH98BQFF/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/LPFXRBHW/2203.html}
-}
-
-@article{code_persistable,
-  title = {Persistable: Persistent and Stable Clustering},
-  shorttitle = {Persistable},
-  author = {Scoccola, Luis and Rolle, Alexander},
-  year = {2023},
-  month = mar,
-  journal = {Journal of Open Source Software},
-  volume = {8},
-  number = {83},
-  pages = {5022},
-  issn = {2475-9066},
-  doi = {10.21105/joss.05022},
-  urldate = {2024-05-09},
-  abstract = {Scoccola et al., (2023). Persistable: persistent and stable clustering. Journal of Open Source Software, 8(83), 5022, https://doi.org/10.21105/joss.05022},
-  langid = {english},
-  file = {/user/dloiseau/home/Zotero/storage/3DS9DWXU/Scoccola et Rolle - 2023 - Persistable persistent and stable clustering.pdf}
-}
-
-@article{erstaircode,
-  title = {Elder-{{Rule-Staircodes}} for {{Augmented Metric Spaces}}},
-  author = {Cai, Chen and Kim, Woojin and Memoli, Facundo and Wang, Yusu},
-  year = {2021},
-  month = jan,
-  journal = {SIAM Journal on Applied Algebra and Geometry},
-  volume = {5},
-  number = {3},
-  eprint = {2003.04523},
-  primaryclass = {cs, math},
-  pages = {417--454},
-  issn = {2470-6566},
-  doi = {10.1137/20M1353605},
-  urldate = {2022-05-20},
[…]
