### 2024-07-16 #6773 @RMeli
> I have managed to get multipers installed now using pip (still unsure what the initial issue was, but a fresh environment sorted it all out).

I'm glad this worked out in the end!
