### 2024-07-17 0988d22
doc: Persistence Matrix column types
