### 2024-07-17 f330952
fix: rank sm plot was ignoring alpha
