### 2024-07-19 a2274f1
optim: faster synthetic point cloud generation
