### 2024-07-26 32efc46
fix: wrong arg in sm pipeline
