### 2024-07-26 19ffd1b
fix: saner defaults
