### 2024-07-29 b14f161
simpler workflow for PRs
