+++ multipers/__init__.py
@@ -3,6 +3,8 @@ import multipers.io
 import multipers.multiparameter_module_approximation
 import multipers.simplex_tree_multi
 import multipers.slicer
+import multipers.data
+import multipers.grids
 from multipers._signed_measure_meta import signed_measure
 from multipers._slicer_meta import Slicer
 from multipers.multiparameter_module_approximation import module_approximation
