### 2024-07-30 32f6779
workaround: only allow minpreses for multiple computations
