### 2024-07-30 63ae3cb
defaults: multithread mma by default
