### 2024-08-02 c4b6dd7
st/slicer uniformization, FilteredComplex type
