### 2024-08-02 c568c95
renamed SimplexTree2MMA -> FilteredComplex2MMA + allows slicer.
