### 2024-08-02 7a9bdab
tests:faster sm tests
