+++ multipers/tests/test_point_clouds.py
@@ -2,7 +2,6 @@ import numpy as np
 
 import multipers as mp
 import multipers.ml.point_clouds as mmp
-from multipers.tests import assert_sm
 
 mp.simplex_tree_multi.SAFE_CONVERSION = True
 np.random.seed(0)
@@ -10,20 +9,33 @@ np.random.seed(0)
 
 def test_throw_test():
     pts = np.array([[1, 1], [2, 2]], dtype=np.float32)
-    st = mmp.PointCloud2SimplexTree(masses=[0.1]).fit_transform([pts])[0][0]
+    st = mmp.PointCloud2FilteredComplex(masses=[0.1]).fit_transform([pts])[0][0]
     assert isinstance(st, mp.simplex_tree_multi.SimplexTreeMulti_type)
-    st = mmp.PointCloud2SimplexTree(bandwidths=[-0.1], complex="alpha").fit_transform(
+    st = mmp.PointCloud2FilteredComplex(bandwidths=[-0.1], complex="alpha").fit_transform(
         [pts]
     )[0][0]
     assert isinstance(st, mp.simplex_tree_multi.SimplexTreeMulti_type)
-    st1, st2 = mmp.PointCloud2SimplexTree(bandwidths=[0.1], masses=[0.1]).fit_transform(
+
+    st = mmp.PointCloud2FilteredComplex(masses=[0.1], complex="alpha", output_type="slicer_novine").fit_transform(
+        [pts]
+    )[0][0]
+    assert isinstance(st, mp.slicer.Slicer_type)
+    assert st.is_vine is False
+
+    st = mmp.PointCloud2FilteredComplex(masses=[0.1], complex="alpha", output_type="slicer").fit_transform(
+        [pts]
+    )[0][0]
+    assert isinstance(st, mp.slicer.Slicer_type)
+    assert st.is_vine
+
+    st1, st2 = mmp.PointCloud2FilteredComplex(bandwidths=[0.1], masses=[0.1]).fit_transform(
         [pts]
     )[0]
     assert isinstance(st1, mp.simplex_tree_multi.SimplexTreeMulti_type)
     assert isinstance(st2, mp.simplex_tree_multi.SimplexTreeMulti_type)
     ## ensures it doesn't throw
     assert isinstance(
-        st.persistence_approximation(),
+        mp.module_approximation(st),
         mp.multiparameter_module_approximation.PyModule_type,
     )
     assert mp.signed_measure(st, degree=None, invariant="euler")[0][0].ndim == 2
@@ -31,29 +43,3 @@ def test_throw_test():
     assert mp.signed_measure(st, degree=0, invariant="rank")[0][0].ndim == 2
     assert mp.signed_measure(st, degree=0, invariant="rank")[0][0].shape[1] == 4
 
-
-# def test_1():
-#     import multipers.data
-#
-#     pts = mp.data.noisy_annulus(20, 20)
-#     st = mmp.PointCloud2SimplexTree(bandwidths=[-0.5], expand_dim=2).fit_transform(
-#         [pts]
-#     )[0][0]
-#     F = st.get_filtration_grid()
-#     for invariant in ["hilbert", "rank"]:
-#         sm1 = mp.signed_measure(
-#             st.grid_squeeze(F), degree=1, invariant=invariant, mass_default=None
-#         )
-#         st.collapse_edges(-2, ignore_warning=True)
-#         s = mp.Slicer(st, dtype=np.float64)
-#         mp.io._init_external_softwares()
-#         if mp.io.pathes["mpfree"] is None:
-#             from warnings import warn
-#
-#             warn("Could not find mpfree, skipping this test")
-#         else:
-#             s = s.minpres(degree=1).grid_squeeze(F)
[…]
