+++ multipers/ml/point_clouds.py
@@ -1,4 +1,4 @@
-from typing import Literal, Optional
+from typing import Literal, Optional,Iterable
 
 import gudhi as gd
 import numpy as np
@@ -16,7 +16,7 @@ def _throw_nofit(any):
     raise Exception("Fit first")
 
 
-class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
+class PointCloud2FilteredComplex(BaseEstimator, TransformerMixin):
     def __init__(
         self,
         bandwidths=[],
@@ -24,7 +24,7 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
         threshold: float = np.inf,
         complex: Literal["alpha", "rips", "delaunay"] = "rips",
         sparse: float | None = None,
-        num_collapses: int | Literal["full"] = "full",
+        num_collapses:int= -2,
         kernel: str = "gaussian",
         log_density: bool = True,
         expand_dim: int = 1,
@@ -33,6 +33,8 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
         fit_fraction: float = 1,
         verbose: bool = False,
         safe_conversion: bool = False,
+        output_type:Optional[Literal["slicer","simplextree", "slicer_vine", "slicer_novine"]]=None,
+        reduce_degrees:Optional[Iterable[int]]=None,
     ) -> None:
         """
         (Rips or Alpha or Delaunay) + (Density Estimation or DTM) 1-critical 2-filtration.
@@ -73,6 +75,12 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
         self.sparse = sparse
         self._get_sts = _throw_nofit
         self.safe_conversion = safe_conversion
+        self.output_type = output_type
+        self._output_type = None
+        self.reduce_degrees = reduce_degrees
+        self._vineyard=None
+        
+        assert output_type != "simplextree" or reduce_degrees is None, "Reduced complex are not simplicial. Cannot return a simplextree."
         return
 
     def _get_distance_quantiles(self, X, qs):
@@ -110,7 +118,7 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
             # RIPS has contigus vertices, so vertices are ordered.
             st_copy.fill_lowerstar(codensity, parameter=1)
 
-        def collapse_edges(st):
+        def reduce(st):
             if self.verbose:
                 print("Num simplices :", st.num_simplices)
             if isinstance(self.num_collapses, int):
@@ -127,10 +135,14 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
                     print(", after expansion :", st.num_simplices, end="")
             if self.verbose:
                 print("")
+            if self._output_type == "slicer":
+                st = mp.Slicer(st, vineyard=self._vineyard)
+                if self.reduce_degrees is not None:
+                    st = mp.minimal_presentation(st, degrees=self.reduce_degrees)
             return st
 
         return Parallel(backend="threading", n_jobs=self.n_jobs)(
-            delayed(collapse_edges)(st) for st in sts
+            delayed(reduce)(st) for st in sts
         )
 
     def _get_sts_alpha(self, x: np.ndarray, return_alpha=False):
@@ -154,6 +166,12 @@ class PointCloud2SimplexTree(BaseEstimator, TransformerMixin):
             alligned_codensity[vertices] = codensity
             # alligned_codensity = np.array([codensity[i] if i in vertices else np.nan for i in range(max_vertices)])
             st_copy.fill_lowerstar(alligned_codensity, parameter=1)
+        if self._output_type == "slicer":
+            sts2 = (mp.Slicer(st, vineyard=self._vineyard) for st in sts)
+            if self.reduce_degrees is not None:
[…]
