### 2024-08-03 7327341
hilbert from slicer & related fixes
