+++ multipers/gudhi/mma_interface_matrix.h
@@ -68,7 +68,7 @@ class Persistence_backend_matrix {
 public:
   using matrix_type = Gudhi::persistence_matrix::Matrix<Matrix_options>;
   using cycle_type = typename matrix_type::cycle_type;
-  static const bool is_vine  = Matrix_options::has_vine_update;
+  static const bool is_vine = Matrix_options::has_vine_update;
 
   using bar = typename matrix_type::Bar;
   //   using index = typename matrix_type::index;
@@ -123,10 +123,12 @@ public:
         auto &b = barcode_->operator[](currPos_);
         currBar_.dim = b.dim;
         currBar_.birth = perm_->operator[](b.birth);
-        if (debug && currBar_.birth > std::numeric_limits<decltype(currBar_.birth)>::max()/2){
-          std::cout <<currBar_ << std::endl;
-          std::cout <<"while "  <<  b.birth;
-          std::cout <<"  "<<  perm_->size();
+        if (debug &&
+            currBar_.birth >
+                std::numeric_limits<decltype(currBar_.birth)>::max() / 2) {
+          std::cout << currBar_ << std::endl;
+          std::cout << "while " << b.birth;
+          std::cout << "  " << perm_->size();
         }
         currBar_.death = b.death == static_cast<pos_index>(-1)
                              ? -1
@@ -140,7 +142,7 @@ public:
     Barcode(matrix_type &matrix, const std::vector<std::size_t> *perm)
         : barcode_(&matrix.get_current_barcode()) {
       const bool debug = false;
-      if constexpr(Matrix_options::has_vine_update){
+      if constexpr (Matrix_options::has_vine_update) {
         perm_ = perm;
       } else {
         perm_.reserve(perm->size());
@@ -149,18 +151,18 @@ public:
             perm_.push_back(stuff);
       }
       if constexpr (debug) {
-        std::cout << "Built matrix of size " << matrix.get_number_of_columns() << " / " << perm->size() 
-                  << std::endl;
+        std::cout << "Built matrix of size " << matrix.get_number_of_columns()
+                  << " / " << perm->size() << std::endl;
       }
     }
 
-    iterator begin() const { 
-      if constexpr(Matrix_options::has_vine_update){
+    iterator begin() const {
+      if constexpr (Matrix_options::has_vine_update) {
         return Barcode_iterator(barcode_, perm_);
+      } else {
+        return Barcode_iterator(barcode_, &this->perm_);
       }
-      else{
-        return Barcode_iterator(barcode_, &this->perm_); 
-      }}
+    }
 
     iterator end() const { return Barcode_iterator(); }
 
@@ -184,11 +186,13 @@ public:
 
   private:
     const typename matrix_type::barcode_type *barcode_;
-    typename std::conditional<Matrix_options::has_vine_update, const std::vector<std::size_t>*, std::vector<std::size_t>>::type perm_;
+    typename std::conditional<Matrix_options::has_vine_update,
+                              const std::vector<std::size_t> *,
+                              std::vector<std::size_t>>::type perm_;
   };
 
   Persistence_backend_matrix() : permutation_(nullptr){};
-  Persistence_backend_matrix(Boundary_matrix_type &boundaries,
+  Persistence_backend_matrix(const Boundary_matrix_type &boundaries,
                              std::vector<std::size_t> &permutation)
       : matrix_(boundaries.size()), permutation_(&permutation) {
 
@@ -199,27 +203,30 @@ public:
     std::vector<std::size_t> boundary_container;
     std::size_t c = 0;
     for (std::size_t i : *permutation_) {
-      if (i >= boundaries.size())
-        { c++;continue; }
+      if (i >= boundaries.size()) {
+        c++;
+        continue;
+      }
       permutationInv[i] = c++;
[…]
