### 2024-08-03 608cf05
hilbert from slicer & related fixes
