+++ multipers/multi_parameter_rank_invariant/hilbert_function.h
@@ -5,6 +5,7 @@
 #include "tensor/tensor.h"
 #include <algorithm>
 #include <gudhi/Simplex_tree/multi_filtrations/Finitely_critical_filtrations.h>
+#include <gudhi/truc.h>
 #include <iostream>
 #include <limits>
 #include <oneapi/tbb/enumerable_thread_specific.h>
@@ -12,16 +13,16 @@
 #include <ostream>
 #include <utility> // std::pair
 #include <vector>
-#include <gudhi/truc.h>
 
 namespace Gudhi::multiparameter::hilbert_function {
 
 // TODO : this function is ugly
 template <typename value_type, typename indices_type>
-inline typename Simplex_tree_std::Filtration_value horizontal_line_filtration2(
-    const std::vector<value_type> &x, indices_type height, indices_type i,
-    indices_type j, const std::vector<indices_type> &fixed_values) {
-  constexpr const Simplex_tree_std::Filtration_value inf = std::numeric_limits<typename Simplex_tree_std::Filtration_value>::infinity();
+inline double
+horizontal_line_filtration2(const std::vector<value_type> &x,
+                            indices_type height, indices_type i, indices_type j,
+                            const std::vector<indices_type> &fixed_values) {
+  constexpr const double inf = std::numeric_limits<double>::infinity();
   for (indices_type k = 0u; k < static_cast<indices_type>(x.size()); k++) {
     if (k == i || k == j)
       continue;                 // coordinate in the plane
@@ -104,8 +105,7 @@ inline void compute_2d_hilbert_surface(
 
     coordinates_container[j + 1] = height;
 
-    Filtration multi_filtration(
-        st_multi.get_number_of_parameters());
+    Filtration multi_filtration(st_multi.get_number_of_parameters());
     auto sh_standard = st_std.complex_simplex_range().begin();
     auto _end = st_std.complex_simplex_range().end();
     auto sh_multi = st_multi.complex_simplex_range().begin();
@@ -114,18 +114,20 @@ inline void compute_2d_hilbert_surface(
       // std::ranges::views::zip(st_std.complex_simplex_range(),
       // st_multi.complex_simplex_range())){ // too bad apple clang exists
       multi_filtration = st_multi.filtration(*sh_multi);
-typename Simplex_tree_std::Filtration_value horizontal_filtration ;
-      if constexpr (Filtration::is_multi_critical)
-      {
-        horizontal_filtration = std::numeric_limits<typename Simplex_tree_std::Filtration_value>::infinity();
-        for (const auto& stuff : multi_filtration)
-          horizontal_filtration = std::min(
-              horizontal_filtration,
-              horizontal_line_filtration2(stuff, height, i, j, fixed_values));
+      typename Simplex_tree_std::Filtration_value horizontal_filtration;
+      if constexpr (Filtration::is_multi_critical) {
+        horizontal_filtration = std::numeric_limits<
+            typename Simplex_tree_std::Filtration_value>::infinity();
+        for (const auto &stuff : multi_filtration)
+          horizontal_filtration =
+              std::min(horizontal_filtration,
+                       static_cast<Simplex_tree_std::Filtration_value>(
+                           horizontal_line_filtration2(stuff, height, i, j,
+                                                       fixed_values)));
+      } else {
+        horizontal_filtration = horizontal_line_filtration2(
+            multi_filtration, height, i, j, fixed_values);
       }
-      else
-        {horizontal_filtration = horizontal_line_filtration2(
-            multi_filtration, height, i, j, fixed_values);}
       st_std.assign_filtration(*sh_standard, horizontal_filtration);
 
       if constexpr (verbose) {
@@ -288,15 +290,15 @@ void _rec_get_hilbert_surface(
 }
 
 template <typename Filtration, typename dtype, typename index_type>
[…]
