+++ multipers/gudhi/truc.h
@@ -285,6 +285,7 @@ public:
     return this->filtration_container;
   }
 
+  template <bool ignore_inf = true>
   inline PersBackend compute_persistence_out(
       const std::vector<typename MultiFiltration::value_type> &one_filtration,
       std::vector<std::size_t> &out_gen_order)
@@ -304,7 +305,7 @@ public:
                   return true;
                 return one_filtration[i] < one_filtration[j];
               });
-    if constexpr (!PersBackend::is_vine) {
+    if constexpr (!PersBackend::is_vine && ignore_inf) {
       for (std::size_t &i : out_gen_order)
         if (one_filtration[i] == MultiFiltration::T_inf) {
           // TODO : later
@@ -329,9 +330,7 @@ public:
       }
       std::cout << "]" << std::endl;
     }
-    return PersBackend(
-        structure,
-        out_gen_order); // FIXME : PersBackend is not const on struct
+    return PersBackend(structure, out_gen_order);
   }
 
   inline const bool has_persistence() { return this->persistence.size(); };
@@ -386,7 +385,7 @@ public:
                       1); // TODO : This doesn't allow for negative dimensions
     const bool verbose = false;
     const bool debug = false;
-    auto inf = MultiFiltration::T_inf;
+    const auto inf = MultiFiltration::T_inf;
     for (const auto &bar : barcode_indices) {
       if constexpr (verbose)
         std::cout << "BAR : " << bar.birth << " " << bar.death << "\n";
@@ -441,9 +440,7 @@ public:
     if (num_bars <= 0)
       return out;
     auto idx = 0u;
-    value_type inf = std::numeric_limits<value_type>::has_infinity
-                         ? std::numeric_limits<value_type>::infinity()
-                         : std::numeric_limits<value_type>::max();
+    const value_type inf = MultiFiltration::T_inf;
     for (const auto &bar : barcode_indices) {
       value_type birth_filtration = inf;
       value_type death_filtration = -birth_filtration;
@@ -483,9 +480,7 @@ public:
     if (num_bars <= 0)
       return out;
     auto idx = 0u;
-    value_type inf = std::numeric_limits<value_type>::has_infinity
-                         ? std::numeric_limits<value_type>::infinity()
-                         : std::numeric_limits<value_type>::max();
+    const value_type inf = MultiFiltration::T_inf;
     for (const auto &bar : barcode_indices) {
       value_type birth_filtration = inf;
       value_type death_filtration = -birth_filtration;
@@ -583,7 +578,7 @@ public:
     }
     return {a, b};
   }
-  std::vector<typename MultiFiltration::OneCritical>
+  inline std::vector<typename MultiFiltration::OneCritical>
   get_filtration_values() const {
     if constexpr (MultiFiltration::is_multi_critical) {
       std::vector<typename MultiFiltration::OneCritical> out;
@@ -599,13 +594,13 @@ public:
       return generator_filtration_values; // copy not necessary for Onecritical
     } // (could return const&)
   }
-  std::vector<MultiFiltration> &get_filtrations() {
+  inline std::vector<MultiFiltration> &get_filtrations() {
     return generator_filtration_values;
   }
-  const std::vector<MultiFiltration> &get_filtrations() const {
+  inline const std::vector<MultiFiltration> &get_filtrations() const {
     return generator_filtration_values;
   }
-  const std::vector<int> get_dimensions() const {
+  inline const std::vector<int> get_dimensions() const {
     std::size_t n = this->num_generators();
     std::vector<int> out(n);
     for (std::size_t i = 0; i < n; ++i) {
@@ -621,7 +616,7 @@ public:
     filtration_container.resize(idx);
   }
 
-  const std::vector<std::vector<unsigned int>> get_boundaries() {
[…]
