### 2024-08-24 4bb5df0
feat: slicer (parallel+vineyard) python interface
