+++ multipers/gudhi/truc.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "gudhi/Simplex_tree/multi_filtrations/Line.h"
 #include "multiparameter_module_approximation/format_python-cpp.h"
 #include <algorithm>
 #include <boost/mpl/aux_/na_fwd.hpp>
@@ -9,6 +10,8 @@
 #include <iostream>
 #include <limits>
 #include <numeric>
+#include <oneapi/tbb/enumerable_thread_specific.h>
+#include <oneapi/tbb/parallel_for.h>
 #include <oneapi/tbb/task_group.h>
 #include <sstream>
 #include <string>
@@ -133,7 +136,7 @@ public:
   SimplicialStructure(const std::vector<std::vector<unsigned int>> &boundaries,
                       unsigned int num_vertices, unsigned int max_dimension)
       : boundaries(boundaries), num_vertices_(num_vertices),
-        max_dimension_(max_dimension){
+        max_dimension_(max_dimension) {
 
         };
 
@@ -232,7 +235,7 @@ public:
     std::iota(generator_order.begin(), generator_order.end(), 0); // range
   }
 
-  Truc(){};
+  Truc() {};
 
   Truc &operator=(const Truc &other) {
     generator_filtration_values = other.generator_filtration_values;
@@ -796,6 +799,62 @@ public:
     PersBackend persistence; // generated by the structure, and generator_order.
 
   }; // class TrucThread
+
+  /*
+   * returns barcodes of the f(multipers)
+   *
+   */
+  template <typename Fun, typename Fun_arg>
+  inline std::vector<split_barcode> barcodes(Fun &&f,
+                                             const std::vector<Fun_arg> &args) {
+    if (args.size() == 0) {
+      return {};
+    }
+    std::vector<split_barcode> out(args.size());
+
+    if constexpr (PersBackend::is_vine) {
+      this->push_to(f(args[0]));
+      this->compute_persistence();
+      out[0] = this->get_barcode();
+      for (auto i = 1u; i < args.size(); ++i) {
+        this->push_to(f(args[i]));
+        this->vineyard_update();
+        out[i] = this->get_barcode();
+      }
+
+    } else {
+      ThreadSafe local_template = this->weak_copy();
+      tbb::enumerable_thread_specific<ThreadSafe> thread_locals(local_template);
+      tbb::parallel_for(static_cast<std::size_t>(0), args.size(),
+                        [&](const std::size_t &i) {
+                          ThreadSafe &s = thread_locals.local();
+                          s.push_to(f(args[i]));
+                          s.compute_persistence();
+                          out[i] = s.get_barcode();
+                        });
+    }
+    return out;
+  }
+  // FOR Python interface, but I'm not fan. Todo: do the lambda function in
+  // cython?
+  inline std::vector<split_barcode>
+  persistence_on_lines(const std::vector<std::vector<value_type>> &basepoints) {
+    return barcodes(
+        [](const std::vector<value_type> &basepoint) {
+          return multi_filtrations::Line<value_type>(basepoint);
+        },
+        basepoints);
+  }
+  inline std::vector<split_barcode> persistence_on_lines(
+      const std::vector<std::pair<std::vector<value_type>,
+                                  std::vector<value_type>>> &bp_dirs) {
+    return barcodes(
+        [](const std::pair<std::vector<value_type>, std::vector<value_type>>
+               &bpdir) {
+          return multi_filtrations::Line<value_type>(bpdir.first, bpdir.second);
+        },
+        bp_dirs);
+  }
 
 public:
   using ThreadSafe = TrucThread; // for outside
