### 2024-08-24 12e08c1
fix: sm plot only triggers newfig if necessary
