### 2024-08-25 021501c
feat: barcode from rank signed measure
