### 2024-08-25 ad5b809
feat: more doc for signed measure
