### 2024-08-26 04a3301
doc: added some info in signed measure
