### 2024-08-26 8c530de
workaround for external libraries
