### 2024-08-26 1c89795
optim: slightly faster io
