### 2024-08-26 187255c
fix: pruning too aggressive
