+++ multipers/torch/rips_density.py
@@ -157,8 +157,10 @@ def function_rips_signed_measure(
     theta: Optional[float] = None,
     function: Literal["dtm", "gaussian", "exponential"] | Callable = "gaussian",
     threshold: Optional[float] = None,
-    grid_strategy: Literal["regular_closest", "exact", "quantile", "regular_left"] = "exact",
-    complex:Literal["rips", "delaunay"] = "rips",
+    grid_strategy: Literal[
+        "regular_closest", "exact", "quantile", "regular_left"
+    ] = "exact",
+    complex: Literal["rips", "delaunay"] = "rips",
     resolution: int = 100,
     safe_conversion: bool = False,
     num_collapses: Optional[int] = None,
@@ -168,7 +170,7 @@ def function_rips_signed_measure(
     # return_st: bool = False,
     *,
     log_density: bool = True,
-    vineyard:bool = False,
+    vineyard: bool = False,
     **sm_kwargs,
 ):
     """
@@ -206,7 +208,9 @@ def function_rips_signed_measure(
             .type(dtype)
         )
     elif isinstance(function, torch.Tensor):
-        assert function.ndim == 1 and codensity.shape[0] == x.shape[0], """
+        assert (
+            function.ndim == 1 and codensity.shape[0] == x.shape[0]
+        ), """
         When function is a tensor, it is interpreted as the value of some function over x. 
         """
         codensity = function
@@ -217,17 +221,20 @@ def function_rips_signed_measure(
         else:
             codensity = function(x, theta=theta).type(dtype)
 
-    
     distance_matrix = torch.cdist(x, x).type(dtype)
     distances = distance_matrix.ravel()
-    if  complex == "rips":
-        threshold = distance_matrix.max(axis=1).values.min() if threshold is None else threshold
+    if complex == "rips":
+        threshold = (
+            distance_matrix.max(axis=1).values.min() if threshold is None else threshold
+        )
         distances = distances[distances <= threshold]
     elif complex == "delaunay":
         distances /= 2
     else:
-        raise ValueError(f"Unimplemented with complex {complex}. You can use rips or delaunay ftm.")
-    
+        raise ValueError(
+            f"Unimplemented with complex {complex}. You can use rips or delaunay ftm."
+        )
+
     # simplificates the simplextree for computation, the signed measure will be recovered from the copy afterward
     reduced_grid = get_grid(strategy=grid_strategy)((distances, codensity), resolution)
 
@@ -247,30 +254,36 @@ def function_rips_signed_measure(
         if None in degrees:
             expansion_degree = st.num_vertices
         else:
-            expansion_degree = (
-                max(degrees) + 1
-            )
+            expansion_degree = max(degrees) + 1
         st.collapse_edges(num=num_collapses)
         if not expand_collapse:
             st.expansion(expansion_degree)  # edge collapse
 
         s = mp.Slicer(st, vineyard=vineyard)
     elif complex == "delaunay":
-        s = mp.slicer.from_function_delaunay(x.detach().numpy(),codensity.detach().numpy())
+        s = mp.slicer.from_function_delaunay(
+            x.detach().numpy(), codensity.detach().numpy()
+        )
         st = mp.slicer.to_simplextree(s)
         st.flagify(2)
         s = mp.Slicer(st, vineyard=vineyard)
-    
+
     if None not in degrees:
         s = s.minpres(degrees=degrees)
     else:
         from joblib import Parallel, delayed
-        s = tuple(Parallel(n_jobs = -1, backend="threading")(delayed(lambda d : s if d is None else s.minpres(degree=d))(d) for d in degrees))
+
+        s = tuple(
[…]
