### 2024-09-02 ecbb380
fix: point cloud pipeline issue
