### 2024-09-02 7c88a6f
merged tempita options
