### 2024-09-02 4f6dfba
merged tempita options
