### 2024-09-02 2101cdb
fix: mma of multiple minpres
