+++ multipers/ml/mma.py
@@ -12,12 +12,15 @@ from multipers.grids import compute_grid as reduce_grid
 from multipers.ml.tools import filtration_grid_to_coordinates
 from multipers.mma_structures import PyBox_f64, PyModule_type
 
+_FilteredComplexType = Union[
+    mp.slicer.Slicer_type, mp.simplex_tree_multi.SimplexTreeMulti_type
+]
 
-_FilteredComplexType=Union[mp.slicer.Slicer_type|mp.simplex_tree_multi.SimplexTreeMulti_type]
 
 class FilteredComplex2MMA(BaseEstimator, TransformerMixin):
     """
-    Turns a list of simplextrees to MMA approximations
+    Turns a list of list of simplextrees or slicers to MMA approximations.
+
     """
 
     def __init__(
@@ -26,84 +29,86 @@ class FilteredComplex2MMA(BaseEstimator, TransformerMixin):
         expand_dim: Optional[int] = None,
         prune_degrees_above: Optional[int] = None,
         progress=False,
-        minpres_degrees:Optional[Iterable[int]]=None,
+        minpres_degrees: Optional[Iterable[int]] = None,
         **persistence_kwargs,
     ) -> None:
         super().__init__()
         self.persistence_args = persistence_kwargs
         self.n_jobs = n_jobs
-        self._has_axis = None
         self._num_axis = None
         self.prune_degrees_above = prune_degrees_above
         self.progress = progress
         self.expand_dim = expand_dim
         self._boxes = None
-        self.minpres_degrees=minpres_degrees
+        self._is_minpres = None
+        self.minpres_degrees = minpres_degrees
         return
-    
+
     @staticmethod
     def _is_filtered_complex(input):
-        return mp.simplex_tree_multi.is_simplextree_multi(input) or mp.slicer.is_slicer(input)
+        return mp.simplex_tree_multi.is_simplextree_multi(input) or mp.slicer.is_slicer(
+            input, allow_minpres=True
+        )
 
     def _input_checks(self, X):
-        self._has_axis = not self._is_filtered_complex(X[0])
-        if self._has_axis:
-            assert len(X) > 0, "No filtered complex found. Cannot fit."
-            assert self._is_filtered_complex(X[0][0]), f"X[0] is not a known filtered complex, {X[0]=}, nor X[0][0]."
-            self._num_axis = len(X[0])
-        first = X[0][0] if self._has_axis else X[0]
-        assert not mp.slicer.is_slicer(first) or self.expand_dim is None, "Cannot expand slicers."
+        assert len(X) > 0, "No filtered complex found. Cannot fit."
+        assert self._is_filtered_complex(
+            X[0][0]
+        ), f"X[0] is not a known filtered complex, {X[0]=}, nor X[0][0]."
+        self._num_axis = len(X[0])
+        first = X[0][0]
+        assert (
+            not mp.slicer.is_slicer(first) or self.expand_dim is None
+        ), "Cannot expand slicers."
+        self._is_minpres = mp.slicer.is_slicer(first) and isinstance(
+            first, Union[tuple, list]
+        )
+        assert not (
+            self._is_minpres and self.minpres_degrees is not None
+        ), "Input is already a minpres. Cannot reduce again."
 
-    def _infer_bounding_box(self,X):
-        if self._has_axis:
-            filtration_values = np.asarray(
+    def _infer_bounding_box(self, X):
+        assert self._num_axis is not None, "Fit first"
+        filtration_values = (
+            np.asarray(
+                [
+                    [s.filtration_bounds() for x in X for s in x[axis]]
+                    for axis in range(self._num_axis)
+                ]
+            )
+            if self._is_minpres
+            else np.asarray(
                 [
[…]
