### 2024-09-02 921c500
typing stuff
