+++ multipers/tests/test_mma.py
@@ -1,5 +1,6 @@
 import numpy as np
-import pytest 
+import pytest
+
 import multipers as mp
 import multipers.ml.mma as mma
 from multipers.tests import random_st
@@ -28,40 +29,50 @@ def test_img():
     assert np.isclose(img[0, 1, 1, 1, 1], 0)
     assert np.isclose(img[0, 3, 4, 5, 5], 1)
 
-@pytest.mark.parametrize("n_jobs", [1,2])
-@pytest.mark.parametrize("prune_degrees_above", [0,1, None])
-def test_pipeline1(prune_degrees_above,n_jobs):
+
+@pytest.mark.parametrize("n_jobs", [1, 2])
+@pytest.mark.parametrize("prune_degrees_above", [0, 1, None])
+def test_pipeline1(prune_degrees_above, n_jobs):
     args = locals()
-    st = random_st(npts=50,max_dim=1).collapse_edges(-2)
+    st = random_st(npts=50, max_dim=1).collapse_edges(-2)
     st.expansion(2)
-    truc1 = mma.FilteredComplex2MMA(**args).fit_transform([st])[0]
-    truc2 = mma.FilteredComplex2MMA(**args).fit_transform([mp.Slicer(st)])[0]
+    (truc1,) = mma.FilteredComplex2MMA(**args).fit_transform([[st]])[0]
+    (truc2,) = mma.FilteredComplex2MMA(**args).fit_transform([[mp.Slicer(st)]])[0]
     box = st.filtration_bounds()
     st_copy = st.copy()
     if prune_degrees_above is not None:
         st_copy.prune_above_dimension(prune_degrees_above)
-    output = mp.module_approximation(st_copy, box = box).representation(bandwidth=.1, kernel="linear")
-    assert np.sum(output)>0, "Invalid mma rpz"
-    assert np.array_equal(truc1.representation(bandwidth=.1, kernel="linear"), truc2.representation(bandwidth=.1, kernel="linear")), "Slicer == Simplextree not satisfied"
-    assert np.array_equal(truc1.representation(bandwidth=.1, kernel="linear"), output)
-    
+    output = mp.module_approximation(st_copy, box=box).representation(
+        bandwidth=0.1, kernel="linear"
+    )
+    assert np.sum(output) > 0, "Invalid mma rpz"
+    assert np.array_equal(
+        truc1.representation(bandwidth=0.1, kernel="linear"),
+        truc2.representation(bandwidth=0.1, kernel="linear"),
+    ), "Slicer == Simplextree not satisfied"
+    assert np.array_equal(truc1.representation(bandwidth=0.1, kernel="linear"), output)
+
     st = [random_st(npts=50).collapse_edges(-2, ignore_warning=True) for _ in range(5)]
     some_fited_pipeline = mma.FilteredComplex2MMA(**args).fit([st])
-    truc1 = some_fited_pipeline.transform([st])[0]
-    truc2 = mma.FilteredComplex2MMA(**args).fit_transform([[mp.Slicer(truc) for truc in st]])[0]
-    for a,b in zip(truc1,truc2):
-        assert np.array_equal(a.representation(bandwidth=.01, kernel="linear"), b.representation(bandwidth=.01, kernel="linear")), "Slicer == Simplextree not satisfied"
+    truc1 = some_fited_pipeline.transform([st])
+    truc2 = mma.FilteredComplex2MMA(**args).fit_transform(
+        [[mp.Slicer(truc) for truc in st]]
+    )
+    for a, b in zip(truc1, truc2):
+        assert np.array_equal(
+            a[0].representation(bandwidth=0.01, kernel="linear"),
[…]
