### 2024-09-06 b20152d
fix: deprecated typing stuff
