### 2024-09-06 343e031
fix: dense torch measure
