### 2024-09-06 fadc5aa
fix: rank invariant, and moved to slicer file
