+++ setup.py
@@ -1,18 +1,55 @@
 import contextlib
 import os
 import site
+import shutil
+import filecmp
 
 import numpy as np
-import sklearn._build_utils
 from Cython.Build import cythonize
 from Cython.Compiler import Options
-from setuptools import Extension, find_packages, setup
+from Cython import Tempita
+from setuptools import Extension, setup
 
 Options.docstrings = True
 Options.embed_pos_in_docstring = True
 Options.fast_fail = True
 # Options.warning_errors = True
 
+os.system("mkdir -p ./build/tmp")
+
+def was_modified(file):
+    tail = os.path.basename(file)
+    new_file = "build/tmp/" + tail
+    if not os.path.isfile(new_file):
+        shutil.copyfile(file, new_file)
+        return True
+    else:
+        res = not filecmp.cmp(new_file, file, shallow=False)
+        if res:
+            shutil.copyfile(file, new_file)
+        return res
+
+# credit to sklearn with just a few modifications:
+# https://github.com/scikit-learn/scikit-learn/blob/156ef1b7fe9bc0ee5b281634cfd56b9c54e83277/sklearn/_build_utils/tempita.py
+# took it out to not having to depend on a sklearn version in addition to a cython version
+def process_tempita(fromfile):
+    """Process tempita templated file and write out the result.
+
+    The template file is expected to end in `.c.tp` or `.pyx.tp`:
+    E.g. processing `template.c.tp` generates `template.c`.
+
+    """
+    if not was_modified(fromfile):
+        return
+    with open(fromfile, "r", encoding="utf-8") as f:
+        template_content = f.read()
+
+    template = Tempita.Template(template_content)
+    content = template.substitute()
+
+    outfile = os.path.splitext(fromfile)[0]
+    with open(outfile, "w", encoding="utf-8") as f:
+        f.write(content)
 
 cython_modules = [
     "simplex_tree_multi",
@@ -21,15 +58,10 @@ cython_modules = [
     "function_rips",
     "mma_structures",
     "multiparameter_module_approximation",
-    # "diff_helper",
-    # "hilbert_function",
-    # "euler_characteristic",
-    # 'cubical_multi_complex',
     "point_measure",
     "grids",
     "slicer",
 ]
-
 
 templated_cython_modules = [
     "filtration_conversions.pxd",
@@ -39,25 +71,23 @@ templated_cython_modules = [
     "slicer.pyx",
 ]
 
+## generates some parameter files (Tempita fails with python<3.12)
+# TODO: see if there is a way to avoid _tempita_grid_gen.py or a nicer way to do it
+os.system("python _tempita_grid_gen.py")
 
-os.system(
-    "rm _simplextrees_.pkl _slicer_names.pkl & python _tempita_grid_gen.py"
-)  ## generates some parameter files (Tempita fails with python<3.12)
-sklearn._build_utils.gen_from_templates(
-    (f"multipers/{mod}.tp" for mod in templated_cython_modules)
-)
+[process_tempita(f"multipers/{mod}.tp") for mod in templated_cython_modules]
 
+## Broken on mac
+# n_jobs = 1
+# with contextlib.suppress(ImportError):
+#     import joblib
 
-n_jobs = 1
-with contextlib.suppress(ImportError):
-    import joblib
-
-    n_jobs = joblib.cpu_count()
+#     n_jobs = joblib.cpu_count()
 
 cythonize_flags = {
-    # "depfile":True,
+    # "depfile": True,
     # "nthreads": n_jobs,  # Broken on mac
-    # "show_all_warnings":True,
+    # "show_all_warnings": True,
 }
 
[…]
