+++ pyproject.toml
@@ -1,15 +1,67 @@
 [build-system]
 requires = [
-    "numpy>=1.24",
-    "cython>=3.0",
-    "scikit-learn>=1.3",
-    "setuptools",
-    "wheel",
-    "gudhi>=3.9",
-    "tqdm",
-    "joblib",
-    "filtration-domination",
-    "pykeops",
-    "pot",
+  "Cython",
+  "setuptools"
+]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "multipers"
+version = "2.0.5"
+description = "Scikit-style Multiparameter persistence toolkit"
+readme = "README.md"
+authors = [
+  { name = "David Loiseaux", email = "david.loiseaux@inria.fr" },
+  { name = "Hannah Schreiber" },
+  { name = "Luis Scoccola" },
+  { name = "Mathieu Carrière" },
+]
+maintainers = [{ name = "David Loiseaux", email = "david.loiseaux@inria.fr" }]
+dependencies = [
+  #for setup.py, eventhough has to be already installed to call the setup in the first place
+  "numpy",
+  #to do `import multipers`
+  "tbb",
+  "tbb-devel",
+  "gudhi",
+  "tqdm",
+  "scipy",
+  "joblib",
+  "matplotlib",
+  #to do `pytest multipers`
+  "scikit-learn",
+  "filtration-domination",
+  "pykeops",
+  "pot"
+]
+requires-python = ">=3.10"
+license = { file = "LICENSE" }
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: Implementation :: CPython",
+  "Topic :: Scientific/Engineering :: Artificial Intelligence",
+  "Topic :: Scientific/Engineering :: Mathematics",
+  "Topic :: Scientific/Engineering :: Visualization",
+  "Topic :: Software Development :: Libraries :: Python Modules",
+  "License :: OSI Approved :: MIT License",
+]
+keywords = ["TDA", "Persistence", "Multiparameter", "sklearn"]
+
+[project.urls]
+source = "https://github.com/DavidLapous/multipers"
+download = "https://pypi.org/project/multipers/#files"
+tracker = "https://github.com/DavidLapous/multipers/issues"
+"release notes" = "https://github.com/DavidLapous/multipers/releases"
+
+# automatic search for packages
+[tool.setuptools.packages.find]
+include = ["multipers*"]
+exclude = [
+  "multipers.gudhi",
+  "multipers.multi_parameter_rank_invariant",
+  "multipers.multiparameter_module_approximation"
 ]
 
