### 2024-09-12 bb15a40
add .clang-format file for automatic formatting + minor corrections
