### 2024-09-15 DavidLapous/multipers#18 @peekxc
latex not rendering in docs 
The latex is not rendering in the documentation for me

Inspecting the source, it seems the latex source is just rendering as text, i.e. it's not a MathML / MathJax / KaTeX related issue

Related [#6773](https://github.com/openjournals/joss-reviews/issues/6773)
