### 2024-09-15 DavidLapous/multipers#18 @peekxc
Interestingly, this seems to only affect some of the docs, as the docs on the main page render fine:
