### 2024-09-15 DavidLapous/multipers#12 @peekxc
I've been able to resolve my installation issues. In particular, using a fresh Python 3.10 (or 3.11) environment + installing `multipers` the recommended way looks to now be working.
