### 2024-09-15 DavidLapous/multipers#19 @peekxc
Contributor instructions
Per the JOSS standard, there should be some community guidelines regarding contribution, reporting issues, etc. The examples in the package are great, but I didn't find either a CONTRIBUTING.md or a doc page [akin to this](https://www.pandoc.org/CONTRIBUTING.html) giving such instructions.

@DavidLapous Could you add one of the above to the repo, or show me where one already exists? It would help potential contributors immensely!

[6773](https://github.com/openjournals/joss-reviews/issues/6773)
