### 2024-09-15 #6773 @peekxc
Hey @RMeli, sorry I've been a bit busy of late. I spent a couple hours this morning and have managed to reproduce the notebooks and look a bit more through the package. I'm filling some additional issues now that I think should be met before proceeding.
