### 2024-09-16 DavidLapous/multipers#18 @DavidLapous
Dear @peekxc , thanks for your investigation!
I think it is because I activated a "lazy" update of the documentation which doesn't recompile ---and simply reads---
the notebooks in the documentation, which seems to prevent sphinx from compiling latex?
IIRC, I did that to have a faster compilation of the library in my scripts (which is already quite long),
this should be fixed by removing the lines in and recompiling the doc
https://github.com/DavidLapous/multipers/blob/3f664fff99cc7576019a698eda76f9e0efb4885f/docs/conf.py#L34-L36

I'll do that for the next release.
