+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,31 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug:**
+A clear and concise description of what the bug is.
+
+**(Minimal) example:**
+This code doesn't work/crashes:
+```python
+import multipers as mp
+[...]
+```
+which outputs 
+```
+[...]
+```
+instead of 
+```
+[...]
+```
+
+
+**Setup:**
+ - OS[e.g., macOS 14 on arm, fedora]:
+ - version [e.g. 2.0, or commit hash if compiled]:
