+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,20 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: enhancement
+assignees: ''
+
+---
+
+**Is your feature request related to a problem that you think `multipers` should handle? Please describe.**
+A clear and concise description of what the problem is, e.g., I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+Ideally, this could take the form of
+```python
+import multipers as mp
+[...]
+mp.stuff.new_function(something) # should return `stuff`
+```
