+++ multipers/gudhi/gudhi/One_critical_filtration.h
@@ -1,0 +1,1375 @@
+/*    This file is part of the Gudhi Library - https://gudhi.inria.fr/ - which is released under MIT.
+ *    See file LICENSE or go to https://gudhi.inria.fr/licensing/ for full license details.
+ *    Author(s):       David Loiseaux
+ *
+ *    Copyright (C) 2023 Inria
+ *
+ *    Modification(s):
+ *      - 2024/08 Hannah Schreiber: Generalization to all signed arithmetic types for T + doc
+ *      - YYYY/MM Author: Description of the modification
+ */
+
+/**
+ * @file One_critical_filtration.h
+ * @author David Loiseaux
+ * @brief Contains the @ref Gudhi::multi_filtration::One_critical_filtration class.
+ */
+
+#ifndef ONE_CRITICAL_FILTRATIONS_H_
+#define ONE_CRITICAL_FILTRATIONS_H_
+
+#include <algorithm>  //std::lower_bound
+#include <cmath>      //std::isnan
+#include <cstddef>    //std::size_t
+#include <cstdint>    //std::int32_t
+#include <ostream>    //std::ostream
+#include <limits>     //std::numerical_limits
+#include <stdexcept>  //std::logic_error
+#include <vector>
+
+#include <gudhi/Debug_utils.h>
+
+namespace Gudhi::multi_filtration {
+
+/**
+ * @class One_critical_filtration one_critical_filtration.h gudhi/one_critical_filtration.h
+ * @ingroup multi_filtration
+ *
+ * @brief Class encoding the apparition time, i.e., filtration value of an object 
+ * (e.g., simplex, cell, abstract algebraic generator) in the setting of 1-critical multiparameter filtrations. 
+ * The class can be used as a vector whose indices correspond to one parameter each.
+ * It also follows numpy-like broadcast semantic.
+ *
+ * @details Inherits of `std::vector<T>`. Overloads `std::numeric_limits` such that:
+ * - `std::numeric_limits<One_critical_filtration<T> >::has_infinity` returns `true`,
+ * - `std::numeric_limits<One_critical_filtration<T> >::infinity()` returns @ref One_critical_filtration<T>::inf() "",
+ * - `std::numeric_limits<One_critical_filtration<T> >::minus_infinity()` returns
+ *   @ref One_critical_filtration<T>::minus_inf() "",
+ * - `std::numeric_limits<One_critical_filtration<T> >::max()` throws,
+ * - `std::numeric_limits<One_critical_filtration<T> >::max(n)` returns a @ref One_critical_filtration<T> with `n`
+ *   parameters evaluated at value `std::numeric_limits<T>::max()`,
+ * - `std::numeric_limits<One_critical_filtration<T> >::quiet_NaN()` returns @ref One_critical_filtration<T>::nan() "".
+ *
+ * One critical simplicial filtrations are filtrations such that the lifetime of each object is a positive cone, e.g.
+ *  - \f$ \{ x \in  \mathbb R^2 : x>=(1,2)\} \f$ is valid, while
+ *  - \f$ \{ x \in  \mathbb R^2 : x>=(1,2)\} \cap \{x \in \mathbb R^2 : x>=(2,1)\} \f$ is not.
+ *
+ * If the lifetime corresponds to a union of such positive cones, the filtration is called a multi-critical filtration.
+ * For those cases, use @ref Multi_critical_filtration instead.
+ *
+ * @tparam T Arithmetic type of an entry for one parameter of the filtration value. Has to be **signed** and
+ * to implement `std::isnan(T)`, `std::numeric_limits<T>::has_quiet_NaN`, `std::numeric_limits<T>::quiet_NaN()`,
+ * `std::numeric_limits<T>::has_infinity`, `std::numeric_limits<T>::infinity()` and `std::numeric_limits<T>::max()`.
[…]
