### 2024-09-17 2d72876
Update issue templates
