### 2024-09-17 DavidLapous/multipers#19 @DavidLapous
There is a small section already at the end of the readme, and I've added templates for issues / feature requests in the commits 0e4878461ae0286efea39aa0f6e6325f9e8c61c7 and 2d7287612a494eb0ebfb922c1803a48187089d73.
You can check that by trying to open an issue. I'll add a contributing page in the documentation for the next release.
