+++ multipers/tests/test_mma.py
@@ -15,7 +15,7 @@ def test_1():
     st.insert([0, 1], [1, 1])
     mma_pymodule = st.persistence_approximation()
     assert np.array_equal(mma_pymodule[0].get_birth_list(), [[0.0, 1.0], [1.0, 0.0]])
-    assert np.array_equal(mma_pymodule[0].get_death_list(), [[np.inf]])
+    assert np.array_equal(mma_pymodule[0].get_death_list(), [[np.inf, np.inf]])
 
 
 def test_img():
