+++ multipers/multiparameter_module_approximation/approximation.h
@@ -30,13 +30,13 @@
 #include <utility>
 #include <vector>
 
-#include "gudhi/Multi_persistence/Box.h"
 #include "utilities.h"
 
 #include "debug.h"
 #include <Persistence_slices_interface.h>
 #include <gudhi/One_critical_filtration.h>
 #include <gudhi/Multi_critical_filtration.h>
+#include <gudhi/Multi_persistence/Box.h>
 #include <gudhi/Multi_persistence/Line.h>
 #include <tbb/parallel_for.h>
 
@@ -70,18 +70,14 @@ public:
   Summand<value_type> &operator[](size_t index);
 
   const Summand<value_type> &operator[](const size_t index) const;
-  template <class Barcode> void add_barcode(const Barcode &barcode);
-  template <class Filtration_type>
-  void add_barcode(
-      const Filtration_type &basepoint,
-      const std::vector<std::pair<int, std::pair<value_type, value_type>>>
-          &barcode,
-      const bool verbose, const Filtration_type &direction = {});
-  template <class Filtration_type>
-  void
-  add_barcode(const Filtration_type &basepoint,
-              const std::vector<std::pair<value_type, value_type>> &barcode,
-              const bool verbose, const Filtration_type &direction = {});
+  template <class Barcode>
+  void add_barcode(const Barcode &barcode);
+  void add_barcode(const Line<value_type> &line,
+                   const std::vector<std::pair<int, std::pair<value_type, value_type> > > &barcode,
+                   const bool threshold);
+  void add_barcode(const Line<value_type> &line,
+                   const std::vector<std::pair<value_type, value_type> > &barcode,
+                   const bool threshold);
   typename module_type::iterator begin();
   typename module_type::iterator end();
 
@@ -209,6 +205,11 @@ private:
                               const filtration_type x, const value_type delta,
                               const value_type p, const bool normalize,
                               const value_type moduleWeight) const;
+
+  void _add_bar_with_threshold(const Line<value_type> &line,
+                               const std::pair<value_type, value_type> &bar,
+                               const bool threshold_in,
+                               Summand<value_type> &summand);
 };
 
 template <typename value_type> class Summand {
@@ -322,7 +323,7 @@ public:
   }
   inline Summand<int64_t> grid_squeeze(
       const std::vector<std::vector<value_type>> &grid) const;
-  
+
 
 private:
   Gudhi::multi_filtration::Multi_critical_filtration<value_type,false> birth_corners_; // TODO : use Multi_critical_filtration
@@ -422,14 +423,13 @@ inline void __add_vineyard_trajectory_to_module(
                 << std::endl;
     if constexpr (verbose)
       std::cout << "Line basepoint " << new_line.base_point() << std::endl;
-    slicer.push_to_least_common_upper_bound(new_line);
+    slicer.push_to(new_line);
 
     slicer.vineyard_update();
     if constexpr (verbose2)
       std::cout << slicer << std::endl;
     const auto &diagram = slicer.get_flat_nodim_barcode();
-    module.add_barcode(new_line.base_point(), std::move(diagram), threshold,
-                       new_line.direction());
+    module.add_barcode(new_line, std::move(diagram), threshold);
   };
 };
 
@@ -456,7 +456,7 @@ void _rec_mma(Module<typename Filtration_value::value_type> &module,
     _rec_mma(module, basepoint_copy, grid_size, dim_to_iterate - 1, pers_copy,
              precision, threshold);
     basepoint[dim_to_iterate] += precision;
-    // current_persistence.push_to_least_common_upper_bound(Line(basepoint));
+    // current_persistence.push_to(Line(basepoint));
[…]
