### 2024-09-21 af84c3e
Added sinc convolution kernel, better typings
