### 2024-09-21 80c699c
fix: optimized speed for custom mma rpz
