+++ multipers/multiparameter_module_approximation/approximation.h
@@ -44,24 +44,26 @@ namespace Gudhi::multiparameter::mma {
 using Debug::Timer;
 using multi_filtrations::Box;
 using multi_filtrations::Line;
-template <typename T> class Module;
-template <typename T> class Summand;
+template <typename T>
+class Module;
+template <typename T>
+class Summand;
 
-void threshold_filters_list(std::vector<value_type> &filtersList,
-                            const Box<value_type> &box);
+void threshold_filters_list(std::vector<value_type> &filtersList, const Box<value_type> &box);
 
-template <typename value_type> class Module {
-public:
-  using filtration_type = Gudhi::multiparameter::multi_filtrations::
-      Finitely_critical_multi_filtration<value_type>;
+template <typename value_type>
+class Module {
+ public:
+  using filtration_type = Gudhi::multiparameter::multi_filtrations::Finitely_critical_multi_filtration<value_type>;
   using module_type = std::vector<Summand<value_type>>;
   using image_type = std::vector<std::vector<value_type>>;
-  using get_2dpixel_value_function_type = std::function<value_type(
-      const typename module_type::const_iterator,
-      const typename module_type::const_iterator, value_type, value_type)>;
-  using get_pixel_value_function_type = std::function<value_type(
-      const typename module_type::const_iterator,
-      const typename module_type::const_iterator, std::vector<value_type> &)>;
+  using get_2dpixel_value_function_type = std::function<value_type(const typename module_type::const_iterator,
+                                                                   const typename module_type::const_iterator,
+                                                                   value_type,
+                                                                   value_type)>;
+  using get_pixel_value_function_type = std::function<value_type(const typename module_type::const_iterator,
+                                                                 const typename module_type::const_iterator,
+                                                                 std::vector<value_type> &)>;
 
   Module();
   Module(Box<value_type> &box);
@@ -71,51 +73,53 @@ public:
   Summand<value_type> &operator[](size_t index);
 
   const Summand<value_type> &operator[](const size_t index) const;
-  template <class Barcode> void add_barcode(const Barcode &barcode);
+  template <class Barcode>
+  void add_barcode(const Barcode &barcode);
   template <class Filtration_type>
-  void add_barcode(
-      const Filtration_type &basepoint,
-      const std::vector<std::pair<int, std::pair<value_type, value_type>>>
-          &barcode,
-      const bool verbose, const Filtration_type &direction = {});
+  void add_barcode(const Filtration_type &basepoint,
+                   const std::vector<std::pair<int, std::pair<value_type, value_type>>> &barcode,
+                   const bool verbose,
+                   const Filtration_type &direction = {});
   template <class Filtration_type>
-  void
-  add_barcode(const Filtration_type &basepoint,
-              const std::vector<std::pair<value_type, value_type>> &barcode,
-              const bool verbose, const Filtration_type &direction = {});
+  void add_barcode(const Filtration_type &basepoint,
+                   const std::vector<std::pair<value_type, value_type>> &barcode,
+                   const bool verbose,
+                   const Filtration_type &direction = {});
   typename module_type::iterator begin();
   typename module_type::iterator end();
 
   void clean();
   void fill(const value_type precision);
 
-  std::vector<image_type> get_vectorization(
-      const value_type delta, const value_type p, const bool normalize,
-      const multi_filtrations::Box<value_type> &box,
-      unsigned int horizontalResolution, unsigned int verticalResolution);
-  std::vector<image_type>
-  get_vectorization(unsigned int horizontalResolution,
-                    unsigned int verticalResolution,
-                    get_2dpixel_value_function_type get_pixel_value) const;
-  image_type get_vectorization_in_dimension(
[…]
