### 2024-09-23 cb93a05
Add build/test status in readme
