### 2024-09-23 6b34fbd
add more debug flags in gh action
