+++ .github/workflows/python_PR.yml
@@ -9,7 +9,8 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-        os: ["ubuntu-latest", "macos-13", "macos-14",] 
+        os: ["ubuntu-latest", "macos-13", "macos-14",]
+      fail-fast: false
     runs-on: ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
