### 2024-09-24 #6773 @DavidLapous
Hey @RMeli,
@peekxc found some issues that I will fix in the next release (DavidLapous/multipers#18 and DavidLapous/multipers#19)
that I will update for the next release (so that the doc match).
We currently are merging some Gudhi upstream files,
which require some testing for a proper stable release.
