### 2024-09-24 a1a95c9
assert fix and forgoten case
