+++ multipers/gudhi/gudhi/Multi_critical_filtration.h
@@ -511,31 +511,7 @@ class Multi_critical_filtration {
    * @return false Otherwise.
    */
   bool add_generator(const Generator &x) {
-    if constexpr (co){
-      if (is_plus_inf()){
-        return false;
-      }
-      if (is_minus_inf()){
-        if (x.is_minus_inf()){
-          return false;
-        }
-        multi_filtration_[0] = x;
-        return true;
-      }
-    } else {
-      if (is_minus_inf()){
-        return false;
-      }
-      if (is_plus_inf()){
-        if (x.is_plus_inf()){
-          return false;
-        }
-        multi_filtration_[0] = x;
-        return true;
-      }
-    }
-
-    GUDHI_CHECK(x.num_parameters() == multi_filtration_[0].num_parameters() || !x.is_finite(),
+    GUDHI_CHECK(x.num_parameters() == multi_filtration_[0].num_parameters() || !is_finite() || !x.is_finite(),
                 "Cannot add a generator with different number of parameters.");
 
     std::size_t end = multi_filtration_.size();
@@ -618,8 +594,6 @@ class Multi_critical_filtration {
    * constructor or with @ref add_guaranteed_generator "", etc.
    */
   void simplify() {
-    if (is_plus_inf() || is_minus_inf()) return;
-
     std::size_t end = 0;
 
     for (std::size_t curr = 0; curr < multi_filtration_.size(); ++curr) {
@@ -884,23 +858,37 @@ class Multi_critical_filtration {
   bool _generator_can_be_added(const Generator &x, std::size_t curr, std::size_t &end) {
     if (x.empty() || x.is_nan()) return false;
 
-    if constexpr (co){
-      if (x.is_minus_inf() && end - curr != 0) return false;
-      
+    // assumes that everything between curr and end is simplified
+    // so, only multi_filtration_[curr] can be at inf or -inf.
+    if constexpr (co) {
+      if (multi_filtration_[curr].is_plus_inf() || (x.is_minus_inf() && end - curr != 0)) {
+        return false;
+      }
+      if (multi_filtration_[curr].is_minus_inf()) {
+        if (x.is_minus_inf()) {
+          return false;
+        }
+        end = curr;
+        return true;
+      }
       if (x.is_plus_inf()) {
-        if (end - curr == 1 && multi_filtration_[curr].is_plus_inf()) return false;
-        // assumes that everything between curr and end is already simplified
-        // so, if end - curr != 1, there can be no minus_inf anymore.
+        if (multi_filtration_[curr].is_plus_inf()) return false;
         end = curr;
         return true;
       }
     } else {
-      if (x.is_plus_inf() && end - curr != 0) return false;
-
+      if (multi_filtration_[curr].is_minus_inf() || (x.is_plus_inf() && end - curr != 0)) {
+        return false;
+      }
+      if (multi_filtration_[curr].is_plus_inf()) {
+        if (x.is_plus_inf()) {
+          return false;
+        }
+        end = curr;
+        return true;
+      }
       if (x.is_minus_inf()) {
-        if (end - curr == 1 && multi_filtration_[curr].is_minus_inf()) return false;
-        // assumes that everything between curr and end is already simplified
-        // so, if end - curr != 1, there can be no minus_inf anymore.
+        if (multi_filtration_[curr].is_minus_inf()) return false;
         end = curr;
         return true;
       }
