+++ .github/workflows/python_PR.yml
@@ -1,7 +1,7 @@
 name: Build, test
 
 on: 
-  [pull_request_target,workflow_dispatch,push]
+  [pull_request,workflow_dispatch,push]
 
 jobs:
   build_test:
