### 2024-09-24 1f4a76f
doc: fixed notebooks not rendering latex
