+++ docs/paper.bib
@@ -10,23 +10,6 @@
   file = {/user/dloiseau/home/Zotero/storage/WWRX5EQI/Alonso et al. - Filtration-domination in bifiltered graphs.pdf}
 }
 
-@misc{botnanIntroductionMultiparameterPersistence2022,
-  title = {An {{Introduction}} to {{Multiparameter Persistence}}},
-  author = {Botnan, Magnus Bakke and Lesnick, Michael},
-  year = {2022},
-  month = mar,
-  number = {arXiv:2203.14289},
-  eprint = {2203.14289},
-  primaryclass = {cs, math},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2203.14289},
-  urldate = {2022-06-11},
-  abstract = {In topological data analysis (TDA), one often studies the shape of data by constructing a filtered topological space, whose structure is then examined using persistent homology. However, a single filtered space often does not adequately capture the structure of interest in the data, and one is led to consider multiparameter persistence, which associates to the data a space equipped with a multiparameter filtration. Multiparameter persistence has become one of the most active areas of research within TDA, with exciting progress on several fronts. In this article, we introduce multiparameter persistence and survey some of this recent progress, with a focus on ideas likely to lead to practical applications in the near future.},
-  archiveprefix = {arXiv},
-  keywords = {Computer Science - Computational Geometry,Mathematics - Algebraic Topology,Mathematics - Representation Theory},
-  file = {/user/dloiseau/home/Zotero/storage/AVCY36LB/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/QH98BQFF/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/LPFXRBHW/2203.html}
-}
-
 @inproceedings{botnanSignedBarcodesMultiParameter2022,
   title = {Signed {{Barcodes}} for {{Multi-Parameter Persistence}} via {{Rank Decompositions}}},
   booktitle = {38th {{International Symposium}} on {{Computational Geometry}} ({{SoCG}} 2022)},
@@ -43,7 +26,7 @@
   urldate = {2022-06-14},
   isbn = {978-3-95977-227-3},
   keywords = {multi-parameter persistent homology,Topological data analysis},
-  file = {/user/dloiseau/home/Zotero/storage/PQ2SWJA6/Botnan et al. - 2022 - Signed Barcodes for Multi-Parameter Persistence vi.pdf;/user/dloiseau/home/Zotero/storage/FTU8JRV7/16027.html}
+  file = {/user/dloiseau/home/Zotero/storage/PQ2SWJA6/Botnan et al. - 2022 - Signed Barcodes for Multi-Parameter Persistence vi.pdf;/user/dloiseau/home/Zotero/storage/ZFIATDM8/Botnan et al. - 2021 - Signed barcodes for multi-parameter persistence vi.pdf;/user/dloiseau/home/Zotero/storage/FTU8JRV7/16027.html}
 }
 
 @article{caiElderRuleStaircodesAugmentedMetric2021,
@@ -97,7 +80,7 @@
   file = {/user/dloiseau/home/Zotero/storage/YG5FLRKT/Flamary et al. - 2021 - POT Python optimal transport.pdf}
 }
 
-@article{lesnickInteractiveVisualization2D2015a,
+@article{lesnickInteractiveVisualization2D2015,
   title = {Interactive Visualization of 2-{{D}} Persistence Modules},
   author = {Lesnick, Michael and Wright, Matthew},
   year = {2015},
