### 2024-09-24 DavidLapous/multipers#19 @DavidLapous
I've added a contributions page in the doc [here](https://www-sop.inria.fr/members/David.Loiseaux/doc/multipers/contributions.html). I still need to do a release to ensure that doc<->library are coherent, but this point should be fixed.
