### 2024-09-24 DavidLapous/multipers#18 @DavidLapous
Some extensions were missing to compile the latex inside the notebooks, this is fixed with commit 1f4a76f1a74b526cb2dcbae2400249ef49c35a9b.
