### 2024-09-24 ed855fb
remove assert in box
