+++ multipers/tensor/tensor.h
@@ -10,8 +10,7 @@
 namespace tensor {
 
 template <typename indices_type>
-inline std::vector<indices_type>
-compute_backward_cumprod(const std::vector<indices_type> &resolution) {
+inline std::vector<indices_type> compute_backward_cumprod(const std::vector<indices_type> &resolution) {
   constexpr bool verbose = false;
   std::vector<indices_type> cum_prod_resolution_(resolution.size());
   cum_prod_resolution_.back() = 1;
@@ -21,27 +20,24 @@ compute_backward_cumprod(const std::vector<indices_type> &resolution) {
   }
   if constexpr (verbose) {
     std::cout << "Cum resolution ";
-    for (auto c : cum_prod_resolution_)
-      std::cout << c << " ";
+    for (auto c : cum_prod_resolution_) std::cout << c << " ";
     std::cout << std::endl;
   }
   return cum_prod_resolution_;
 }
 
 template <typename dtype, typename indices_type>
-class static_tensor_view { // Python handles the construction - destruction of
-                           // the data,
-public:
-  using sparse_type =
-      std::pair<std::vector<std::vector<indices_type>>, std::vector<dtype>>;
+class static_tensor_view {  // Python handles the construction - destruction of
+                            // the data,
+ public:
+  using sparse_type = std::pair<std::vector<std::vector<indices_type>>, std::vector<dtype>>;
   static_tensor_view();
-  static_tensor_view(dtype *data_ptr,
-                     const std::vector<indices_type> &resolution)
+
+  static_tensor_view(dtype *data_ptr, const std::vector<indices_type> &resolution)
       : data_ptr_(data_ptr),
         size_(resolution.size() == 0
                   ? 0
-                  : std::accumulate(begin(resolution), end(resolution), 1,
-                                    std::multiplies<indices_type>())),
+                  : std::accumulate(begin(resolution), end(resolution), 1, std::multiplies<indices_type>())),
         resolution_(resolution)
   // cum_prod_resolution_(compute_backward_cumprod(resolution))
   {
@@ -60,9 +56,12 @@ public:
   // 	return *data_ptr;
   // }
   inline std::size_t size() const { return size_; }
+
   inline bool empty() const { return size_ == 0; }
+
   inline dtype &data_back() const { return *(data_ptr_ + size_ - 1); }
-  inline indices_type ndim() const { return resolution_.size(); }
+
+  inline std::size_t ndim() const { return resolution_.size(); }
 
   template <class oned_array_like = std::initializer_list<indices_type>>
   inline dtype &operator[](const oned_array_like &coordinates) const {
@@ -73,8 +72,7 @@ public:
     if constexpr (check) {
       if (coordinates.size() != resolution_.size()) {
         auto it = coordinates.begin();
-        for (size_t i = 0u; i < coordinates.size(); i++)
-          std::cerr << *(it++) << "/" << resolution_[i] << ", ";
+        for (size_t i = 0u; i < coordinates.size(); i++) std::cerr << *(it++) << "/" << resolution_[i] << ", ";
         std::cerr << ")" << std::endl;
         throw std::invalid_argument("Invalid coordinate dimension.");
       }
@@ -89,8 +87,7 @@ public:
         if (ci >= res) [[unlikely]] {
           std::cerr << "Crash log. Coordinates : (";
           auto it = coordinates.begin();
-          for (auto i = 0u; i < coordinates.size(); i++)
-            std::cerr << *(it++) << "/" << resolution_[i] << ", ";
+          for (auto i = 0u; i < coordinates.size(); i++) std::cerr << *(it++) << "/" << resolution_[i] << ", ";
           // for (auto [c, r] : std::views::zip(coordinates, resolution_))
[…]
