### 2024-09-26 bb007e6
feat: added deprecation warning
