### 2024-09-26 ef681e6
feat: added some checks in sm pipeline
