### 2024-09-26 fbe7cbd
fix: optimized sm tests
