+++ multipers/ml/signed_measures.py
@@ -1,6 +1,6 @@
-from collections.abc import Callable, Iterable
+from collections.abc import Callable, Iterable, Sequence
 from itertools import product
-from typing import Optional
+from typing import Optional, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -55,7 +55,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         degrees: list[int | None] = [],
         rank_degrees: list[int] = [],  # same for rank invariant
         filtration_grid: (
-            Iterable[np.ndarray]
+            Sequence[Sequence[np.ndarray]]
             # filtration values to consider. Format : [ filtration values of Fi for Fi:filtration values of parameter i]
             | None
         ) = None,
@@ -100,8 +100,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         self.normalize_filtrations = normalize_filtrations
         self.grid_strategy = grid_strategy
         self.num_parameter = None
-        self._is_input_delayed = None
-        self._möbius_inversion = _möbius_inversion
+        # self._möbius_inversion = _möbius_inversion
         self._reconversion_grid = None
         self.expand = expand
         # will only refit the grid if filtration_grid has never been given.
@@ -109,7 +108,6 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         self.seed = seed
         self.fit_fraction = fit_fraction
         self._transform_st = None
-        self._to_simplex_tree: Callable
         self.out_resolution = out_resolution
         self.individual_grid = individual_grid
         self.enforce_null_mass = enforce_null_mass
@@ -135,59 +133,68 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         assert (
             not mp.slicer.is_slicer(first) or self.expand is None
         ), "Cannot expand slicers."
-        self._is_minpres = mp.slicer.is_slicer(first) and isinstance(
-            first, Union[tuple, list]
+        self._is_minpres = mp.slicer.is_slicer(first) and (
+            (first.is_minpres)
+            or (isinstance(first, Union[tuple, list]) and first[0].is_minpres)
         )
-        assert not (
-            self._is_minpres and self.minpres_degrees is not None
-        ), "Input is already a minpres. Cannot reduce again."
 
     def _infer_filtration(self, X):
         indices = np.random.choice(
             len(X), min(int(self.fit_fraction * len(X)) + 1, len(X)), replace=False
         )
 
-        def get_st_filtration(x) -> np.ndarray:
-            return self._to_simplex_tree(x).get_filtration_grid(grid_strategy="exact")
+        # def get_st_filtration(x) -> np.ndarray:
+        #     return x.get_filtration_grid(grid_strategy="exact")
 
-        filtrations: list = Parallel(n_jobs=self.n_jobs, backend="threading")(
-            delayed(get_st_filtration)(x) for x in (X[idx] for idx in indices)
+        # filtrations: list = Parallel(n_jobs=self.n_jobs, backend="threading")(
+        #     delayed(get_st_filtration)(x) for x in (X[idx] for idx in indices)
+        # )
+        ## ax, num_x
+        filtrations = tuple(
+            tuple(
+                reduce_grid(x, strategy="exact")
+                for x in (X[ax][idx] for idx in indices)
+                for ax in range(self._num_axis)
+            )
         )
-        num_parameters = len(filtrations[0])
+        num_parameters = len(filtrations[0][0])
+        assert num_parameters == self.num_parameters
+
         filtrations_values = [
-            np.unique(np.concatenate([x[i] for x in filtrations]))
-            for i in range(num_parameters)
+            [
[…]
