### 2024-09-26 235af2e
fix: sm pipelines
