+++ multipers/ml/signed_measures.py
@@ -139,6 +139,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         )
 
     def _infer_filtration(self, X):
+        self.num_parameters = X[0][0].num_parameters
         indices = np.random.choice(
             len(X), min(int(self.fit_fraction * len(X)) + 1, len(X)), replace=False
         )
@@ -153,12 +154,14 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         filtrations = tuple(
             tuple(
                 reduce_grid(x, strategy="exact")
-                for x in (X[ax][idx] for idx in indices)
-                for ax in range(self._num_axis)
+                for x in (X[idx][ax] for idx in indices)
             )
+            for ax in range(self._num_axis)
         )
         num_parameters = len(filtrations[0][0])
-        assert num_parameters == self.num_parameters
+        assert (
+            num_parameters == self.num_parameters
+        ), f"Internal error, got {num_parameters=} and {self.num_parameters=}"
 
         filtrations_values = [
             [
@@ -191,7 +194,6 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
         self._params_check()
         self._input_checks(X)
 
-        self.num_parameter = X[0][0].num_parameters
         if isinstance(self.resolution, int):
             self.resolution = [self.resolution] * self.num_parameters
 
@@ -237,8 +239,8 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
     def transform1(
         self,
         simplextree,
-        filtration_grid,
-        _reconversion_grid,
+        ax,
+        # _reconversion_grid,
         thread_id: str = "",
     ):
         # st = mp.SimplexTreeMulti(st, num_parameters=st.num_parameters)  # COPY
@@ -246,11 +248,18 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
             filtration_grid = reduce_grid(
                 simplextree, strategy=self.grid_strategy, resolution=self.resolution
             )
+            mass_default = self._default_mass_location[ax]
             if self.enforce_null_mass:
                 filtration_grid = [
                     np.concatenate([f, [d]], axis=0)
-                    for f, d in zip(filtration_grid, self._default_mass_location)
+                    for f, d in zip(filtration_grid, mass_default)
                 ]
+            _reconversion_grid = filtration_grid
+        else:
+            filtration_grid = self.filtration_grid[ax]
+            _reconversion_grid = self._reconversion_grid[ax]
+            mass_default = self._default_mass_location[ax]
+
         st = simplextree.grid_squeeze(
             filtration_grid=filtration_grid, coordinate_values=True
         )
@@ -265,7 +274,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
                 st,
                 degrees=[None],
                 plot=self.plot,
-                mass_default=self._default_mass_location,
+                mass_default=mass_default,
                 invariant="euler",
                 # thread_id=thread_id,
                 backend=self.backend,
@@ -285,7 +294,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
             mp.signed_measure(
                 st,
                 degrees=int_degrees,
-                mass_default=self._default_mass_location,
+                mass_default=mass_default,
                 plot=self.plot,
                 invariant="hilbert",
                 thread_id=thread_id,
@@ -307,7 +316,7 @@ class SimplexTree2SignedMeasure(BaseEstimator, TransformerMixin):
             mp.signed_measure(
                 st,
                 degrees=self.rank_degrees,
-                mass_default=self._default_mass_location,
+                mass_default=mass_default,
[…]
