### 2024-09-26 00abee2
fix: sm pipeline
