### 2024-09-26 fc68fcd
fix: add deprecation warning
