### 2024-09-26 44e5acd
fix: deprecate SimplexTree2SignedMeasure
