### 2024-09-26 2fb3e71
fix: deprecate SimplexTree2MMA
