### 2024-09-26 72ec332
fix: add assert
