### 2024-09-26 9804099
feat: re-added parallelizable sm pipeline
