+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "multipers"
-version = "2.0.5"
+version = "2.1.0"
 description = "Multiparameter Topological Persistence for Machine Learning"
 readme = "README.md"
 authors = [
