### 2024-09-27 #6773 @DavidLapous
[The latest release](https://github.com/DavidLapous/multipers/releases/tag/v2.1.0) should have fixed these issues. I've also updated the website accordingly.
I've also put the repo on zenodo; with the DOI:
 [![DOI](https://zenodo.org/badge/438254685.svg)](https://zenodo.org/doi/10.5281/zenodo.13844727)
