### 2024-09-27 731ce19
first drafts
