+++ setup.py
@@ -79,6 +79,7 @@ templated_cython_modules = [
     "mma_structures.pyx",
     "simplex_tree_multi.pyx",
     "slicer.pyx",
+    "io.pxd",
 ]
 
 ## generates some parameter files (Tempita fails with python<3.12)
@@ -147,8 +148,8 @@ extensions = [
         ],
         language="c++",
         extra_compile_args=[
-            "-O3",  # -Ofast disables infinity values for filtration values
-            # "-g",
+            "-Og",  # -Ofast disables infinity values for filtration values
+            "-g",
             # "-march=native",
             "-std=c++20",  # Windows doesn't support this yet. TODO: Wait.
             # "-fno-aligned-new", # Uncomment this if you have trouble compiling on macos.
