+++ multipers/gudhi/scc_io.h
@@ -1,0 +1,302 @@
+#ifndef MULTIPERS_SCC_IO_H
+#define MULTIPERS_SCC_IO_H
+
+#include <cassert>
+#include <cstddef>
+#include <ostream>
+#include <fstream>
+#include <iostream>
+#include <stdexcept>
+#include <string>
+#include <vector>
+#include <limits>
+#include <iomanip>
+#include <cmath>
+
+// #include "truc.h"
+
+// using Gudhi::multiparameter::interface::Truc;
+
+bool is_comment_or_empty_line(const std::string& line) {
+  size_t current = line.find_first_not_of(' ', 0);
+  if (current == std::string::npos) return true;  // is empty line
+  if (line[current] == '#') return true;  // is comment
+  return false;
+}
+
+template <class Slicer>
+Slicer read_scc_file(std::string inFilePath, bool isRivetCompatible = false){
+  std::string line;
+  std::ifstream file(inFilePath);
+  Slicer slicer;
+
+  auto error = [&file](std::string msg){
+    file.close();
+    throw std::invalid_argument(msg);
+  };
+
+  if (file.is_open()) {
+    while (getline(file, line, '\n') && is_comment_or_empty_line(line));
+    if (!file) error("Empty file!");
+
+    if (isRivetCompatible && line.compare("firep") != 0) error("Wrong file format. Should start with 'firep'.");
+    if (!isRivetCompatible && line.compare("scc2020") != 0) error("Wrong file format. Should start with 'scc2020'.");
+
+    while (getline(file, line, '\n') && is_comment_or_empty_line(line));
+    if (!file) error("Premature ending of the file. Stops before numbers of parameters.");
+    
+
+
+    // std::vector<ID_handle> data;
+    // unsigned int id = 0;
+    // double timestamp;
+    // lineType type;
+
+    // while (getline(file, line, '\n') && read_operation(line, data, timestamp) == COMMENT);
+    // double lastTimestamp = timestamp;
+    // // first operation has to be an insertion.
+    // zp.insert_face(id, data, 0, timestamp);
+
+    // while (getline(file, line, '\n')) {
+    //   type = read_operation(line, data, timestamp);
+    //   if (type != COMMENT && lastTimestamp != timestamp) {
+    //     lastTimestamp = timestamp;
+    //   }
+
+    //   if (type == INCLUSION) {
+    //     ++id;
+    //     int dim = data.size() == 0 ? 0 : data.size() - 1;
+    //     zp.insert_face(id, data, dim, timestamp);
+    //   } else if (type == REMOVAL) {
+    //     ++id;
+    //     zp.remove_face(data[0], timestamp);
+    //   }
+    // }
+
+    file.close();
+  } else {
+    std::cerr << "Unable to open input file." << std::endl;
+    file.setstate(std::ios::failbit);
+  }
+
+  return slicer;
+};
+
+template <class Slicer>
+void write_scc_file(std::string outFilePath,
+                    const Slicer& slicer,
+                    int numberOfParameters = -1,
+                    int degree = -1,
+                    bool rivetCompatible = false,
+                    bool IgnoreLastGenerators = false,
+                    bool stripComments = false,
+                    bool reverse = false) {
+  if (numberOfParameters < 0){
+    numberOfParameters = slicer.num_parameters();
+  }
+  assert(numberOfParameters > 0 && "Invalid number of parameters!");
+
+  std::ofstream file(outFilePath);
+
+  if (rivetCompatible) file << "firep\n";
+  else file << "scc2020\n";
+
[…]
