### 2024-09-29 be1a1bc
tests: added test for distance matrix filling
