+++ pyproject.toml
@@ -1,7 +1,8 @@
 [build-system]
 requires = [
+  "numpy",
   "Cython",
-  "setuptools"
+  "setuptools",
 ]
 build-backend = "setuptools.build_meta"
 
@@ -16,15 +17,14 @@ authors = [
 ]
 maintainers = [{ name = "David Loiseaux", email = "david.lapous@proton.me" }]
 dependencies = [
-  #for setup.py, eventhough has to be already installed to call the setup in the first place
+  #`import multipers`
   "numpy",
-  #to do `import multipers`
   "gudhi",
   "tqdm",
   "scipy",
   "joblib",
   "matplotlib",
-  #to do `pytest multipers`
+  #`pytest multipers`
   "scikit-learn",
   "filtration-domination",
   "pykeops",
