### 2024-09-30 0419e51
cpp scc read
