+++ multipers/gudhi/scc_io.h
@@ -1,6 +1,7 @@
 #ifndef MULTIPERS_SCC_IO_H
 #define MULTIPERS_SCC_IO_H
 
+#include <algorithm>
 #include <cassert>
 #include <cstddef>
 #include <ostream>
@@ -13,29 +14,27 @@
 #include <iomanip>
 #include <cmath>
 
-// #include "truc.h"
-
-// using Gudhi::multiparameter::interface::Truc;
-
-bool is_comment_or_empty_line(const std::string& line) {
-  size_t current = line.find_first_not_of(' ', 0);
-  if (current == std::string::npos) return true;  // is empty line
-  if (line[current] == '#') return true;  // is comment
-  return false;
-}
-
 template <class Slicer>
-Slicer read_scc_file(std::string inFilePath, bool isRivetCompatible = false){
+Slicer read_scc_file(const std::string& inFilePath, bool isRivetCompatible = false, bool isReversed = false){
+  using Filtration_value = typename Slicer::Filtration_value;
+
   std::string line;
   std::ifstream file(inFilePath);
   Slicer slicer;
-
-  auto error = [&file](std::string msg){
-    file.close();
-    throw std::invalid_argument(msg);
-  };
+  unsigned int numberOfParameters;
 
   if (file.is_open()) {
+    auto error = [&file](std::string msg){
+      file.close();
+      throw std::invalid_argument(msg);
+    };
+    auto is_comment_or_empty_line = [](const std::string& line)->bool{
+      size_t current = line.find_first_not_of(' ', 0);
+      if (current == std::string::npos) return true;  // is empty line
+      if (line[current] == '#') return true;  // is comment
+      return false;
+    };
+
     while (getline(file, line, '\n') && is_comment_or_empty_line(line));
     if (!file) error("Empty file!");
 
@@ -45,33 +44,106 @@ Slicer read_scc_file(std::string inFilePath, bool isRivetCompatible = false){
     while (getline(file, line, '\n') && is_comment_or_empty_line(line));
     if (!file) error("Premature ending of the file. Stops before numbers of parameters.");
     
+    if (isRivetCompatible){
+      numberOfParameters = 2;
+      getline(file, line, '\n');  //second rivet label
+    } else {
+      std::size_t current = line.find_first_not_of(' ', 0);
+      std::size_t next = line.find_first_of(' ', current);
+      numberOfParameters = std::stoi(line.substr(current, next - current));
+    }
 
+    while (getline(file, line, '\n') && is_comment_or_empty_line(line));
+    if (!file) error("Premature ending of the file. Not a single cell was specified.");
 
-    // std::vector<ID_handle> data;
-    // unsigned int id = 0;
-    // double timestamp;
-    // lineType type;
+    std::vector<unsigned int> counts;
+    unsigned int numberOfCells = 0;
+    counts.reserve(line.size());
+    std::size_t current = line.find_first_not_of(' ', 0);
+    while (current != std::string::npos) {
+      std::size_t next = line.find_first_of(' ', current);
+      counts.push_back(std::stoi(line.substr(current, next - current)));
+      numberOfCells += counts.back();
+      current = line.find_first_not_of(' ', next);
+    }
 
-    // while (getline(file, line, '\n') && read_operation(line, data, timestamp) == COMMENT);
-    // double lastTimestamp = timestamp;
-    // // first operation has to be an insertion.
-    // zp.insert_face(id, data, 0, timestamp);
+    std::size_t dimIt = 0;
+    while (dimIt < counts.size() && counts[dimIt] == 0) ++dimIt;
 
-    // while (getline(file, line, '\n')) {
-    //   type = read_operation(line, data, timestamp);
[…]
