### 2024-10-02 DavidLapous/multipers#19 @peekxc
Thanks @DavidLapous! Looks good
