### 2024-10-02 84f4a00
tests: workaround for macos arm
