### 2024-10-03 04b60dc
feat!: changed the default to non-vine (+ warning+copy if necessary)
