### 2024-10-03 0b172ce
fix: warn in stderr
