### 2024-10-03 4a9a721
fix: point cloud pipeline was ignoring reduce in some cases
