### 2024-10-03 5828019
feat: bounding box from stuff
