### 2024-10-04 34cc39b
fix: bad defaults
