### 2024-10-04 e0570d5
feat: better __init__
