+++ multipers/__init__.py
@@ -1,13 +1,31 @@
+from importlib.metadata import version as _version
+
+__version__ = _version("multipers")
 # Doc
-import multipers.io
-import multipers.multiparameter_module_approximation
-import multipers.simplex_tree_multi
-import multipers.slicer
-import multipers.data
-import multipers.grids
+from multipers import (
+    data,
+    grids,
+    io,
+    multiparameter_module_approximation,
+    simplex_tree_multi,
+    slicer,
+)
 from multipers._signed_measure_meta import signed_measure
 from multipers._slicer_meta import Slicer
 from multipers.multiparameter_module_approximation import module_approximation
 
 # Shortcuts
 from multipers.simplex_tree_multi import SimplexTreeMulti
+
+__all__ = [
+    "signed_measure",
+    "module_approximation",
+    "Slicer",
+    "SimplexTreeMulti",
+    "data",
+    "grids",
+    "io",
+    "multiparameter_module_approximation",
+    "slicer",
+    "simplex_tree_multi",
+]
