### 2024-10-04 b203952
bench: added gudhi slices
