+++ multipers/gudhi/gudhi/Persistence_matrix/columns/unordered_set_column.h
@@ -31,22 +31,22 @@
 #include <unordered_set>
 #endif
 
-#include <gudhi/Persistence_matrix/allocators/cell_constructors.h>
+#include <gudhi/Persistence_matrix/allocators/entry_constructors.h>
 
 namespace Gudhi {
 namespace persistence_matrix {
 
 // For unordered_set container. Outside of Unordered_set_column because of a msvc bug who can't compile properly
 // unordered_flat_set if the hash method is nested.
-template <class Cell>
-struct CellPointerHash
+template <class Entry>
+struct EntryPointerHash
 {
-  size_t operator()(const Cell* c) const { return std::hash<Cell>()(*c); }
+  size_t operator()(const Entry* c) const { return std::hash<Entry>()(*c); }
 };
-template <class Cell>
-struct CellPointerEq
+template <class Entry>
+struct EntryPointerEq
 {
-  bool operator()(const Cell* c1, const Cell* c2) const { return *c1 == *c2; }
+  bool operator()(const Entry* c1, const Entry* c2) const { return *c1 == *c2; }
 };
 
 /**
@@ -55,12 +55,12 @@ struct CellPointerEq
  *
  * @brief Column class following the @ref PersistenceMatrixColumn concept.
  *
- * Column based on an unordered set structure. The cells are not ordered, but only non-zero values
- * are stored uniquely in the underlying container. When adding a cell range into it, the given cell range
+ * Column based on an unordered set structure. The entries are not ordered, but only non-zero values
+ * are stored uniquely in the underlying container. When adding an entry range into it, the given entry range
  * also does not need to be ordered (contrary to most other column types).
  *
  * @tparam Master_matrix An instantiation of @ref Matrix from which all types and options are deduced.
- * @tparam Cell_constructor Factory of @ref Cell classes.
+ * @tparam Entry_constructor Factory of @ref Entry classes.
  */
 template <class Master_matrix>
 class Unordered_set_column : public Master_matrix::Row_access_option,
@@ -73,21 +73,21 @@ class Unordered_set_column : public Master_matrix::Row_access_option,
   using ID_index = typename Master_matrix::ID_index;
   using Dimension = typename Master_matrix::Dimension;
   using Field_element = typename Master_matrix::Element;
-  using Cell = typename Master_matrix::Matrix_cell;
+  using Entry = typename Master_matrix::Matrix_entry;
   using Column_settings = typename Master_matrix::Column_settings;
 
  private:
   using Field_operators = typename Master_matrix::Field_operators;
-  using Cell_constructor = typename Master_matrix::Cell_constructor;
+  using Entry_constructor = typename Master_matrix::Entry_constructor;
 
-  struct CellPointerComp {
-    bool operator()(const Cell* c1, const Cell* c2) const { return *c1 < *c2; }
+  struct EntryPointerComp {
+    bool operator()(const Entry* c1, const Entry* c2) const { return *c1 < *c2; }
   };
 
 #if BOOST_VERSION >= 108100
-  using Column_support = boost::unordered_flat_set<Cell*, CellPointerHash<Cell>, CellPointerEq<Cell>>;
+  using Column_support = boost::unordered_flat_set<Entry*, EntryPointerHash<Entry>, EntryPointerEq<Entry>>;
 #else
-  using Column_support = std::unordered_set<Cell*, CellPointerHash<Cell>, CellPointerEq<Cell>>;
+  using Column_support = std::unordered_set<Entry*, EntryPointerHash<Entry>, EntryPointerEq<Entry>>;
 #endif
 
  public:
@@ -137,30 +137,32 @@ class Unordered_set_column : public Master_matrix::Row_access_option,
[…]
