+++ multipers/gudhi/gudhi/Persistence_matrix/columns/intrusive_list_column.h
@@ -25,7 +25,7 @@
 
 #include <boost/intrusive/list.hpp>
 
-#include <gudhi/Persistence_matrix/allocators/cell_constructors.h>
+#include <gudhi/Persistence_matrix/allocators/entry_constructors.h>
 #include <gudhi/Persistence_matrix/columns/column_utilities.h>
 
 namespace Gudhi {
@@ -37,11 +37,11 @@ namespace persistence_matrix {
  *
  * @brief Column class following the @ref PersistenceMatrixColumn concept.
  *
- * Column based on a intrusive list structure. The cells are always ordered by row index and only non-zero values
+ * Column based on a intrusive list structure. The entries are always ordered by row index and only non-zero values
  * are stored uniquely in the underlying container.
  *
  * @tparam Master_matrix An instantiation of @ref Matrix from which all types and options are deduced.
- * @tparam Cell_constructor Factory of @ref Cell classes.
+ * @tparam Entry_constructor Factory of @ref Entry classes.
  */
 template <class Master_matrix>
 class Intrusive_list_column : public Master_matrix::Row_access_option,
@@ -54,15 +54,15 @@ class Intrusive_list_column : public Master_matrix::Row_access_option,
   using ID_index = typename Master_matrix::ID_index;
   using Dimension = typename Master_matrix::Dimension;
   using Field_element = typename Master_matrix::Element;
-  using Cell = typename Master_matrix::Matrix_cell;
+  using Entry = typename Master_matrix::Matrix_entry;
   using Column_settings = typename Master_matrix::Column_settings;
 
  private:
   using Field_operators = typename Master_matrix::Field_operators;
   using Column_support =
-      boost::intrusive::list<Cell, boost::intrusive::constant_time_size<false>,
+      boost::intrusive::list<Entry, boost::intrusive::constant_time_size<false>,
                              boost::intrusive::base_hook<typename Master_matrix::Base_hook_matrix_list_column> >;
-  using Cell_constructor = typename Master_matrix::Cell_constructor;
+  using Entry_constructor = typename Master_matrix::Entry_constructor;
 
  public:
   using iterator = typename Column_support::iterator;
@@ -117,20 +117,22 @@ class Intrusive_list_column : public Master_matrix::Row_access_option,
   reverse_iterator rend() noexcept;
   const_reverse_iterator rend() const noexcept;
 
-  template <class Cell_range>
-  Intrusive_list_column& operator+=(const Cell_range& column);
+  template <class Entry_range>
+  Intrusive_list_column& operator+=(const Entry_range& column);
   Intrusive_list_column& operator+=(Intrusive_list_column& column);
 
   Intrusive_list_column& operator*=(const Field_element& val);
 
   // this = v * this + column
-  template <class Cell_range>
-  Intrusive_list_column& multiply_target_and_add(const Field_element& val, const Cell_range& column);
+  template <class Entry_range>
+  Intrusive_list_column& multiply_target_and_add(const Field_element& val, const Entry_range& column);
   Intrusive_list_column& multiply_target_and_add(const Field_element& val, Intrusive_list_column& column);
   // this = this + column * v
-  template <class Cell_range>
-  Intrusive_list_column& multiply_source_and_add(const Cell_range& column, const Field_element& val);
+  template <class Entry_range>
+  Intrusive_list_column& multiply_source_and_add(const Entry_range& column, const Field_element& val);
   Intrusive_list_column& multiply_source_and_add(Intrusive_list_column& column, const Field_element& val);
+
+  void push_back(const Entry& entry);
 
[…]

[…]
