### 2024-10-05 511b28c
fix: kcritical conversion in mma
