### 2024-10-06 c859167
fix: smaller default threshold to rips
