### 2024-10-07 e8d6477
factorize some code
