### 2024-10-11 f67ca6f
c++ io part1
