+++ multipers/gudhi/truc.h
@@ -66,7 +66,7 @@ public:
   } // needs to be iterable (begin, end, size)
   inline int dimension(std::size_t i) const { return generator_dimensions[i]; };
   inline friend std::ostream &operator<<(std::ostream &stream,
-                                         PresentationStructure &structure) {
+                                         const PresentationStructure &structure) {
     stream << "Boundary:\n";
     stream << "{";
     for (const auto &stuff : structure.generators) {
@@ -248,6 +248,47 @@ public:
     return *this;
   };
 
+  template<bool ignore_inf>
+  std::vector<std::pair<int, std::vector<unsigned int> > > get_current_boundary_matrix(){
+    std::vector<std::size_t> permutation(generator_order.size());
+    std::iota(permutation.begin(), permutation.end(), 0);
+    if constexpr (ignore_inf) {
+      permutation.erase(std::remove_if(permutation.begin(),
+                                       permutation.end(),
+                                       [&](std::size_t val) {
+                                         return filtration_container[val] == MultiFiltration::Generator::T_inf;
+                                       }),
+                        permutation.end());
+      std::sort(permutation.begin(), permutation.end());
+    }
+    std::sort(permutation.begin(), permutation.end(),
+              [&](std::size_t i, std::size_t j) {
+                if (structure.dimension(i) > structure.dimension(j))
+                  return false;
+                if (structure.dimension(i) < structure.dimension(j))
+                  return true;
+                return filtration_container[i] < filtration_container[j];
+              });
+
+    std::vector<std::pair<int, std::vector<unsigned int> > > matrix(permutation.size());
+
+    std::vector<std::size_t> permutationInv(generator_order.size());
+    std::size_t newPos = 0;
+    for (std::size_t oldPos : permutation) {
+      permutationInv[oldPos] = newPos;
+      auto& boundary = matrix[newPos].second;
+      boundary.resize(structure[oldPos].size());
+      for (std::size_t j = 0; j < structure[oldPos].size(); ++j) {
+        boundary[j] = permutationInv[structure[oldPos][j]];
+      }
+      std::sort(boundary.begin(), boundary.end());
+      matrix[newPos].first = structure.dimension(oldPos);
+      ++newPos;
+    }
+
+    return matrix;
+  }
+
   inline std::size_t num_generators() const { return structure.size(); }
   inline std::size_t num_parameters() const {
     return num_generators() == 0
@@ -319,6 +360,7 @@ public:
         }
     }
     if constexpr (false) {
+      std::cout << structure << std::endl;
       std::cout << "[";
       for (auto i : out_gen_order) {
 
@@ -846,6 +888,24 @@ public:
           return Gudhi::multi_persistence::Line<value_type>(bpdir.first, bpdir.second);
         },
         bp_dirs);
+  }
+
+  void build_from_scc_file(const std::string &inFilePath,
+                           bool isRivetCompatible = false,
+                           bool isReversed = false,
+                           int shiftDimensions = 0) {
+    *this = read_scc_file<Truc>(inFilePath, isRivetCompatible, isReversed, shiftDimensions);
+  }
+
+  void write_to_scc_file(const std::string &outFilePath,
+                         int numberOfParameters = -1,
+                         int degree = -1,
+                         bool rivetCompatible = false,
+                         bool IgnoreLastGenerators = false,
+                         bool stripComments = false,
+                         bool reverse = false) {
+    write_scc_file(
+        outFilePath, *this, numberOfParameters, degree, rivetCompatible, IgnoreLastGenerators, stripComments, reverse);
   }
 
 public:

[…]
