### 2024-10-11 64f063e
cpp io part2
