### 2024-10-14 9e24c16
cpp io part3
