### 2024-10-14 7abb613
cpp io part4
