### 2024-10-15 #6773 @RMeli
@DavidLapous while I go through the paper one final time, can you please create a new release and a corresponding archive?

Please also have a look at the DOIs missing above. For PyTorch there is a citation I think; the other items that look like titles of the paper also can be found at leas on arXiv, so a DOI should be added.
