### 2024-10-15 #6773 @DavidLapous
OK, no problem! Thanks for the update @RMeli !
 - For the release, I just need to take care of a pull request, and check that it is stable. I'll do that ASAP.
 - I'll also update the DOIs; do you want me to replace published articles with no DOIs (e.g., NeurIPS papers) with their draft version on arxiv so that they have a DOI?
