### 2024-10-15 #6773 @RMeli
>  do you want me to replace published articles with no DOIs (e.g., NeurIPS papers) with their draft version on arxiv so that they have a DOI

Ah, I did not realise these publications had no DOIs for the actual publication. I'll inquire internally, and let you know.
