### 2024-10-16 9624213
fix: squeezed slicers should be unsqueezed before dumped
