+++ multipers/_slicer_meta.py
@@ -113,6 +113,7 @@ def Slicer(
     is_kcritical: Optional[bool] = None,
     column_type: Optional[_column_type] = None,
     max_dim: Optional[int] = None,
+    return_type_only: bool = False,
 ) -> mps.Slicer_type:
     """
     Given a simplextree or blocks (a.k.a scc for python),
@@ -155,6 +156,8 @@ def Slicer(
     _Slicer = mps.get_matrix_slicer(
         is_vineyard=vineyard, is_k_critical=is_kcritical, dtype=dtype, col=column_type
     )
+    if return_type_only:
+        return _Slicer
     if st is None:
         return _Slicer()
     elif mps.is_slicer(st):
@@ -183,9 +186,9 @@ You can try using `multipers.slicer.to_simplextree`."""
         )
     else:
         filtration_grid = None
-        if max_dim is not None: #no test for simplex tree?
+        if max_dim is not None:  # no test for simplex tree?
             st.prune_above_dimension(max_dim)
-        if isinstance(st, str): #is_kcritical should be false
+        if isinstance(st, str):  # is_kcritical should be false
             slicer = _Slicer()._build_from_scc_file(st)
         else:
             if is_simplextree_multi(st):
