### 2024-10-16 184d6f2
feat: add test for slicer-scc
