+++ multipers/gudhi/mma_interface_matrix.h
@@ -28,34 +28,27 @@
 
 namespace Gudhi::multiparameter::interface {
 
-template <Gudhi::persistence_matrix::Column_types column_type =
-              Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
-struct Multi_persistence_options
-    : Gudhi::persistence_matrix::Default_options<column_type, true> {
+template <Gudhi::persistence_matrix::Column_types column_type = Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
+struct Multi_persistence_options : Gudhi::persistence_matrix::Default_options<column_type, true> {
   static const bool has_matrix_maximal_dimension_access = false;
   static const bool has_column_pairings = true;
   static const bool has_vine_update = true;
   static const bool can_retrieve_representative_cycles = true;
 };
 
-template <Gudhi::persistence_matrix::Column_types column_type =
-              Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
-struct Multi_persistence_Clement_options
-    : Gudhi::persistence_matrix::Default_options<column_type, true> {
+template <Gudhi::persistence_matrix::Column_types column_type = Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
+struct Multi_persistence_Clement_options : Gudhi::persistence_matrix::Default_options<column_type, true> {
   static const bool has_matrix_maximal_dimension_access = false;
   static const bool has_column_pairings = true;
   static const bool has_vine_update = true;
   static const bool is_of_boundary_type = false;
-  static const Gudhi::persistence_matrix::Column_indexation_types
-      column_indexation_type =
-          Gudhi::persistence_matrix::Column_indexation_types::POSITION;
+  static const Gudhi::persistence_matrix::Column_indexation_types column_indexation_type =
+      Gudhi::persistence_matrix::Column_indexation_types::POSITION;
   static const bool can_retrieve_representative_cycles = true;
 };
 
-template <Gudhi::persistence_matrix::Column_types column_type =
-              Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
-struct No_vine_multi_persistence_options
-    : Gudhi::persistence_matrix::Default_options<column_type, true> {
+template <Gudhi::persistence_matrix::Column_types column_type = Gudhi::persistence_matrix::Column_types::INTRUSIVE_SET>
+struct No_vine_multi_persistence_options : Gudhi::persistence_matrix::Default_options<column_type, true> {
   static const bool has_matrix_maximal_dimension_access = false;
   static const bool has_column_pairings = true;
   static const bool has_vine_update = false;
@@ -63,8 +56,7 @@ struct No_vine_multi_persistence_options
 
 template <class Matrix_options, class Boundary_matrix_type>
 class Persistence_backend_matrix {
-
-public:
+ public:
   using matrix_type = Gudhi::persistence_matrix::Matrix<Matrix_options>;
   using cycle_type = typename matrix_type::Cycle;
   static const bool is_vine = Matrix_options::has_vine_update;
@@ -76,27 +68,21 @@ public:
   using Index = typename matrix_type::Index;
   using dimension_type = typename matrix_type::Dimension;
 
-  class Barcode_iterator
-      : public boost::iterator_facade<Barcode_iterator, const bar &,
-                                      boost::forward_traversal_tag> {
-  public:
-    Barcode_iterator(const typename matrix_type::Barcode *barcode,
-                     const std::vector<std::size_t> *inv)
-        : barcode_(barcode->size() == 0 ? nullptr : barcode),
-          perm_(barcode->size() == 0 ? nullptr : inv), currPos_(0) {
[…]
