### 2024-10-17 bb6ae18
feat: Gudhi cohomology. TODO: TEST
