+++ multipers/gudhi/truc.h
@@ -23,24 +23,22 @@
 #include "scc_io.h"
 
 namespace Gudhi::multiparameter::interface {
+using index_type = std::uint32_t;
 
 class PresentationStructure {
-public:
+ public:
   PresentationStructure() {}
+
   /* SimplicialStructure &operator=(const SimplicialStructure &) = default; */
 
-  PresentationStructure(
-      const std::vector<std::vector<unsigned int>> &generators,
-      const std::vector<int> &generator_dimensions)
-      : generators(generators), generator_dimensions(generator_dimensions),
-        num_vertices_(0) {
+  PresentationStructure(const std::vector<std::vector<index_type>> &generators,
+                        const std::vector<int> &generator_dimensions)
+      : generators(generators), generator_dimensions(generator_dimensions), num_vertices_(0) {
     for (const auto &stuff : generator_dimensions) {
-      if (stuff == 0)
-        num_vertices_++;
+      if (stuff == 0) num_vertices_++;
     }
     max_dimension_ = generator_dimensions.size() > 0
-                         ? *std::max_element(generator_dimensions.begin(),
-                                             generator_dimensions.end())
+                         ? *std::max_element(generator_dimensions.begin(), generator_dimensions.end())
                          : 0;
   };
 
@@ -49,6 +47,7 @@ public:
         generator_dimensions(other.generator_dimensions),
         num_vertices_(other.num_vertices_),
         max_dimension_(other.max_dimension_) {}
+
   /* PresentationStructure &operator=(const PresentationStructure &other) { */
   /*   generators = other.generators; */
   /*   generator_dimensions = other.generator_dimensions; */
@@ -60,180 +59,181 @@ public:
 
   const std::vector<unsigned int> &operator[](std::size_t i) const {
     return generators[i];
-  } // needs to be iterable (begin, end, size)
+  }  // needs to be iterable (begin, end, size)
+
   std::vector<unsigned int> &operator[](std::size_t i) {
     return generators[i];
-  } // needs to be iterable (begin, end, size)
+  }  // needs to be iterable (begin, end, size)
+
   inline int dimension(std::size_t i) const { return generator_dimensions[i]; };
-  inline friend std::ostream &operator<<(std::ostream &stream,
-                                         const PresentationStructure &structure) {
+
+  inline friend std::ostream &operator<<(std::ostream &stream, const PresentationStructure &structure) {
     stream << "Boundary:\n";
     stream << "{";
     for (const auto &stuff : structure.generators) {
       stream << "{";
-      for (auto truc : stuff)
-        stream << truc << ", ";
+      for (auto truc : stuff) stream << truc << ", ";
 
-      if (!stuff.empty())
-        stream << "\b" << "\b ";
+      if (!stuff.empty()) stream << "\b" << "\b ";
 
       stream << "},\n";
     }
     stream << "}\n";
     stream << "Degrees: (max " << structure.max_dimension() << ")\n";
     stream << "{";
-    for (const auto &stuff : structure.generator_dimensions)
-      stream << stuff << ", ";
+    for (const auto &stuff : structure.generator_dimensions) stream << stuff << ", ";
     if (structure.size() > 0) {
       stream << "\b" << "\b";
     }
     stream << "}\n";
     return stream;
   }
-  inline void to_stream(std::ostream &stream,
-                        const std::vector<std::size_t> &order) {
+
+  inline void to_stream(std::ostream &stream, const std::vector<index_type> &order) {
     for (const auto &i : order) {
       const auto &stuff = this->operator[](i);
       stream << i << " : [";
-      for (const auto &truc : stuff)
-        stream << truc << ", ";
+      for (const auto &truc : stuff) stream << truc << ", ";
[…]
