+++ multipers/ml/point_clouds.py
@@ -266,7 +266,9 @@ class PointCloud2FilteredComplex(BaseEstimator, TransformerMixin):
                     f"Invalid complex \
                 {self.complex}. Possible choises are rips, delaunay, or alpha."
                 )
-        self._vineyard = self.output_type is None or "novine" not in self.output_type
+        self._vineyard = (
+            False if self.output_type is None else "novine" not in self.output_type
+        )
         self._output_type = (
             _pref_output
             if self.output_type is None
