+++ multipers/multi_parameter_rank_invariant/rank_invariant.h
@@ -9,14 +9,15 @@
 #include <oneapi/tbb/enumerable_thread_specific.h>
 #include <oneapi/tbb/parallel_for.h>
 #include <ostream>
-#include <utility> // std::pair
+#include <utility>  // std::pair
 #include <vector>
+
 namespace Gudhi::multiparameter::rank_invariant {
+using Index = interface::index_type;
 
 // using Elbow = std::vector<std::pair<>>;grid
 template <typename index_type>
-inline void push_in_elbow(index_type &i, index_type &j, const index_type I,
-                          const index_type J) {
+inline void push_in_elbow(index_type &i, index_type &j, const index_type I, const index_type J) {
   if (j < J) {
     j++;
     return;
@@ -28,37 +29,33 @@ inline void push_in_elbow(index_type &i, index_type &j, const index_type I,
   j++;
   return;
 }
+
 template <typename index_type, typename value_type>
-inline value_type
-get_slice_rank_filtration(const value_type x, const value_type y,
-                          const index_type I, const index_type J) {
+inline value_type get_slice_rank_filtration(const value_type x,
+                                            const value_type y,
+                                            const index_type I,
+                                            const index_type J) {
   if (x > static_cast<value_type>(I))
-    return std::numeric_limits<value_type>::has_infinity
-               ? std::numeric_limits<value_type>::infinity()
-               : std::numeric_limits<value_type>::max();
-  if (y > static_cast<value_type>(J))
-    return I + static_cast<index_type>(y);
+    return std::numeric_limits<value_type>::has_infinity ? std::numeric_limits<value_type>::infinity()
+                                                         : std::numeric_limits<value_type>::max();
+  if (y > static_cast<value_type>(J)) return I + static_cast<index_type>(y);
   return J + static_cast<index_type>(x);
 }
 
 template <typename index_type>
-inline std::pair<index_type, index_type>
-get_coordinates(index_type in_slice_value, index_type I, index_type J) {
-  if (in_slice_value <= J)
-    return {0, J};
-  if (in_slice_value <= I + J)
-    return {in_slice_value - J, J};
+inline std::pair<index_type, index_type> get_coordinates(index_type in_slice_value, index_type I, index_type J) {
+  if (in_slice_value <= J) return {0, J};
+  if (in_slice_value <= I + J) return {in_slice_value - J, J};
   return {I, in_slice_value - I};
 }
 
 template <typename dtype, typename index_type, typename Filtration>
 inline void compute_2d_rank_invariant_of_elbow(
-    Simplex_tree<Gudhi::multi_persistence::Simplex_tree_options_multidimensional_filtration<Filtration>>
-        &st_multi,
-    Simplex_tree_std &_st_container, // copy of st_multi
-    const tensor::static_tensor_view<dtype, index_type>
-        &out, // assumes its a zero tensor
-    const index_type I, const index_type J,
+    Simplex_tree<Gudhi::multi_persistence::Simplex_tree_options_multidimensional_filtration<Filtration>> &st_multi,
+    Simplex_tree_std &_st_container,                           // copy of st_multi
+    const tensor::static_tensor_view<dtype, index_type> &out,  // assumes its a zero tensor
+    const index_type I,
+    const index_type J,
     const std::vector<index_type> &grid_shape,
     const std::vector<index_type> &degrees,
     const int expand_collapse_max_dim = 0) {
@@ -81,8 +78,7 @@ inline void compute_2d_rank_invariant_of_elbow(
       for (const auto &stuff : multi_filtration) {
         value_type x = stuff[0];
         value_type y = stuff[1];
[…]
