+++ multipers/tests/test_slicer.py
@@ -88,7 +88,7 @@ def test_representative_cycles():
         {3, 6},
     ]
     truc = [list(machin) for machin in truc]
-    slicer = mp.slicer._Slicer0_vine_f64(
+    slicer = mp.slicer._Slicer_Matrix0_vine_f64(
         truc,
         np.array([max(len(x) - 1, 0) for x in truc]),
         np.array([list(range(len(truc))), list(range(len(truc)))]).T,

[…]
