### 2024-10-19 0459cff
fix: slicer args not taken into account
