### 2024-10-20 ca2417c
fix: GudhiCohomology inf trick is not ready yet
