+++ multipers/multi_parameter_rank_invariant/persistence_slices.h
@@ -7,35 +7,34 @@
 
 namespace Gudhi::multiparameter {
 
-struct Simplex_tree_float { // smaller simplextrees
+struct Simplex_tree_float {  // smaller simplextrees
   typedef linear_indexing_tag Indexing_tag;
   typedef std::int32_t Vertex_handle;
   typedef float Filtration_value;
   typedef std::uint32_t Simplex_key;
   static const bool store_key = true;
   static const bool store_filtration = true;
-  static const bool contiguous_vertices =
-      false; // TODO OPTIMIZATION : maybe make the simplextree contiguous when
-             // calling grid_squeeze ?
+  static const bool contiguous_vertices = false;  // TODO OPTIMIZATION : maybe make the simplextree contiguous when
+                                                  // calling grid_squeeze ?
   static const bool link_nodes_by_label = true;
   static const bool stable_simplex_handles = false;
   static const bool is_multi_parameter = false;
 };
+
 // using Simplex_tree_float = Simplex_tree_options_fast_persistence;
 
 using Gudhi::multi_persistence::Box;
 using Simplex_tree_std = Simplex_tree<Simplex_tree_float>;
 
-using Barcode = std::vector<std::pair<Simplex_tree_std::Filtration_value,
-                                      Simplex_tree_std::Filtration_value>>;
+using Barcode = std::vector<std::pair<Simplex_tree_std::Filtration_value, Simplex_tree_std::Filtration_value>>;
+
 inline Barcode compute_dgm(Simplex_tree_std &st, int degree) {
   st.initialize_filtration(true);
   constexpr int coeff_field_characteristic = 11;
   constexpr Simplex_tree_std::Filtration_value min_persistence = 0;
   bool persistence_dim_max = st.dimension() == degree;
-  Gudhi::persistent_cohomology::Persistent_cohomology<
-      Simplex_tree_std, Gudhi::persistent_cohomology::Field_Zp>
-      pcoh(st, persistence_dim_max);
+  Gudhi::persistent_cohomology::Persistent_cohomology<Simplex_tree_std, Gudhi::persistent_cohomology::Field_Zp> pcoh(
+      st, persistence_dim_max);
   pcoh.init_coefficients(coeff_field_characteristic);
   pcoh.compute_persistent_cohomology(min_persistence);
   const auto &persistent_pairs = pcoh.intervals_in_dimension(degree);
@@ -46,9 +45,10 @@ inline Barcode compute_dgm(Simplex_tree_std &st, int degree) {
 }
 
 template <typename degree_type, class interface_std_like>
-inline std::vector<Barcode>
-compute_dgms(interface_std_like &st, const std::vector<degree_type> &degrees,
-             int num_collapses, int expansion_dim) {
+inline std::vector<Barcode> compute_dgms(interface_std_like &st,
+                                         const std::vector<degree_type> &degrees,
+                                         int num_collapses,
+                                         int expansion_dim) {
   std::vector<Barcode> out(degrees.size());
   static_assert(!interface_std_like::Options::is_multi_parameter,
                 "Can only compute persistence for 1-parameter simplextrees.");
@@ -65,7 +65,7 @@ compute_dgms(interface_std_like &st, const std::vector<degree_type> &degrees,
     st.expansion(expansion_dim);
   }
 
-  st.initialize_filtration(true); // true is ignore_infinite_values
+  st.initialize_filtration(true);  // true is ignore_infinite_values
   constexpr int coeff_field_characteristic = 11;
   constexpr typename interface_std_like::Filtration_value min_persistence = 0;
 
[…]
