### 2024-10-21 ab29505
fix: remove debug print
