+++ multipers/data/MOL2.py
@@ -1,350 +1,458 @@
-import numpy as np
-from os.path import expanduser
-import pandas as pd
-from sklearn.preprocessing import LabelEncoder
-from os.path import expanduser
-from os import listdir
 import os
-import MDAnalysis as mda
+from os import listdir
+from os.path import expanduser
+from typing import Iterable
+
 import matplotlib.pyplot as plt
+import MDAnalysis as mda
+import numpy as np
+import pandas as pd
+from joblib import Parallel, delayed
 from MDAnalysis.topology.guessers import guess_masses
-import multipers as mp
+from sklearn.base import BaseEstimator, TransformerMixin
+from sklearn.preprocessing import LabelEncoder
+
 # from numba import njit
 from tqdm import tqdm
-from typing import Iterable
-from joblib import Parallel, delayed
-from sklearn.base import BaseEstimator,TransformerMixin
 
+import multipers as mp
 
 DATASET_PATH = expanduser("~/Datasets/")
 JC_path = DATASET_PATH + "Cleves-Jain/"
 DUDE_path = DATASET_PATH + "DUD-E/"
 
 
-#pathes = get_data_path()
-#imgs = apply_pipeline(pathes=pathes, pipeline=pipeline_img)
-#distances_to_letter, ytest = img_distances(imgs)
+# pathes = get_data_path()
+# imgs = apply_pipeline(pathes=pathes, pipeline=pipeline_img)
+# distances_to_letter, ytest = img_distances(imgs)
 
 
 def _get_mols_in_path(folder):
-	with open(folder+"/TargetList", "r") as f:
-		train_data =  [folder + "/" + mol.strip() for mol in f.readlines()]
-	criterion = lambda dataset : dataset.endswith(".mol2") and not dataset.startswith("final") and dataset not in train_data
-	test_data = [folder + "/" + dataset for dataset in listdir(folder) if criterion(folder + "/" + dataset)]
-	return train_data, test_data
-def get_data_path_JC(type="dict"):
-	if type == "dict": out = {}
-	elif type == "list": out = []
-	else: raise TypeError(f"Type {out} not supported")
-	for stuff in listdir(JC_path):
-		if stuff.startswith("target_"):
-			current_letter = stuff[-1]
-			to_add = _get_mols_in_path(JC_path + stuff)
-			if type == "dict":	out[current_letter] = to_add
-			elif type == "list": out.append(to_add)
-	decoy_folder = JC_path + "RognanRing850/"
-	to_add = [decoy_folder + mol for mol in listdir(decoy_folder) if mol.endswith(".mol2")]
-	if type == "dict":	out["decoy"] = to_add
-	elif type == "list": out.append(to_add)
-	return out
-def get_all_JC_path():
-	out = []
-	for stuff in listdir(JC_path):
-		if stuff.startswith("target_"):
-			train_data, test_data =  _get_mols_in_path(JC_path + stuff)
-			out += train_data
-			out += test_data
-	decoy_folder = JC_path + "RognanRing850/"
-	out +=[decoy_folder + mol for mol in listdir(decoy_folder) if mol.endswith(".mol2")]
-	return out
-		
+    with open(folder + "/TargetList", "r") as f:
+        train_data = [folder + "/" + mol.strip() for mol in f.readlines()]
+    criterion = (
+        lambda dataset: dataset.endswith(".mol2")
+        and not dataset.startswith("final")
+        and dataset not in train_data
+    )
+    test_data = [
+        folder + "/" + dataset
+        for dataset in listdir(folder)
+        if criterion(folder + "/" + dataset)
+    ]
+    return train_data, test_data
 
[…]
