### 2024-10-21 5518b18
fix: torch rips-density + allow tune the backend
