### 2024-10-22 46ffb80
fix: factorize code
