### 2024-10-22 a60f389
feat: added a notebook for cubical cplx
