+++ docs/index.rst
@@ -115,6 +115,7 @@ For theoretical references,
   notebooks/graph_classification
   notebooks/molecular_embedding
   notebooks/degree_rips_interface
+  notebooks/immuno_img
 
 .. toctree::
   :caption: Differentiation
