### 2024-10-23 7958a1e
fix: dim order for convolutions
