### 2024-10-23 #6773 @DavidLapous
I just did a new release, archived with zenodo at https://zenodo.org/records/13981409.
@RMeli do you have an update on the DOIs issues ? The references without DOI all seem to be Neurips / JMLR / other publications that don't provide DOIs.
