### 2024-10-24 42b7660
make doc manual
