+++ .github/workflows/doc.yml
@@ -3,8 +3,8 @@ name: Build documentation + upload
 
 on:
   # Runs on pushes targeting the default branch
-  push:
-    branches: ["main"]
+  # push:
+  #   branches: ["main"]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
