### 2024-10-28 f9e2cda
fix: indices for function_delaunay
