### 2024-10-28 a3cb22b
tests: add tests for scc indices
