### 2024-10-28 #6773 @RMeli
Looks good @DavidLapous. I left a few final comments on the paper in https://github.com/DavidLapous/multipers/pull/30 and then it is good to go.
