### 2024-10-29 f557bcf
updated paper .bib
