+++ paper/paper.bib
@@ -9,10 +9,18 @@
   publisher = {arXiv},
   doi = {10.48550/arXiv.2203.14289},
   urldate = {2022-06-11},
-  abstract = {In topological data analysis (TDA), one often studies the shape of data by constructing a filtered topological space, whose structure is then examined using persistent homology. However, a single filtered space often does not adequately capture the structure of interest in the data, and one is led to consider multiparameter persistence, which associates to the data a space equipped with a multiparameter filtration. Multiparameter persistence has become one of the most active areas of research within TDA, with exciting progress on several fronts. In this article, we introduce multiparameter persistence and survey some of this recent progress, with a focus on ideas likely to lead to practical applications in the near future.},
+  abstract = {In topological data analysis (TDA), one often studies the shape of data by
+              constructing a filtered topological space, whose structure is then examined using
+              persistent homology. However, a single filtered space often does not adequately capture
+              the structure of interest in the data, and one is led to consider multiparameter
+              persistence, which associates to the data a space equipped with a multiparameter
+              filtration. Multiparameter persistence has become one of the most active areas of
+              research within TDA, with exciting progress on several fronts. In this article, we
+              introduce multiparameter persistence and survey some of this recent progress, with a
+              focus on ideas likely to lead to practical applications in the near future.},
   archiveprefix = {arxiv},
-  keywords = {Computer Science - Computational Geometry,Mathematics - Algebraic Topology,Mathematics - Representation Theory},
-  file = {/user/dloiseau/home/Zotero/storage/AVCY36LB/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/QH98BQFF/Botnan et Lesnick - 2022 - An Introduction to Multiparameter Persistence.pdf;/user/dloiseau/home/Zotero/storage/LPFXRBHW/2203.html}
+  keywords = {Computer Science - Computational Geometry,Mathematics - Algebraic Topology,Mathematics
+              - Representation Theory},
 }
 
 @article{code_persistable,
@@ -28,9 +36,9 @@
   issn = {2475-9066},
   doi = {10.21105/joss.05022},
   urldate = {2024-05-09},
-  abstract = {Scoccola et al., (2023). Persistable: persistent and stable clustering. Journal of Open Source Software, 8(83), 5022, https://doi.org/10.21105/joss.05022},
+  abstract = {Scoccola et al., (2023). Persistable: persistent and stable clustering. Journal of
+              Open Source Software, 8(83), 5022, https://doi.org/10.21105/joss.05022},
   langid = {english},
-  file = {/user/dloiseau/home/Zotero/storage/3DS9DWXU/Scoccola et Rolle - 2023 - Persistable persistent and stable clustering.pdf}
 }
 
 @article{erstaircode,
@@ -47,10 +55,33 @@
   issn = {2470-6566},
   doi = {10.1137/20M1353605},
   urldate = {2022-05-20},
[…]
