+++ paper/paper.md
@@ -91,7 +91,7 @@ Although relying on distinct methods, `multipers` can also be used to compute Ma
 <!-- However, this library doesn't leverage on recently developped significant optimizations ([@filtration_domination], [@mpfree]), takes only very specific text-file inputs, doesn't support general multi-critical filtrations, and can only compute 1-parameter slices.  -->
 <!-- Furthermore, although rivet can be used to compute some very specific multiparameter invariants, such as MPL, it needs other libraries to do so. -->
 
-[**Multiparameter Persistent Landscape.**](https://github.com/OliverVipond/Multiparameter_Persistence_Landscapes) Implemented on top of [Rivet](https://github.com/rivetTDA/rivet) [@rivet], this library computes a multiparameter persistent descriptor by computing 1-parameter persistence landscape vectorizations of slices in multi-filtrations [@mpl], called Multiparameter Persistent Landscape (MPL). This library also features some multiparameter persistence visualizations. However, it is limited to `Rivet` capabilities and landscapes computations, which on one hand does not leverage on recently developped optimizations, e.g., [@filtration_domination], or [@mpfree], and on the other hand can only work with very specific text file inputs.
+[**Multiparameter Persistent Landscape.**](https://github.com/OliverVipond/Multiparameter_Persistence_Landscapes) Implemented on top of [Rivet](https://github.com/rivetTDA/rivet) [@rivet], this library computes a multiparameter persistent descriptor by computing 1-parameter persistence landscape vectorizations of slices in multi-filtrations [@mpl], called Multiparameter Persistent Landscape (MPL). This library also features some multiparameter persistence visualizations. However, it is limited to `Rivet` capabilities and landscapes computations, which on one hand does not leverage on recently developed optimizations, e.g., [@filtration_domination], or [@mpfree], and on the other hand can only work with very specific text file inputs.
 
 <!--**Multiparameter Persistent Kernel.** Similar to landscapes, this invariant can be computed on diagonal persistence slices given by, e.g., `Rivet`.-->
 
