### 2024-10-30 #6773 @RMeli
Thank you @DavidLapous for submitting to JOSS. Based on the extensive comments from the reviewers (and my own assessment), I'm happy to recommend this paper for acceptance! The EiC will now perform the final checks. Thank you for engaging so well with the review process.

Thank you very much again @yossibokorbleile and @peekxc for the engaging in the review process!
