### 2024-10-30 #6773 @RMeli
@DavidLapous using https://doi.org/10/gtt7dq redirects me to https://doi.org/10.1137/22M1489150 Can you please use the latter, which seems to have the expected format for the automated check?
