+++ paper/paper.bib
@@ -504,7 +504,7 @@
   number = {1},
   pages = {54--88},
   publisher = {{Society for Industrial and Applied Mathematics}},
-  doi = {10/gtt7dq},
+  doi = {10.1137/22M1489150},
   urldate = {2024-05-14},
   abstract = {.We introduce harmonic persistent homology spaces for filtrations of finite simplicial
               complexes. As a result, we can associate concrete subspaces of cycles to each bar of
