### 2024-10-30 #6773 @DavidLapous
> @DavidLapous using https://doi.org/10/gtt7dq redirects me to https://doi.org/10.1137/22M1489150 Can you please use the latter, which seems to have the expected format for the automated check?

Done in davidlapous/multipers@69f0ea116a7919827184576f31cee2a579d37f0a.
