### 2024-10-30 #6773 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
