### 2024-10-30 1c2fde9
paper: update .bib
