### 2024-10-30 #6773 @DavidLapous
If it's not too late, I've updated an old reference (draft to published) in davidlapous/multipers@1c2fde9885d49fe6accf5e8de5f9a21e44528e64.
