### 2024-11-04 b453a79
de/serialzation of simplex tree working also for One_critical and Multi_critical filtration values
