+++ multipers/gudhi/gudhi/One_critical_filtration.h
@@ -1196,6 +1196,31 @@ class One_critical_filtration : public std::vector<T> {
     return f1.push_to_least_common_upper_bound(f2);
   }
 
+  friend char* serialize_trivial(const One_critical_filtration& value, char* start)
+  {
+    const auto length = value.size();
+    const std::size_t arg_size = sizeof(T) * length;
+    const std::size_t type_size = sizeof(One_critical_filtration::Base::size_type);
+    memcpy(start, &length, type_size);
+    memcpy(start + type_size, value.data(), arg_size);
+    return start + arg_size + type_size;
+  }
+
+  friend const char* deserialize_trivial(One_critical_filtration& value, const char* start)
+  {
+    const std::size_t type_size = sizeof(One_critical_filtration::Base::size_type);
+    typename One_critical_filtration::Base::size_type length;
+    memcpy(&length, start, type_size);
+    std::size_t arg_size = sizeof(T) * length;
+    value.resize(length);
+    memcpy(value.data(), start + type_size, arg_size);
+    return start + arg_size + type_size;
+  }
+
+  friend std::size_t get_serialization_size_of(const One_critical_filtration& value) {
+    return sizeof(One_critical_filtration::Base::size_type) + sizeof(T) * value.size();
+  }
+
   /**
    * @brief Infinity value of an entry of the filtration value.
    */
