+++ multipers/gudhi/gudhi/Simplex_tree.h
@@ -176,11 +176,12 @@ class Simplex_tree {
   struct Filtration_simplex_base_dummy {
     Filtration_simplex_base_dummy() {}
     Filtration_simplex_base_dummy(Filtration_value GUDHI_CHECK_code(f)) {
-      GUDHI_CHECK(f == 0, "filtration value specified for a complex that does not store them");
+      GUDHI_CHECK(f == Filtration_value(),
+                  "filtration value specified in the constructor for a complex that does not store them");
     }
 
     void assign_filtration(const Filtration_value& GUDHI_CHECK_code(f)) {
-      GUDHI_CHECK(f == Filtration_value(), "filtration value specified for a complex that does not store them");
+      GUDHI_CHECK(f == Filtration_value(), "filtration value assigned for a complex that does not store them");
     }
     const Filtration_value& filtration() const { return null_; }
 
@@ -1230,7 +1231,15 @@ class Simplex_tree {
    * which can be cleared with clear_filtration().
    */
   void initialize_filtration(bool ignore_infinite_values = false) {
-    initialize_filtration(is_before_in_totally_ordered_filtration(this), ignore_infinite_values);
+    if (ignore_infinite_values){
+      initialize_filtration(is_before_in_totally_ordered_filtration(this), [&](Simplex_handle sh) -> bool {
+        return filtration(sh) == Filtration_simplex_base_real::get_infinity();
+      });
+    } else {
+      initialize_filtration(is_before_in_totally_ordered_filtration(this), [](Simplex_handle) -> bool {
+        return false;
+      });
+    }
   }
 
   /**
@@ -1254,17 +1263,22 @@ class Simplex_tree {
    * which can be cleared with @ref clear_filtration().
    * 
    * @tparam Comparator Method type taking two Simplex_handle as input and returns a bool.
+   * @tparam Ignorer Method type taking one Simplex_handle as input and returns a bool.
    * @param is_before_in_filtration Method used to compare two simplices with respect to their position in the
    * wanted filtration. Takes two Simplex_handle as input and returns true if and only if the first simplex appears
    * strictly before the second simplex in the resulting 1-parameter filtration.
-   * @param ignore_infinite_values If true, simplices with filtration values at infinity are not cached.
+   * @param ignore_simplex Method taking a simplex handle as input and returns true if and only if the corresponding
+   * simplex should be ignored in the filtration and therefore not be cached. But note that it is important that
+   * the resulting filtration is still a valid filtration, that is all proper faces of a not ignored simplex
+   * should also not be ignored and have to appear before this simplex. There will be no tests to ensure that, so
+   * it is of the users responsibility.
    */
-  template<typename Comparator>
-  void initialize_filtration(Comparator&& is_before_in_filtration, bool ignore_infinite_values = false) {
+  template<typename Comparator, typename Ignorer>
+  void initialize_filtration(Comparator&& is_before_in_filtration, Ignorer&& ignore_simplex) {
     filtration_vect_.clear();
     filtration_vect_.reserve(num_simplices());
     for (Simplex_handle sh : complex_simplex_range()) {
-      if (ignore_infinite_values && filtration(sh) == Filtration_simplex_base_real::get_infinity()) continue;
+      if (ignore_simplex(sh)) continue;
       filtration_vect_.push_back(sh);
     }
 
[…]
