### 2024-11-04 73356ea
add set/getstate python + some fixes + test
