+++ multipers/gudhi/gudhi/Simplex_tree.h
@@ -54,15 +54,15 @@
 #include <stdexcept>
 #include <limits>  // Inf
 #include <initializer_list>
-#include <algorithm>  // for std::max
-#include <iterator>  // for std::distance
+#include <algorithm>    // for std::max
+#include <iterator>     // for std::distance
 #include <type_traits>  // for std::conditional
 #include <unordered_map>
 #include <iterator>  // for std::prev
 
 namespace Gudhi {
 
-/** \addtogroup simplex_tree 
+/** \addtogroup simplex_tree
  *  @{
  */
 
@@ -78,7 +78,7 @@ namespace Gudhi {
  * Details may be found in \cite Cohen-Steiner2009 and section 2.2 in \cite Carriere16.
  *
  */
-enum class Extended_simplex_type {UP, DOWN, EXTRA};
+enum class Extended_simplex_type { UP, DOWN, EXTRA };
 
 /**
  * \class Simplex_tree Simplex_tree.h gudhi/Simplex_tree.h
@@ -93,7 +93,7 @@ enum class Extended_simplex_type {UP, DOWN, EXTRA};
  *
  */
 
-template<typename SimplexTreeOptions = Simplex_tree_options_default>
+template <typename SimplexTreeOptions = Simplex_tree_options_default>
 class Simplex_tree {
  public:
   typedef SimplexTreeOptions Options;
@@ -120,26 +120,30 @@ class Simplex_tree {
   // flat_set (with our own comparator) where we can control the layout of the struct (put Vertex_handle and
   // Simplex_key next to each other).
   typedef typename boost::container::flat_map<Vertex_handle, Node> flat_map;
-  //Dictionary::iterator remain valid under insertions and deletions,
-  //necessary e.g. when computing oscillating rips zigzag filtrations.
+  // Dictionary::iterator remain valid under insertions and deletions,
+  // necessary e.g. when computing oscillating rips zigzag filtrations.
   typedef typename boost::container::map<Vertex_handle, Node> map;
-  typedef typename std::conditional<Options::stable_simplex_handles,
-                                    map,
-                                    flat_map>::type Dictionary;
+  typedef typename std::conditional<Options::stable_simplex_handles, map, flat_map>::type Dictionary;
 
   /** \brief Set of nodes sharing a same parent in the simplex tree. */
   typedef Simplex_tree_siblings<Simplex_tree, Dictionary> Siblings;
 
   struct Key_simplex_base_real {
     Key_simplex_base_real() : key_(-1) {}
+
     Key_simplex_base_real(Simplex_key k) : key_(k) {}
+
     void assign_key(Simplex_key k) { key_ = k; }
+
     Simplex_key key() const { return key_; }
+
    private:
     Simplex_key key_;
   };
+
   struct Key_simplex_base_dummy {
     Key_simplex_base_dummy() {}
+
     // Undefined so it will not link
     void assign_key(Simplex_key);
     Simplex_key key() const;
@@ -148,20 +152,24 @@ class Simplex_tree {
   struct Extended_filtration_data {
     Filtration_value minval;
     Filtration_value maxval;
+
     Extended_filtration_data() {}
 
-    Extended_filtration_data(const Filtration_value& vmin, const Filtration_value& vmax)
-        : minval(vmin), maxval(vmax)
-    {}
+    Extended_filtration_data(const Filtration_value& vmin, const Filtration_value& vmax) : minval(vmin), maxval(vmax) {}
   };
+
   typedef typename std::conditional<Options::store_key, Key_simplex_base_real, Key_simplex_base_dummy>::type
       Key_simplex_base;
 
   struct Filtration_simplex_base_real {
[…]
