+++ multipers/gudhi/gudhi/Multi_critical_filtration.h
@@ -25,7 +25,6 @@
 #include <cstring>
 #include <iostream>
 #include <limits>
-#include <string>
 #include <utility>
 #include <vector>
 
@@ -759,45 +758,45 @@ class Multi_critical_filtration {
     return stream;
   }
 
-  friend bool unify_lifetimes(Multi_critical_filtration& f1, const Multi_critical_filtration& f2){
+  friend bool unify_lifetimes(Multi_critical_filtration &f1, const Multi_critical_filtration &f2) {
     bool modified = false;
-    for (const Generator& g : f2.multi_filtration_){
+    for (const Generator &g : f2.multi_filtration_) {
       modified |= f1.add_generator(g);
     }
     return modified;
   }
 
-  friend bool intersect_lifetimes(Multi_critical_filtration& f1, const Multi_critical_filtration& f2){
+  friend bool intersect_lifetimes(Multi_critical_filtration &f1, const Multi_critical_filtration &f2) {
     if (f1.is_nan() || f2.is_nan()) return false;
 
-    if constexpr (co){
-      if (f1.is_plus_inf()){
+    if constexpr (co) {
+      if (f1.is_plus_inf()) {
         if (f2.is_plus_inf()) return false;
         f1 = f2;
         return true;
       }
-      if (f1.is_minus_inf()){
+      if (f1.is_minus_inf()) {
         return false;
       }
     } else {
-      if (f1.is_minus_inf()){
+      if (f1.is_minus_inf()) {
         if (f2.is_minus_inf()) return false;
         f1 = f2;
         return true;
       }
-      if (f1.is_plus_inf()){
+      if (f1.is_plus_inf()) {
         return false;
       }
     }
-    
+
     Multi_critical_filtration res(1, -_get_default_value());
     // TODO: see if the order can be used to avoid n^2 complexity and
     // perhaps even to replace add_generator by add_guaranteed_generator
-    for (const Generator& of1 : f1.multi_filtration_){
-      for (const Generator& of2 : f2.multi_filtration_){
+    for (const Generator &of1 : f1.multi_filtration_) {
+      for (const Generator &of2 : f2.multi_filtration_) {
         // TODO: avoid one go-through by constructing nf directly as the max/min
         Generator nf = of1;
-        if constexpr (co){
+        if constexpr (co) {
           nf.pull_to_greatest_common_lower_bound(of2);
         } else {
           nf.push_to_least_common_upper_bound(of2);
@@ -810,34 +809,32 @@ class Multi_critical_filtration {
     return f1 != res;
   }
 
-  friend char* serialize_trivial(const Multi_critical_filtration& value, char* start)
-  {
+  friend char *serialize_trivial(const Multi_critical_filtration &value, char *start) {
     const auto nberOfGenerators = value.num_generators();
     const std::size_t type_size = sizeof(std::size_t);
     memcpy(start, &nberOfGenerators, type_size);
-    char* curr = start + type_size;
-    for (const Generator& g : value){
+    char *curr = start + type_size;
+    for (const Generator &g : value) {
       curr = serialize_trivial(g, curr);
     }
     return curr;
   }
 
-  friend const char* deserialize_trivial(Multi_critical_filtration& value, const char* start)
-  {
+  friend const char *deserialize_trivial(Multi_critical_filtration &value, const char *start) {
     const std::size_t type_size = sizeof(std::size_t);
     std::size_t nberOfGenerators;
     memcpy(&nberOfGenerators, start, type_size);
-    const char* curr = start + type_size;
-    value.resize(nberOfGenerators);
-    for (const Generator& g : value){
+    const char *curr = start + type_size;
+    value.set_num_generators(nberOfGenerators);
+    for (Generator &g : value) {
[…]
