### 2024-11-05 11f715f
ignore infinite values with coh interface + possibility of non simplicial for cohomology
