+++ multipers/gudhi/gudhi/Persistent_cohomology.h
@@ -165,7 +165,8 @@ class Persistent_cohomology {
    *
    * Assumes that the filtration provided by the simplicial complex is
    * valid. Undefined behavior otherwise. */
-  void compute_persistent_cohomology(Filtration_value min_interval_length = 0) {
+  void compute_persistent_cohomology(Filtration_value min_interval_length = 0,
+                                     const bool is_simplicial_and_starting_from_dim_0 = true) {
     interval_length_policy.set_length(min_interval_length);
     Simplex_key idx_fil = -1;
     std::vector<Simplex_key> vertices; // so we can check the connected components at the end
@@ -174,16 +175,20 @@ class Persistent_cohomology {
       cpx_->assign_key(sh, ++idx_fil);
       dsets_.make_set(cpx_->key(sh));
       int dim_simplex = cpx_->dimension(sh);
-      switch (dim_simplex) {
-        case 0:
-          vertices.push_back(idx_fil);
-          break;
-        case 1:
-          update_cohomology_groups_edge(sh);
-          break;
-        default:
-          update_cohomology_groups(sh, dim_simplex);
-          break;
+      if (is_simplicial_and_starting_from_dim_0){
+        switch (dim_simplex) {
+          case 0:
+            vertices.push_back(idx_fil);
+            break;
+          case 1:
+            update_cohomology_groups_edge(sh);
+            break;
+          default:
+            update_cohomology_groups(sh, dim_simplex);
+            break;
+        }
+      } else {
+        update_cohomology_groups(sh, dim_simplex);
       }
     }
     // Compute infinite intervals of dimension 0
