### 2024-11-05 68a97ef
bug fix in R matrix reduction
