+++ multipers/gudhi/gudhi/Persistence_matrix/base_pairing.h
@@ -132,7 +132,7 @@ inline const typename Base_pairing<Master_matrix>::Barcode& Base_pairing<Master_
 template <class Master_matrix>
 inline void Base_pairing<Master_matrix>::_reduce() 
 {
-  std::unordered_map<ID_index, Index> pivotsToColumn(_matrix()->get_number_of_columns());
+  std::unordered_map<Index, Index> negativeColumns(_matrix()->get_number_of_columns());
 
   auto dim = _matrix()->get_max_dimension();
   std::vector<std::vector<Index> > columnsByDim(dim + 1);
@@ -144,17 +144,21 @@ inline void Base_pairing<Master_matrix>::_reduce()
     for (Index i : cols) {
       auto& curr = _matrix()->get_column(i);
       if (curr.is_empty()) {
-        if (pivotsToColumn.find(i) == pivotsToColumn.end()) {
+        if (negativeColumns.find(i) == negativeColumns.end()) {
           barcode_.emplace_back(i, -1, dim);
         }
       } else {
         ID_index pivot = curr.get_pivot();
+        auto it = idToPosition_.find(pivot);
+        Index pivotColumnNumber = it == idToPosition_.end() ? pivot : it->second;
+        auto itNeg = negativeColumns.find(pivotColumnNumber);
+        Index pivotKiller = itNeg == negativeColumns.end() ? -1 : itNeg->second;
 
-        while (pivot != static_cast<ID_index>(-1) && pivotsToColumn.find(pivot) != pivotsToColumn.end()) {
+        while (pivot != static_cast<ID_index>(-1) && pivotKiller != static_cast<Index>(-1)) {
           if constexpr (Master_matrix::Option_list::is_z2) {
-            curr += _matrix()->get_column(pivotsToColumn.at(pivot));
+            curr += _matrix()->get_column(pivotKiller);
           } else {
-            auto& toadd = _matrix()->get_column(pivotsToColumn.at(pivot));
+            auto& toadd = _matrix()->get_column(pivotKiller);
             typename Master_matrix::Element coef = toadd.get_pivot_value();
             auto& operators = _matrix()->colSettings_->operators;
             coef = operators.get_inverse(coef);
@@ -163,12 +167,14 @@ inline void Base_pairing<Master_matrix>::_reduce()
           }
 
           pivot = curr.get_pivot();
+          it = idToPosition_.find(pivot);
+          pivotColumnNumber = it == idToPosition_.end() ? pivot : it->second;
+          itNeg = negativeColumns.find(pivotColumnNumber);
+          pivotKiller = itNeg == negativeColumns.end() ? -1 : itNeg->second;
         }
 
         if (pivot != static_cast<ID_index>(-1)) {
-          pivotsToColumn.emplace(pivot, i);
-          auto it = idToPosition_.find(pivot);
-          auto pivotColumnNumber = it == idToPosition_.end() ? pivot : it->second;
+          negativeColumns.emplace(pivotColumnNumber, i);
           _matrix()->get_column(pivotColumnNumber).clear();
           barcode_.emplace_back(pivotColumnNumber, i, dim - 1);
         } else {
