### 2024-11-05 ddfccd7
fix: readme typos
