+++ README.md
@@ -2,14 +2,14 @@
 [![Build, test](https://github.com/DavidLapous/multipers/actions/workflows/python_PR.yml/badge.svg)](https://github.com/DavidLapous/multipers/actions/workflows/python_PR.yml)
 [![PyPI](https://img.shields.io/pypi/v/multipers?color=green)](https://pypi.org/project/multipers)
 [![Downloads](https://static.pepy.tech/badge/multipers)](https://pepy.tech/project/multipers)
-[Documentation](https://www-sop.inria.fr/members/David.Loiseaux/doc/multipers/index.html)
+[Documentation](https://davidlapous.github.io/multipers/index.html)
 <br>
 Scikit-style PyTorch-autodiff multiparameter persistent homology python library. 
 This library aims to provide easy to use and performant strategies for applied multiparameter topology.
 <br> Meant to be integrated in [the Gudhi library](https://gudhi.inria.fr/).
 
 ## Multiparameter Persistence
-This library allows to compute several representation from "geometrical datasets", e.g., point clouds, images, graphs, that have multiple scales. A well known example is is the following one. 
+This library allows computing several representations from "geometrical datasets", e.g., point clouds, images, graphs, that have multiple scales. A well known example is the following one. 
 <br>
 Pick a point cloud that has diffuse noise, or on which the sampling measure has some interesting properties, e.g., in the following example the measure has three modes.
 <br>
@@ -18,13 +18,13 @@ This filtration $X$, indexed over a radius parameter $r$ and a codensity paramet
 
 $$ X_{r,s} = \bigcup_{x \in P, \, \mathrm{density}(x) \ge s} B(x,r) = \lbrace x\in \mathbb R^2 \mid \exists p \in P, \, \mathrm{density}(p) \ge s \text{ and } d(x,p) \le r \rbrace$$
 
-The green shape on the left represent the lifetime of the biggest annulus. On the right, each cycle appearing on the left gets a colored shape (the color is only a label) and the shape of this colored shape represents the lifetime of this cycle.
+The green shape on the left represents the lifetime of the biggest annulus. On the right, each cycle appearing on the left gets a colored shape (the color is only a label) and the shape of this colored shape represents the lifetime of this cycle.
 <br>
 In our case, the big green shape on the left corresponds to the largest green shape appearing on the right, recovering the structure of the annulus here.
 ![Alt text](docs/1.png)
 
 The **magic** part is that we never had to choose any parameter to remove the noise in this construction, but the annulus still naturally appears!
-<br>A more strinking example is the following one.
+<br>A more striking example is the following one.
 Using the same constructions, we can identify topological structure, and their size, in a parameter free approach, even though the majority of the sampling measure's mass is noise.
 <br> In this example, the lifetime shape associated to each cycle can be identified from 
  - Their radius : the smaller cycle will naturally live more this shape will appear on the "left" (smaller radius)
@@ -36,7 +36,7 @@ We also provide several other descriptors, as well as associated Machine Learnin
 ![alt text](docs/3.png)
 
 
-A non-exhaustive list of features can be found in the **Features** section, and in the [documentation](https://www-sop.inria.fr/members/David.Loiseaux/doc/multipers/index.html).
+A non-exhaustive list of features can be found in the **Features** section, and in the [documentation](https://davidlapous.github.io/multipers/index.html).
 ## Quick start
 
 
@@ -46,7 +46,7 @@ pip install multipers
 ```
 We recommend Windows user to use [WSL](https://learn.microsoft.com/en-us/windows/wsl/). 
 <br>
[…]
