+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "multipers"
-version = "2.2.0"
+version = "2.2.1"
 description = "Multiparameter Topological Persistence for Machine Learning"
 readme = "README.md"
 authors = [
