### 2024-11-06 #6773 @DavidLapous
Once again, if it's not too late @RMeli, I've just release the version to v2.2.1, and updated the corresponding zenodo archive at  https://zenodo.org/records/14042221
