### 2024-11-07 850db26
feat: small vector column
