+++ multipers/gudhi/gudhi/Persistence_matrix/columns/small_vector_column.h
@@ -1,0 +1,1093 @@
+/*    This file is part of the Gudhi Library - https://gudhi.inria.fr/ - which is released under MIT.
+ *    See file LICENSE or go to https://gudhi.inria.fr/licensing/ for full license details.
+ *    Author(s):       Hannah Schreiber
+ *
+ *    Copyright (C) 2022-24 Inria
+ *
+ *    Modification(s):
+ *      - YYYY/MM Author: Description of the modification
+ */
+
+/**
+ * @file naive_vector_column.h
+ * @author Hannah Schreiber
+ * @brief Contains the @ref Gudhi::persistence_matrix::Naive_vector_column class.
+ * Also defines the std::hash method for @ref Gudhi::persistence_matrix::Naive_vector_column.
+ */
+
+#ifndef PM_SMALL_VECTOR_COLUMN_H
+#define PM_SMALL_VECTOR_COLUMN_H
+
+#include <vector>
+#include <stdexcept>
+#include <type_traits>
+#include <algorithm>  //binary_search
+#include <utility>    //std::swap, std::move & std::exchange
+
+#include <boost/iterator/indirect_iterator.hpp>
+#include <boost/container/small_vector.hpp>
+
+#include <gudhi/Persistence_matrix/allocators/entry_constructors.h>
+#include <gudhi/Persistence_matrix/columns/column_utilities.h>
+
+namespace Gudhi {
+namespace persistence_matrix {
+
+/**
+ * @class Naive_vector_column naive_vector_column.h gudhi/Persistence_matrix/columns/naive_vector_column.h
+ * @ingroup persistence_matrix
+ *
+ * @brief Column class following the @ref PersistenceMatrixColumn concept.
+ *
+ * Column based on a vector structure. The entries are always ordered by row index and only non-zero values
+ * are stored uniquely in the underlying container.
+ *
+ * @tparam Master_matrix An instantiation of @ref Matrix from which all types and options are deduced.
+ * @tparam Entry_constructor Factory of @ref Entry classes.
+ */
+template <class Master_matrix>
+class Small_vector_column : public Master_matrix::Row_access_option,
+                            public Master_matrix::Column_dimension_option,
+                            public Master_matrix::Chain_column_option
+{
+ public:
+  using Master = Master_matrix;
+  using Index = typename Master_matrix::Index;
+  using ID_index = typename Master_matrix::ID_index;
+  using Dimension = typename Master_matrix::Dimension;
+  using Field_element = typename Master_matrix::Element;
+  using Entry = typename Master_matrix::Matrix_entry;
+  using Column_settings = typename Master_matrix::Column_settings;
+
+ private:
+  using Field_operators = typename Master_matrix::Field_operators;
+  using Column_support = boost::container::small_vector<Entry*, 10>;
+  using Entry_constructor = typename Master_matrix::Entry_constructor;
+
+ public:
+  using iterator = boost::indirect_iterator<typename Column_support::iterator>;
+  using const_iterator = boost::indirect_iterator<typename Column_support::const_iterator>;
+  using reverse_iterator = boost::indirect_iterator<typename Column_support::reverse_iterator>;
+  using const_reverse_iterator = boost::indirect_iterator<typename Column_support::const_reverse_iterator>;
+
+  Small_vector_column(Column_settings* colSettings = nullptr);
+  template <class Container = typename Master_matrix::Boundary>
+  Small_vector_column(const Container& nonZeroRowIndices, Column_settings* colSettings);
+  template <class Container = typename Master_matrix::Boundary, class Row_container>
+  Small_vector_column(Index columnIndex,
+                      const Container& nonZeroRowIndices,
+                      Row_container* rowContainer,
+                      Column_settings* colSettings);
+  template <class Container = typename Master_matrix::Boundary>
[…]
