+++ multipers/gudhi/gudhi/Matrix.h
@@ -64,6 +64,7 @@
 #include <gudhi/Persistence_matrix/columns/unordered_set_column.h>
 #include <gudhi/Persistence_matrix/columns/vector_column.h>
 #include <gudhi/Persistence_matrix/columns/naive_vector_column.h>
+#include <gudhi/Persistence_matrix/columns/small_vector_column.h>
 #include <gudhi/Persistence_matrix/columns/heap_column.h>
 
 /// Gudhi namespace.
@@ -314,6 +315,7 @@ class Matrix {
   using Matrix_list_column = List_column<Matrix<PersistenceMatrixOptions> >;
   using Matrix_vector_column = Vector_column<Matrix<PersistenceMatrixOptions> >;
   using Matrix_naive_vector_column = Naive_vector_column<Matrix<PersistenceMatrixOptions> >;
+  using Matrix_small_vector_column = Small_vector_column<Matrix<PersistenceMatrixOptions> >;
   using Matrix_set_column = Set_column<Matrix<PersistenceMatrixOptions> >;
   using Matrix_unordered_set_column = Unordered_set_column<Matrix<PersistenceMatrixOptions> >;
   using Matrix_intrusive_list_column = Intrusive_list_column<Matrix<PersistenceMatrixOptions> >;
@@ -345,7 +347,10 @@ class Matrix {
                             typename std::conditional<
                                 PersistenceMatrixOptions::column_type == Column_types::NAIVE_VECTOR,
                                 Matrix_naive_vector_column, 
-                                Matrix_intrusive_set_column
+                                typename std::conditional_t<
+                                    PersistenceMatrixOptions::column_type == Column_types::SMALL_VECTOR,
+                                    Matrix_small_vector_column,
+                                    Matrix_intrusive_set_column>
                                 >::type
                             >::type
                         >::type
