### 2024-11-07 b1b89c3
feat: slicer getstate/setstate
