### 2024-11-12 #6773 @crvernon
:wave: @DavidLapous - I just need you to address the following before I can move to accept this one for publication.

In the archive:

- [ ] Please ensure the license specified in your Zenodo archive is the same as what you specify for your codebase in your repository.

In the paper:

- [ ] LINE 35:  "...Barcodes (Botnan et al., 2022) (Oudot & Scoccola, 2024),..." should appear as "...Barcodes (Botnan et al., 2022; Oudot & Scoccola, 2024),..." where the in-text citations are separated by a semicolon.  Take a look at other JOSS published works if you are unfamiliar with how to structure this in text when referencing your bib records.
- [ ] LINE 77:  "...(Scoccola & Rolle, 2023) (Rolle & Scoccola, 2020)." should appear as "...(Scoccola & Rolle, 2023; Rolle & Scoccola, 2020)." where the in-text citations are separated by a semicolon.

Let me know when you have addressed these.  Thank you!
