### 2024-11-13 #6773 @DavidLapous
:wave: @crvernon, thanks for taking a look at this paper and your comments;

I have fixed the citation issues in davidlapous/multipers@2452c8424a2a9e4cbdb07edc4355a53b66bc5d1b and a figure reference issue in davidlapous/multipers@64edbf75b7b32064f5e133c215a989280a75e2d5.
I also fixed the license mismatch in the current zenodo archive, thank you for noticing it.
