### 2024-11-14 #6773 @crvernon
@DavidLapous one last thing...please remove the formatting around the title of your paper.  Otherwise, it will display as `<code>multipers</code> : Multiparameter Persistence for Machine Learning` in JOSS.  Once you do this, I'll reaccept.  Just let me know here when it is done.  Thanks!
