### 2024-11-14 b81acc5
paper: remove title formatting
