### 2024-11-14 #6773 @DavidLapous
@crvernon: this is fixed in davidlapous/multipers@b81acc55c31b03d11c0373c702db29b1635fb2ce. Sorry for this, I wasn't aware that this could cause an issue.
